### 2019-08-15 ce6c824
drops python 2.7 support
