+++ Makefile
@@ -1,5 +1,4 @@
-.PHONY: tests upload-pypi conda-build-27 conda-build-35 conda-build-36 \
-	conda-build-37
+.PHONY: tests upload-pypi conda-build-35 conda-build-36 conda-build-37
 
 tests:
 	pytest
@@ -25,9 +24,6 @@ requirements:
 
 mock-ci-tests:
 	. ./ci_tests.sh
-
-conda-build-27:
-	conda-build --python=2.7 conda.recipe
 
 conda-build-35:
 	conda-build --python=3.5 conda.recipe
