### 2019-08-15 23c7254
drops python 2.7 support
