+++ setup.py
@@ -2,10 +2,9 @@ from setuptools import setup
 
 with open('README.md') as f:
     long_description = f.read()
-
 setup(
     name="pandera",
-    version="0.1.5",
+    version="0.2.0",
     author="Niels Bantilan",
     author_email="niels.bantilan@gmail.com",
     description="A light-weight and flexible validation package for pandas "
@@ -23,20 +22,18 @@ setup(
         "pandera",
     ],
     install_requires=[
-        "enum34 ; python_version<'3.4'",
         "numpy >= 1.9.0",
         "pandas >= 0.23.0",
         "wrapt",
-        "scipy ; python_version<'2.7'",
+        "scipy ; python_version>='3.5'",
     ],
-    python_requires='>=2.7, !=3.0.*, !=3.1.*, !=3.2.*, !=3.3.*',
+    python_requires='>=3.5',
     platforms='any',
     classifiers=[
         'Development Status :: 5 - Production/Stable',
         'Operating System :: OS Independent',
         'Intended Audience :: Science/Research',
         'Programming Language :: Python',
-        'Programming Language :: Python :: 2.7',
         'Programming Language :: Python :: 3',
         'Programming Language :: Python :: 3.5',
         'Programming Language :: Python :: 3.6',
@@ -45,3 +42,4 @@ setup(
         ],
 
 )
+
