+++ ci_tests.sh
@@ -10,7 +10,6 @@ run_ci() {
         conda remove -q -y -n pandera-ci-env-$1 --all
 }
 
-run_ci 2.7
 run_ci 3.5
 run_ci 3.6
 run_ci 3.7
