+++ pandera/pandera.py
@@ -954,10 +954,7 @@ class Column(SeriesSchemaBase):
 
 
 def _get_fn_argnames(fn):
-    if sys.version_info.major >= 3:
-        arg_spec_args = inspect.getfullargspec(fn).args
-    else:
-        arg_spec_args = inspect.getargspec(fn).args
+    arg_spec_args = inspect.getfullargspec(fn).args
 
     if inspect.ismethod(fn) and arg_spec_args[0] == "self":
         # don't include "self" argument
