### 2019-08-15 4d38a4d
drops python 2.7 support (#88)
* drops python 2.7 support

* removal of unused sys

* increments version number in init.py
