### 2019-08-15 5b864b1
adds more explicit error message
