### 2019-08-15 4b2f849
improve sphinx docs, test python code snippets
