### 2019-08-15 c1a3616
adds explanatory error message to check_input decorator (#86)
* adds explanatory error message to check decorators and tests

* adds more explicit error message
