### 2019-08-15 23d5aa4
make docs in travis
