+++ .travis.yml
@@ -27,4 +27,4 @@ script:
   # Check conformity to flake8 style guide
   - flake8 pandera/pandera.py tests
   # Check docs can build, treating warnings as errors
-  - python -m sphinx -E -W "docs/source" "docs/_build"
+  - make docs
