### 2019-08-15 9deafb5
add make docs to travis
