### 2019-08-15 ca5b39d
improve sphinx docs, test python code snippets (#89)
* improve sphinx docs, test python code snippets

* make docs in travis

* add make docs to travis
