+++ docs/source/dataframe_schemas.rst
@@ -10,7 +10,8 @@ validated against.
 
 The DataFrameSchema object consists of |column|_\s, |index|_\s, whether to
 |coerced|_ the types of all of the columns and |strict|_ which if True will
-error if the DataFrame contains columns that aren’t in the DataFrameSchema.
+raise a ``SchemaError`` if the DataFrame contains columns that aren’t in the
+DataFrameSchema.
 
 
 .. |column| replace:: ``Column``
@@ -18,28 +19,32 @@ error if the DataFrame contains columns that aren’t in the DataFrameSchema.
 .. |coerced| replace:: ``coerce``
 .. |strict| replace:: ``strict``
 
-.. code:: python
+.. testcode:: dataframe_schemas
 
-   schema = DataFrameSchema({
-       "column1": Column(Int),
-       "column2": Column(Float, Check(lambda s: s < -1.2)),
-       # you can provide a list of validators
-       "column3": Column(String, [
-           Check(lambda s: s.str.startswith("value")),
-           Check(lambda s: s.str.split("_", expand=True).shape[1] == 2)
-       ]),
-   },
-      strict = True,
-      coerce = True,
-   )
-   })
+    import pandera as pa
+
+    from pandera import Column, DataFrameSchema, Check
+
+    schema = DataFrameSchema(
+        {
+            "column1": Column(pa.Int),
+            "column2": Column(pa.Float, Check(lambda s: s < -1.2)),
+            # you can provide a list of validators
+            "column3": Column(pa.String, [
+               Check(lambda s: s.str.startswith("value")),
+               Check(lambda s: s.str.split("_", expand=True).shape[1] == 2)
+            ]),
+        },
+        strict=True,
+        coerce=True,
+    )
 
 .. _column:
 
 Column Validation
 -----------------
 
-A ``Column`` must contain a *type* to be validated. It can be optionally
+A ``Column`` must specify a *type* to be validated. It can be optionally
 verified for `null values`_ or duplicate values. The column can be coerced_ into
 the specified type, and the required_ parameter allows control over whether or
 not the column is allowed to be missing.
@@ -62,43 +67,47 @@ By default, SeriesSchema/Column objects assume that values are not
 nullable. In order to accept null values, you need to explicitly specify
 ``nullable=True``, or else you’ll get an error.
 
-.. code:: python
+.. testcode:: null_values_in_columns
 
    import numpy as np
    import pandas as pd
+   import pandera as pa
 
-   from pandera import Check, Column, DataFrameSchema, Int
+   from pandera import Check, Column, DataFrameSchema
 
    df = pd.DataFrame({"column1": [5, 1, np.nan]})
 
    non_null_schema = DataFrameSchema({
-       "column1": Column(Int, Check(lambda x: x > 0))
+       "column1": Column(pa.Int, Check(lambda x: x > 0))
    })
 
    non_null_schema.validate(df)
 
-   # SchemaError: non-nullable series contains null values: {2: nan}
+.. testoutput:: null_values_in_columns
+
+    Traceback (most recent call last):
+    ...
+    SchemaError: non-nullable series contains null values: {2: nan}
 
 .. note:: Due to a known limitation in
     `pandas <http://pandas.pydata.org/pandas-docs/stable/gotchas.html#support-for-integer-na>`__,
     integer arrays cannot contain ``NaN`` values, so this schema will return
     a dataframe where ``column1`` is of type ``float``.
 
-.. code:: python
+.. testcode:: null_values_in_columns
 
[…]
