+++ docs/source/decorators.rst
@@ -16,53 +16,61 @@ functions.
 Validates input pandas DataFrame/Series before entering the wrapped
 function.
 
-.. code:: python
+.. testcode:: check_input_decorators
 
-   import pandas as pd
+    import pandas as pd
+    import pandera as pa
 
-   from pandera import DataFrameSchema, Column, Check, Int, Float, check_input
+    from pandera import DataFrameSchema, Column, Check, check_input
 
 
-   df = pd.DataFrame({
+    df = pd.DataFrame({
        "column1": [1, 4, 0, 10, 9],
        "column2": [-1.3, -1.4, -2.9, -10.1, -20.4],
-   })
+    })
 
-   in_schema = DataFrameSchema({
-       "column1": Column(Int, Check(lambda x: 0 <= x <= 10)),
-       "column2": Column(Float, Check(lambda x: x < -1.2)),
-   })
+    in_schema = DataFrameSchema({
+       "column1": Column(pa.Int,
+                         Check(lambda x: 0 <= x <= 10, element_wise=True)),
+       "column2": Column(pa.Float, Check(lambda x: x < -1.2)),
+    })
+
+    # by default, check_input assumes that the first argument is
+    # dataframe/series.
+    @check_input(in_schema)
+    def preprocessor(dataframe):
+        dataframe["column3"] = dataframe["column1"] + dataframe["column2"]
+        return dataframe
+
+    preprocessed_df = preprocessor(df)
+    print(preprocessed_df)
+
+.. testoutput:: check_input_decorators
+
+       column1  column2  column3
+    0        1     -1.3     -0.3
+    1        4     -1.4      2.6
+    2        0     -2.9     -2.9
+    3       10    -10.1     -0.1
+    4        9    -20.4    -11.4
 
 
-   # by default, assumes that the first argument is dataframe/series.
-   @check_input(in_schema)
-   def preprocessor(dataframe):
-       dataframe["column4"] = dataframe["column1"] + dataframe["column2"]
-       return dataframe
+You can also provide the argument name as a string
 
+.. testcode:: check_input_decorators
 
-   # or you can provide the argument name as a string
-   @check_input(in_schema, "dataframe")
-   def preprocessor(dataframe):
-       ...
+    @check_input(in_schema, "dataframe")
+    def preprocessor(dataframe):
+        ...
 
+Or an integer representing the index in the positional arguments.
 
-   # or integer representing index in the positional arguments.
-   @check_input(in_schema, 1)
-   def preprocessor(foo, dataframe):
-       ...
+.. testcode:: check_input_decorators
 
+    @check_input(in_schema, 1)
+    def preprocessor(foo, dataframe):
+        ...
 
-   preprocessed_df = preprocessor(df)
-   print(preprocessed_df)
-
-   #  Output:
-   #     column1  column2  column3  column4
-   #  0        1     -1.3  value_1     -0.3
-   #  1        4     -1.4  value_2      2.6
-   #  2        0     -2.9  value_3     -2.9
-   #  3       10    -10.1  value_2     -0.1
-   #  4        9    -20.4  value_1    -11.4
 
 ``check_output``
 ~~~~~~~~~~~~~~~~
@@ -70,48 +78,53 @@ function.
 The same as ``check_input``, but this decorator checks the output
 DataFrame/Series of the decorated function.
 
-.. code:: python
+.. testcode:: check_output_decorators
 
-   from pandera import DataFrameSchema, Column, Check, Int, check_output
+    import pandas as pd
+    import pandera as pa
 
-
-   preprocessed_df = ...
-
-   # assert that all elements in "column1" are zero
-   out_schema = DataFrameSchema({
-       "column1": Column(Int, Check(lambda x: x == 0))
-   })
[…]
