### 2019-08-18 pandera-dev/pandera#90 @cosmicBboy
create conda-forge package
https://conda-forge.org/#contribute
