### 2019-08-19 9788012
adds type hinting
