+++ pandera/pandera.py
@@ -4,6 +4,7 @@ import inspect
 import warnings
 import pandas as pd
 import wrapt
+from typing import Union, Optional, List, Dict
 
 from collections import OrderedDict
 from enum import Enum
@@ -50,12 +51,12 @@ class Check(object):
 
     def __init__(
             self,
-            fn,
-            groups=None,
-            groupby=None,
-            element_wise=False,
-            error=None,
-            n_failure_cases=N_FAILURE_CASES):
+            fn: callable,
+            groups: Union[str, List[str], None] = None,
+            groupby: Union[str, List[str], callable, None] = None,
+            element_wise: Union[bool, List[bool]] = False,
+            error: Optional[str] = None,
+            n_failure_cases: Optional[int] = N_FAILURE_CASES):
         """Check object applies function element-wise or series-wise
 
         :param callable fn: A function to check series schema. If element_wise
@@ -64,7 +65,7 @@ class Check(object):
             to be: pd.Series -> bool|pd.Series[bool].
         :param groups: The dict input to the `fn` callable will be constrained
             to the groups specified by `groups`.
-        :type groups: str|list[str]|None
+        :type groups: str, List[str], None
         :param groupby: Only applies to Column Checks. If a string or list of
             strings is provided, then these columns are used to group the
             Column Series by `groupby`. If a callable is passed, the expected
@@ -78,12 +79,12 @@ class Check(object):
             dict[str|tuple[str], Series] -> bool|pd.Series[bool]
 
             Where specific groups can be obtained from the input dict.
-        :type groupby: str|list[str]|callable|None
+        :type groupby: str, List[str], callable, None
         :param element_wise: Whether or not to apply validator in an
             element-wise fashion. If bool, assumes that all checks should be
             applied to the column element-wise. If list, should be the same
             number of elements as checks.
-        :type element_wise: bool|list[bool]
+        :type element_wise: bool, List[bool]
         :param str error: custom error message if series fails validation
             check.
         :type str error:
@@ -110,46 +111,53 @@ class Check(object):
         self.groups = groups
 
     @property
-    def error_message(self):
+    def error_message(self) -> str:
 
         if self.error:
             return "%s: %s" % (self.fn.__name__, self.error)
         return "%s" % self.fn.__name__
 
     def vectorized_error_message(
-            self, parent_schema, check_index, failure_cases):
+            self,
+            parent_schema,
+            check_index: int,
+            failure_cases: pd.DataFrame) -> str:
         """Constructs an error message when an element-wise validator fails.
 
-        :param parent_schema: The schema object that is being checked and that
-            was inherited from the parent class.
-        :param check_index: The validator that failed.
-        :param failure_cases: The failure cases encountered by the element-wise
-            validator.
+        :param DataFrameSchema parent_schema: The schema object that is
+            being checked and that was inherited from the parent class.
+        :param int check_index: The validator that failed.
+        :param pd.DataFrame failure_cases: The failure cases encountered by the
+            element-wise validator.
 
         """
         return (
                 "%s failed element-wise validator %d:\n"
                 "%s\nfailure cases:\n%s" %
-                (parent_schema, check_index,
+                (parent_schema,
+                 check_index,
                  self.error_message,
                  self._format_failure_cases(failure_cases)))
 
-    def generic_error_message(self, parent_schema, check_index):
+    def generic_error_message(self,
+                              parent_schema,
+                              check_index: int) -> str:
         """Constructs an error message when a check validator fails.
 
[…]
