### 2019-08-20 889c523
Enabled Check to take callable objects as check functions
