### 2019-08-20 dad3474
Added ValueRange check class with unit tests
