### 2019-08-21 eba1c9c
Reformatting for flake8
