### 2019-08-21 fd1dc80
Implemented checks StringMatch and StringLength
