### 2019-08-21 9aa8cef
Basic reading of DataFrame schema from yaml. Checks are not yet parsed.
