+++ requirements.txt
@@ -5,6 +5,7 @@ numpy >= 1.9.0
 pandas >= 0.23.0
 pytest
 pytest-cov
+pyyaml
 scipy >= 1.2.0
 sphinx
 sphinx_rtd_theme
