### 2019-08-21 6a847d1
update typehints
