### 2019-08-21 4b0eecb
remove types from docstrings and DataFrameSchema types
