+++ pandera/pandera.py
@@ -78,15 +78,11 @@ class Check(object):
             dict[str|tuple[str], Series] -> bool|pd.Series[bool]
 
             Where specific groups can be obtained from the input dict.
-        :type groupby: str, List[str], callable, None
         :param element_wise: Whether or not to apply validator in an
             element-wise fashion. If bool, assumes that all checks should be
             applied to the column element-wise. If list, should be the same
             number of elements as checks.
-        :type element_wise: bool, List[bool]
-        :param str error: custom error message if series fails validation
-            check.
-        :type str error:
+        :param error: custom error message if series fails validation check.
         :param n_failure_cases: report the top n failure cases. If None, then
             report all failure cases.
         """
@@ -123,11 +119,11 @@ class Check(object):
             failure_cases: Union[pd.DataFrame, pd.Series]) -> str:
         """Constructs an error message when an element-wise validator fails.
 
-        :param DataFrameSchema parent_schema: The schema object that is
-            being checked and that was inherited from the parent class.
-        :param int check_index: The validator that failed.
-        :param pd.DataFrame failure_cases: The failure cases encountered by the
-            element-wise validator.
+        :param DataFrameSchema parent_schema: The schema object that is being
+            checked and that was inherited from the parent class.
+        :param check_index: The validator that failed.
+        :param failure_cases: The failure cases encountered by the element-wise
+            validator.
 
         """
         return (
@@ -143,9 +139,9 @@ class Check(object):
                               check_index: int) -> str:
         """Constructs an error message when a check validator fails.
 
-        :param DataFrameSchema parent_schema: The schema object that is
-            being checked and that was inherited from the parent class.
-        :param int check_index: The validator that failed.
+        :param DataFrameSchema parent_schema: The schema object that is being
+            checked and that was inherited from the parent class.
+        :param check_index: The validator that failed.
 
         """
         return "%s failed series validator %d: %s" % \
@@ -156,7 +152,7 @@ class Check(object):
                               ) -> pd.DataFrame:
         """Constructs readable error messages for vectorized_error_message.
 
-        :param pd.DataFrame failure_cases: The failure cases encountered by the
+        :param failure_cases: The failure cases encountered by the
             element-wise validator.
 
         """
@@ -205,7 +201,6 @@ class Check(object):
         :param groupby_obj:
         :param groups:
         :return: check_obj
-        :rtype: Dict[str, pd.Series]
         """
         if groups is None:
             return {group_key: group for group_key, group in groupby_obj}
@@ -225,14 +220,13 @@ class Check(object):
                              dataframe: pd.DataFrame) -> Dict[str, pd.Series]:
         """Prepare input for Column check.
 
-        :param pd.Series series: One-dimensional ndarray with axis labels
+        :param series: One-dimensional ndarray with axis labels
             (including time series).
-        :param pd.DataFrame dataframe: Two-dimensional size-mutable,
+        :param dataframe: Two-dimensional size-mutable,
             potentially heterogeneous tabular data structure with labeled axes
             (rows and columns)
         :return: a check_obj dictionary of pd.Series to be used by `_check_fn`
             and `_vectorized_series_check`
-        :rtype: dict[str, pd.Series]
 
         """
         if dataframe is None or self.groupby is None:
@@ -260,17 +254,16 @@ class Check(object):
         return self._format_input(groupby_obj, self.groups)
 
     def _vectorized_check(self,
-                          parent_schema: Union[DataFrameSchema, SeriesSchemaBase],
+                          parent_schema,
                           check_index: int,
                           check_obj: Dict[str, pd.Series]):
         """Perform a vectorized check on a series.
 
[…]
