### 2019-08-21 dd0dfc3
add autodoc typehints to pip requirements.txt
