### 2019-08-22 de284db
Moved example yaml for testing to separate file to satisfy flake8
