### 2019-08-22 6f25eb7
enables Check to take callable objects as check functions (#94)
* Enabled Check to take callable objects as check functions

* Reformatting for flake8
