### 2019-08-23 #12 @lwasser
## Editor checks:

- [x] **Automated tests:** Package has a testing suite and is tested via Travis-CI or another CI service. `Might add better dev setup instructions for contributing... but i see a dev envt txt`
- [x] **License:** The package has an OSI accepted license `MIT License`

---

## Editor comments

---

Reviewers: @mbjoseph @xmnlab
Due date: @mbjoseph we agreed to do reviews one at a time. Given that, is a 2 week deadline (which would be September 6) ok for your schedule? if that is ok then @xmnlab i will ping you once Max's review is in and you can begin your review!! @cosmicBboy has agreed to issues and PR's if you want to create a review using that approach rather than all text in this issue (links to the issue and/or PR may be preferred). Thank you all for your time!!
