### 2019-08-26 23e484c
modularize pandera into logical sets of functionality
this diff decomposes the single pandera.py file into logically
coherent modules:

- schemas for DataFrame and Series validations
- schema_components for objects used as part of schemas
- dtypes for datatypes
- errors
- decorators for `check_input` and `check_output`
- checks
- hypotheses
