+++ pandera/schemas.py
@@ -1,0 +1,322 @@
+"""Core pandera schema class definitions."""
+
+import pandas as pd
+
+from typing import Optional
+
+from . import errors, constants
+from .checks import Check
+
+
+class DataFrameSchema(object):
+    """A light-weight pandas DataFrame validator."""
+
+    def __init__(
+            self,
+            columns,
+            checks: callable = None,
+            index=None,
+            transformer: callable = None,
+            coerce: bool = False,
+            strict=False):
+        """Initialize DataFrameSchema validator.
+
+        :param columns: a dict where keys are column names and values are
+            Column objects specifying the datatypes and properties of a
+            particular column.
+        :type columns: dict[str -> Column]
+        :param checks: dataframe-wide checks.
+        :type checks: list[Check].
+        :param index: specify the datatypes and properties of the index.
+        :type index: Index
+        :param transformer: a callable with signature:
+            pandas.DataFrame -> pandas.DataFrame. If specified, calling
+            `validate` will verify properties of the columns and return the
+            transformed dataframe object.
+        :type transformer: callable
+        :param coerce: whether or not to coerce all of the columns on
+            validation.
+        :type coerce: bool
+        :param strict: whether or not to accept columns in the dataframe that
+            aren't in the DataFrameSchema.
+        :type strict: bool
+        """
+        if checks is None:
+            checks = []
+        if isinstance(checks, Check):
+            checks = [checks]
+        self._checks = checks
+        self.index = index
+        self.columns = columns
+        self.transformer = transformer
+        self.coerce = coerce
+        self.strict = strict
+        self._validate_schema()
+
+    def __call__(
+            self,
+            dataframe: pd.DataFrame,
+            head: int = None,
+            tail: int = None,
+            sample: int = None,
+            random_state: Optional[int] = None):
+        """Delegate to `validate` method.
+
+        :param pd.DataFrame dataframe: the dataframe to be validated.
+        :param head: validate the first n rows. Rows overlapping with `tail` or
+            `sample` are de-duplicated.
+        :type head: int
+        :param tail: validate the last n rows. Rows overlapping with `head` or
+            `sample` are de-duplicated.
+        :type tail: int
+        :param sample: validate a random sample of n rows. Rows overlapping
+            with `head` or `tail` are de-duplicated.
+        """
+        return self.validate(dataframe)
+
+    def _validate_schema(self):
+
+        for column_name, column in self.columns.items():
+            for check in column._checks:
+                if check.groupby is None or callable(check.groupby):
+                    continue
+                nonexistent_groupby_columns = [
+                    c for c in check.groupby if c not in self.columns]
+                if nonexistent_groupby_columns:
+                    raise errors.SchemaInitError(
+                        "groupby argument %s in Check for Column %s not "
+                        "specified in the DataFrameSchema." %
+                        (nonexistent_groupby_columns, column_name))
+
+    @staticmethod
+    def _dataframe_to_validate(
+            dataframe: pd.DataFrame,
+            head: int,
+            tail: int,
+            sample: int,
+            random_state: Optional[int]) -> pd.DataFrame:
+        dataframe_subsample = []
+        if head is not None:
+            dataframe_subsample.append(dataframe.head(head))
+        if tail is not None:
+            dataframe_subsample.append(dataframe.tail(tail))
+        if sample is not None:
+            dataframe_subsample.append(
+                dataframe.sample(sample, random_state=random_state))
+        return dataframe if len(dataframe_subsample) == 0 else \
+            pd.concat(dataframe_subsample).drop_duplicates()
+
+    def _check_dataframe(self, dataframe: pd.DataFrame):
+        return all(
+            check(self, check_index, check.prepare_dataframe_input(dataframe))
[…]
