+++ tests/test_pandera.py
@@ -4,16 +4,16 @@ import numpy as np
 import pandas as pd
 import pytest
 
-from pandera import Column, DataFrameSchema, Index, MultiIndex, PandasDtype, \
-    SeriesSchema, Check, Bool, Float, Int, DateTime, String, check_input, \
-    check_output, SchemaError, SchemaInitError, SchemaDefinitionError, \
-    Hypothesis
+from pandera import errors
+from pandera import Column, DataFrameSchema, Index, MultiIndex, \
+    SeriesSchema, Bool, Category, Check, DateTime, Float, Int, Object, \
+    String, Timedelta, check_input, check_output, Hypothesis
 from scipy import stats
 
 
 def test_column():
     schema = DataFrameSchema({
-        "a": Column(PandasDtype.Int, Check(lambda x: x > 0, element_wise=True))
+        "a": Column(Int, Check(lambda x: x > 0, element_wise=True))
     })
     data = pd.DataFrame({"a": [1, 2, 3]})
     assert isinstance(schema.validate(data), pd.DataFrame)
@@ -21,13 +21,13 @@ def test_column():
 
 def test_series_schema():
     schema = SeriesSchema(
-        PandasDtype.Int, Check(lambda x: 0 <= x <= 100, element_wise=True))
+        Int, Check(lambda x: 0 <= x <= 100, element_wise=True))
     validated_series = schema.validate(pd.Series([0, 30, 50, 100]))
     assert isinstance(validated_series, pd.Series)
 
     # error cases
     for data in [-1, 101, 50.1, "foo"]:
-        with pytest.raises(SchemaError):
+        with pytest.raises(errors.SchemaError):
             schema.validate(pd.Series([data]))
 
     for data in [-1, {"a": 1}, -1.0]:
@@ -35,55 +35,55 @@ def test_series_schema():
             schema.validate(TypeError)
 
     non_duplicate_schema = SeriesSchema(
-        PandasDtype.Int, allow_duplicates=False)
-    with pytest.raises(SchemaError):
+        Int, allow_duplicates=False)
+    with pytest.raises(errors.SchemaError):
         non_duplicate_schema.validate(pd.Series([0, 1, 2, 3, 4, 1]))
 
 
 def test_vectorized_checks():
     schema = SeriesSchema(
-        PandasDtype.Int, Check(
+        Int, Check(
             lambda s: s.value_counts() == 2, element_wise=False))
     validated_series = schema.validate(pd.Series([1, 1, 2, 2, 3, 3]))
     assert isinstance(validated_series, pd.Series)
 
     # error case
-    with pytest.raises(SchemaError):
+    with pytest.raises(errors.SchemaError):
         schema.validate(pd.Series([1, 2, 3]))
 
 
 def test_series_schema_multiple_validators():
     schema = SeriesSchema(
-        PandasDtype.Int, [
+        Int, [
             Check(lambda x: 0 <= x <= 50, element_wise=True),
             Check(lambda s: (s == 21).any())])
     validated_series = schema.validate(pd.Series([1, 5, 21, 50]))
     assert isinstance(validated_series, pd.Series)
 
     # raise error if any of the validators fails
-    with pytest.raises(SchemaError):
+    with pytest.raises(errors.SchemaError):
         schema.validate(pd.Series([1, 5, 20, 50]))
 
 
 def test_dataframe_schema():
     schema = DataFrameSchema(
         {
-            "a": Column(PandasDtype.Int,
+            "a": Column(Int,
                         Check(lambda x: x > 0, element_wise=True)),
-            "b": Column(PandasDtype.Float,
+            "b": Column(Float,
                         Check(lambda x: 0 <= x <= 10, element_wise=True)),
[…]
