+++ pandera/errors.py
@@ -1,0 +1,13 @@
+"""pandera-specific errors."""
+
+
+class SchemaInitError(Exception):
+    pass
+
+
+class SchemaDefinitionError(Exception):
+    pass
+
+
+class SchemaError(Exception):
+    pass
