### 2019-08-27 af7e86d
add type hint to MultiIndex call
