### 2019-08-27 f67050d
adds type hinting (#92)
* adds type hinting

* update typehints

* remove types from docstrings and DataFrameSchema types

* add autodoc typehints to pip requirements.txt

* add type hint to MultiIndex call
