### 2019-08-27 e9f60e5
Flexible handling of 'dtype' and 'pandas_dtype' in yaml
