### 2019-08-29 050ee61
refactor for clarity
