### 2019-08-31 ef6cfe1
more dtype support: int and float types
