### 2019-09-05 360add7
Refactoring checks to functional format (ongoing)
