### 2019-09-10 bc9279d
add citation information to the README.md
