+++ conda.recipe/meta.yaml
@@ -4,7 +4,7 @@ package:
 
 source:
   git_rev: 0.1.5
-  git_url: https://github.com/cosmicBboy/pandera.git
+  git_url: https://github.com/pandera-dev/pandera.git
 
 build:
   noarch: python
@@ -30,7 +30,7 @@ test:
     - pandera
 
 about:
-  home: https://github.com/cosmicBboy/pandera
+  home: https://github.com/pandera-dev/pandera
   license: MIT
   summary: 'Light-weight and flexible validation for pandas data structures.'
   description: |
