### 2019-09-10 077141d
add dev installation instructions in README.md
