### 2019-09-10 b01eef0
more dtype support: int and float types (#101)
