+++ tests/test_pandera.py
@@ -8,6 +8,7 @@ from pandera import errors
 from pandera import Column, DataFrameSchema, Index, MultiIndex, \
     SeriesSchema, Bool, Category, Check, DateTime, Float, Int, Object, \
     String, Timedelta, check_input, check_output, Hypothesis
+from pandera import dtypes
 from scipy import stats
 
 
@@ -384,12 +385,42 @@ def test_check_input_method_decorators():
         transformer.transform_secord_arg_with_dict_getter(None, dataframe))
 
 
-def test_string_dtypes():
-    # TODO: add tests for all datatypes
-    schema = DataFrameSchema(
-        {"col": Column("float64", nullable=True)})
-    df = pd.DataFrame({"col": [np.nan, 1.0, 2.0]})
-    assert isinstance(schema.validate(df), pd.DataFrame)
+def test_dtypes():
+    for dtype in [
+            dtypes.Float,
+            dtypes.Float16,
+            dtypes.Float32,
+            dtypes.Float64]:
+        schema = DataFrameSchema({"col": Column(dtype, nullable=False)})
+        validated_df = schema.validate(
+            pd.DataFrame(
+                {"col": [-123.1, -7654.321, 1.0, 1.1, 1199.51, 5.1, 4.6]},
+                dtype=dtype.value))
+        assert isinstance(validated_df, pd.DataFrame)
+
+    for dtype in [
+            dtypes.Int,
+            dtypes.Int8,
+            dtypes.Int16,
+            dtypes.Int32,
+            dtypes.Int64]:
+        schema = DataFrameSchema({"col": Column(dtype, nullable=False)})
+        validated_df = schema.validate(
+            pd.DataFrame(
+                {"col": [-712, -4, -321, 0, 1, 777, 5, 123, 9000]},
+                dtype=dtype.value))
+        assert isinstance(validated_df, pd.DataFrame)
+
+    for dtype in [
+            dtypes.UInt8,
+            dtypes.UInt16,
+            dtypes.UInt32,
+            dtypes.UInt64]:
+        schema = DataFrameSchema({"col": Column(dtype, nullable=False)})
+        validated_df = schema.validate(
+            pd.DataFrame(
+                {"col": [1, 777, 5, 123, 9000]}, dtype=dtype.value))
+        assert isinstance(validated_df, pd.DataFrame)
 
 
 def test_nullable_int():
