### 2019-09-10 7310d0d
improve README and sphinx documentation (#105)
