+++ docs/source/API.rst
@@ -3,10 +3,64 @@
 API
 ===
 
-Module contents
----------------
+``pandera.schemas``
+-------------------
 
-.. automodule:: pandera
+.. automodule:: pandera.schemas
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+
+``pandera.schema_components``
+-----------------------------
+
+.. automodule:: pandera.schema_components
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+
+``pandera.checks``
+------------------
+
+.. automodule:: pandera.checks
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+
+``pandera.hypotheses``
+----------------------
+
+.. automodule:: pandera.hypotheses
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+
+``pandera.decorators``
+----------------------
+
+.. automodule:: pandera.decorators
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+
+``pandera.dtypes``
+------------------
+
+.. automodule:: pandera.dtypes
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+
+``pandera.errors``
+------------------
+
+.. automodule:: pandera.errors
    :members:
    :undoc-members:
    :show-inheritance:
