### 2019-09-10 fc071d6
make error message formatting functions private (#106)
