### 2019-09-10 ed4e8f8
add docstrings to error classes (#107)
