### 2019-09-10 bdbaa6e
add human-readable __str__ and __repr__ methods to dataframeschema
