### 2019-09-10 fd5cf99
remove gt/lt sign in __str__ method
