### 2019-09-10 954d8a2
add human-readable __str__ and __repr__ methods to dataframeschema (#111)
* add human-readable __str__ and __repr__ methods to dataframeschema

* remove gt/lt sign in __str__ method
