### 2019-09-10 db87b21
Update requirements.txt (#100)
removes enum34 which is no longer needed when Python 2 support is dropped in v2.0
