+++ requirements.txt
@@ -1,5 +1,4 @@
 codecov
-enum34
 flake8
 numpy >= 1.9.0
 pandas >= 0.23.0
