### 2019-09-11 4e1f828
update conda recipe to 0.2.0
