### 2019-09-11 d8bc901
add setup_py_install: True option in .readthedocs.yml
