+++ .readthedocs.yml
@@ -19,5 +19,6 @@ formats: all
 # Optionally set the version of Python and requirements required to build your docs
 python:
   version: 3.7
+  setup_py_install: true
   install:
     - requirements: requirements.txt
