### 2019-09-11 ef1aab6
fix .readthedocs.yml, use v2 syntax
https://docs.readthedocs.io/en/stable/config-file/v2.html#python
