### 2019-09-11 3ae8aab
bump to v0.2.1: bugfix on sphinx docs
