### 2019-09-11 #12 @cosmicBboy
thanks @lwasser!

@mbjoseph thanks for the review! I've released v0.2.1, where I addressed many of the points that you raised in the review, check out the [release notes](https://github.com/pandera-dev/pandera/releases/tag/v0.2.0).

## Most notable changes:

- add citation information
- add dev installation instructions
- improve formatting and wording of sphinx documentation (this addresses several of the points you made about formatting and wording in the documentation)
- make SchemaError message formatting functions private (`generic_error_message` and this other methods should have been private all along)
- add docstrings to error classes

## Minor points:

> Test coverage is pretty high - any particular reason why the remaining lines are not tested?

I haven't really had to much time to prioritize covering the rest, though I'd like to prioritize the biggest holes and cover those.

> There are some deprecation warnings that arise in running the tests: https://travis-ci.org/pandera-dev/pandera/jobs/579197344#L2287

Planning to do this as part of https://github.com/pandera-dev/pandera/issues/110

> CI testing on OSX and Windows might be nice too.

Made an issue for this: https://github.com/pandera-dev/pandera/issues/109

> Why not conda-forge instead of the cosmicbboy conda channel?

Yes, would love to get a conda-forge recipe going: https://github.com/pandera-dev/pandera/issues/90

> pylint points out some places where the code could be streamlined a bit (e.g., unnecessary else statements, and some cases where object is explicitly declared as a parent class), but none of the output is indicative of major problems. Feel free to address or ignore any of these checks:

Cool, made an issue to add pylint to CI: https://github.com/pandera-dev/pandera/issues/108
