### 2019-09-11 #12 @xmnlab
just one question. the version submitted for review is `0.1.5`
but it seems pandera has more 2 version after that.

should I review just `0.1.5`? the same applies to documentation on readthedocs?
