### 2019-09-14 #12 @xmnlab
- [x] **Function Documentation:** for all user-facing functions
- [ ] **Examples** for all user-facing functions
- [x] **Community guidelines** including contribution guidelines in the README or CONTRIBUTING, and DESCRIPTION with `URL`, `BugReports` and `Maintainer`.

Readme requirements

- [x] Badges for continuous integration and test coverage, the badge for pyOpenSci peer-review once it has started (see below), a repostatus.org badge, and any other badges. If the README has many more badges, you might want to consider using a table for badges, see this example, that one and that one. Such a table should be more wide than high.
- [x] Short description of goals of package, with descriptive links to all vignettes (rendered, i.e. readable, cf the documentation website section) unless the package is small and there’s only one vignette repeating the README.
- [x] Any additional setup required (authentication tokens, etc)
- [x] Brief demonstration usage
- [x] Direction to more detailed documentation (e.g. your documentation files or website).
- [x] If applicable, how the package compares to other similar packages and/or how it relates to other packages

- [x] **Continuous Integration:** Has continuous integration, such as Travis CI, AppVeyor, CircleCI, and/or others.

Estimated hours spent reviewing: 4:30

---

The package looks very good: package structure, documentation, tests and CI looks in very good shape. Some points reported by @mbjoseph were already fixed or already added as an GitHub issue.

I am adding just 2 more comments. Actually the 1st is just a comment related to an issue that was already partial fixed (installation for development) but maybe it could be improved.

- **Installation instructions:**: probably the documentation should recommend `python setup.py develop` or `pip install -e .` for the installation in development mode (as @mbjoseph  suggested)
- **Examples**: maybe it should consider the usage of example sections for docstrings. It seems the project is using sphinx style for docstrings. I didn't find an official documentation for that but maybe it could help: http://queirozf.com/entries/python-docstrings-reference-examples
