### 2019-09-27 18a3fe7
Make dtype optional
