+++ pandera/schemas.py
@@ -7,6 +7,7 @@ import pandas as pd
 from typing import Optional
 
 from . import errors, constants
+from .dtypes import PandasDtype
 from .checks import Check
 
 
@@ -220,7 +221,7 @@ class SeriesSchemaBase(object):
 
     def __init__(
             self,
-            pandas_dtype,
+            pandas_dtype: PandasDtype = None,
             checks: callable = None,
             nullable: bool = False,
             allow_duplicates: bool = True,
@@ -269,8 +270,11 @@ class SeriesSchemaBase(object):
             raise errors.SchemaError(
                 "Expected %s to have name '%s', found '%s'" %
                 (type(self), self._name, series.name))
-        expected_dtype = _dtype = self._pandas_dtype if \
-            isinstance(self._pandas_dtype, str) else self._pandas_dtype.value
+
+        expected_dtype = _dtype = self._pandas_dtype if (
+            isinstance(self._pandas_dtype, str) or self._pandas_dtype is None
+        ) else self._pandas_dtype.value
+
         if self._nullable:
             series = series.dropna()
             if dataframe_context is not None:
@@ -313,7 +317,7 @@ class SeriesSchemaBase(object):
                      series[duplicates].head(
                         constants.N_FAILURE_CASES).to_dict()))
 
-        if series.dtype != _dtype:
+        if _dtype is not None and series.dtype != _dtype:
             raise errors.SchemaError(
                 "expected series '%s' to have type %s, got %s" %
                 (series.name, expected_dtype, series.dtype))
@@ -332,7 +336,7 @@ class SeriesSchema(SeriesSchemaBase):
 
     def __init__(
             self,
-            pandas_dtype,
+            pandas_dtype: PandasDtype = None,
             checks: callable = None,
             nullable: bool = False,
             allow_duplicates: bool = True,
