### 2019-09-27 23ef65e
Add assert for series
