### 2019-09-29 f7e8d5c
add examples to all public classes/methods
- use `pip install -e .` for dev install instructions
