+++ pandera/hypotheses.py
@@ -5,7 +5,7 @@ import pandas as pd
 from functools import partial
 
 from scipy import stats
-from typing import Union, Optional, List, Dict
+from typing import Callable, Union, Optional, List, Dict
 
 from . import errors
 from .checks import Check
@@ -15,9 +15,9 @@ DEFAULT_ALPHA = 0.01
 
 
 class Hypothesis(Check):
-    """Extends Check to perform a hypothesis test on a Column or DataFrame."""
+    """Extends Check to perform a hypothesis test on a Column."""
 
-    RELATIONSHIPS = {
+    _RELATIONSHIPS = {
         "greater_than": (lambda stat, pvalue, alpha=DEFAULT_ALPHA:
                          stat > 0 and pvalue / 2 < alpha),
         "less_than": (lambda stat, pvalue, alpha=DEFAULT_ALPHA:
@@ -28,16 +28,16 @@ class Hypothesis(Check):
     }
 
     def __init__(self,
-                 test: callable,
-                 samples: Union[str, List[str], None],
-                 groupby: Union[str, List[str], callable, None] = None,
-                 relationship: Union[str, callable] = "equal",
+                 test: Callable,
+                 samples: Optional[Union[str, List[str]]] = None,
+                 groupby: Optional[Union[str, List[str], Callable]] = None,
+                 relationship: Union[str, Callable] = "equal",
                  test_kwargs: Dict = None,
                  relationship_kwargs: Dict = None,
                  error: Optional[str] = None):
-        """Initialises hypothesis to perform a hypothesis test on a Column.
+        """Perform a hypothesis test on a Series or DataFrame.
 
-            Can function on a single column or be grouped by another column.
+        Can function on a single column or be grouped by another column.
 
         :param callable test: A function to check a series schema.
         :param samples: for `Column` or `SeriesSchema` hypotheses, this refers
@@ -49,7 +49,6 @@ class Hypothesis(Check):
             multiple columns to pass into the `test` function. The `samples`
             column(s) are passed into the `test`  function as positional
             arguments.
-        :type samples: str|list[str]|None
         :param groupby: If a string or list of strings is provided, then these
             columns are used to group the Column Series by `groupby`. If a
             callable is passed, the expected signature is
@@ -62,39 +61,78 @@ class Hypothesis(Check):
             dict[str|tuple[str], Series] -> bool|pd.Series[bool]
 
             Where specific groups can be obtained from the input dict.
-        :type groupby: str|list[str]|callable|None
         :param relationship: Represents what relationship conditions are
             imposed on the hypothesis test. A function or lambda function can
             be supplied.
 
-            If a string is provided, a lambda function will be returned from
-            Hypothesis.relationships. Available relationships are:
-            "greater_than", "less_than", "not_equal" or "equal".
+            Available built-in relationships are: "greater_than", "less_than",
+            "not_equal" or "equal", where "equal" is the null hypothesis.
 
             If callable, the input function signature should have the signature
-            `(stat: float, pvalue: float, **kwargs)` where `stat` is the
+            ``(stat: float, pvalue: float, **kwargs)`` where `stat` is the
             hypothesis test statistic, `pvalue` assesses statistical
-            significance, and `**kwargs` are other arguments supplied by the
+            significance, and `**kwargs` are other arguments supplied bia the
             `**relationship_kwargs` argument.
 
             Default is "equal" for the null hypothesis.
-
-        :type relationship: str|callable
         :param dict test_kwargs: Key Word arguments to be supplied to the test.
         :param dict relationship_kwargs: Key Word arguments to be supplied to
             the relationship function. e.g. `alpha` could be used to specify a
[…]
