+++ pandera/schemas.py
@@ -4,9 +4,9 @@ import json
 
 import pandas as pd
 
-from typing import Optional
+from typing import List, Optional, Union
 
-from . import errors, constants
+from . import errors, constants, dtypes
 from .checks import Check
 
 
@@ -19,7 +19,7 @@ class DataFrameSchema(object):
     def __init__(
             self,
             columns,
-            checks: callable = None,
+            checks: Optional[List[Check]] = None,
             index=None,
             transformer: callable = None,
             coerce: bool = False,
@@ -29,22 +29,50 @@ class DataFrameSchema(object):
         :param columns: a dict where keys are column names and values are
             Column objects specifying the datatypes and properties of a
             particular column.
-        :type columns: dict[str -> Column]
+        :type columns: Dict[str, Column]
         :param checks: dataframe-wide checks.
-        :type checks: list[Check].
         :param index: specify the datatypes and properties of the index.
-        :type index: Index
         :param transformer: a callable with signature:
             pandas.DataFrame -> pandas.DataFrame. If specified, calling
             `validate` will verify properties of the columns and return the
             transformed dataframe object.
-        :type transformer: callable
         :param coerce: whether or not to coerce all of the columns on
             validation.
-        :type coerce: bool
         :param strict: whether or not to accept columns in the dataframe that
             aren't in the DataFrameSchema.
-        :type strict: bool
+
+        :examples:
+
+        >>> import pandera as pa
+        >>> from pandera import DataFrameSchema, Column
+        >>>
+        >>> schema = DataFrameSchema({
+        ...     "str_column": Column(pa.String),
+        ...     "float_column": Column(pa.Float),
+        ...     "int_column": Column(pa.Int),
+        ...     "date_column": Column(pa.DateTime),
+        ... })
+
+        Use the pandas API to define checks, which takes a function with
+        the signature: ``pd.Series -> Union[bool, pd.Series[bool]]``
+
+        >>> from pandera import Check
+        >>>
+        >>> schema_with_checks = DataFrameSchema({
+        ...     "probability": Column(
+        ...         pa.Float, Check(lambda s: (s >= 0) & (s <= 1))),
+        ...
+        ...     # check that the "category" column contains a few discrete
+        ...     # values, and the majority of the entries are dogs.
+        ...     "category": Column(
+        ...         pa.String, [
+        ...             Check(lambda s: s.isin(["dog", "cat", "duck"])),
+        ...             Check(lambda s: (s == "dog").mean() > 0.5),
+        ...         ]),
+        ... })
+
+        See :ref:`here<DataFrameSchemas>` for more usage details.
+
         """
         if checks is None:
             checks = []
@@ -62,9 +90,9 @@ class DataFrameSchema(object):
     def __call__(
             self,
             dataframe: pd.DataFrame,
-            head: int = None,
-            tail: int = None,
-            sample: int = None,
+            head: Optional[int] = None,
+            tail: Optional[int] = None,
+            sample: Optional[int] = None,
             random_state: Optional[int] = None):
         """Delegate to `validate` method.
 
@@ -95,7 +123,7 @@ class DataFrameSchema(object):
 
     def _set_column_names(self):
         self.columns = {
-            column_name: column.set_name(column_name)
+            column_name: column._set_name(column_name)
             for column_name, column in self.columns.items()
         }
 
@@ -124,7 +152,7 @@ class DataFrameSchema(object):
                 check(
                     self,
                     check_index,
-                    check.prepare_dataframe_input(dataframe)))
+                    check._prepare_dataframe_input(dataframe)))
[…]

[…]
