### 2019-09-29 300ea2b
fix docstring tests
