### 2019-09-29 78d0c74
improve series schema test coverage
