### 2019-09-29 4eeeb16
add examples to all public classes/methods (#115)
* add examples to all public classes/methods

- use `pip install -e .` for dev install instructions

* fix tests

* fix test warnings

* fix docstring tests

* improve series schema test coverage

* fix docstring
