+++ pandera/schema_components.py
@@ -1,8 +1,11 @@
 """Components used in pandera schemas."""
 
+from typing import Union
+
 import pandas as pd
 
 from .dtypes import PandasDtype
+from .checks import Check, List
 from .schemas import DataFrameSchema, SeriesSchemaBase
 
 
@@ -10,32 +13,42 @@ class Column(SeriesSchemaBase):
 
     def __init__(
             self,
-            pandas_dtype,
-            checks: callable = None,
+            pandas_dtype: Union[str, PandasDtype],
+            checks: Union[Check, List[Check]] = None,
             nullable: bool = False,
             allow_duplicates: bool = True,
             coerce: bool = False,
             required: bool = True):
-        """Initialize column validator object.
+        """Create column validator object.
 
         :param pandas_dtype: datatype of the column. If a string is specified,
             then assumes one of the valid pandas string values:
             http://pandas.pydata.org/pandas-docs/stable/basics.html#dtypes
         :type pandas_dtype: str|PandasDtype
-        :param checks: if element_wise is True, then callable signature should
-            be: x -> bool where x is a scalar element in the column. Otherwise,
-            x is assumed to be a pandas.Series object.
-        :type checks: callable
+        :param checks: checks to verify validity of the column
         :param nullable: Whether or not column can contain null values.
-        :type nullable: bool
         :param allow_duplicates: Whether or not to coerce the column to the
             specified pandas_dtype before validation
-        :type allow_duplicates: bool
         :param coerce: If True, when schema.validate is called the column will
             be coerced into the specified dtype.
-        :type coerce:  bool
         :param required: Whether or not column is allowed to be missing
-        :type required:  bool
+
+        :example:
+
+        >>> import pandas as pd
+        >>> import pandera as pa
+        >>> from pandera import DataFrameSchema, Column
+        >>>
+        >>> schema = DataFrameSchema({
+        ...     "column": Column(pa.String)
+        ... })
+        >>>
+        >>> schema.validate(pd.DataFrame({"column": ["foo", "bar"]}))
+          column
+        0    foo
+        1    bar
+
+        See :ref:`here<column>` for more usage details.
         """
         super(Column, self).__init__(
             pandas_dtype, checks, nullable, allow_duplicates)
@@ -46,7 +59,7 @@ class Column(SeriesSchemaBase):
     def _allow_groupby(self):
         return True
 
-    def set_name(self, name):
+    def _set_name(self, name):
         """Used to set or modify the name of a column object.
 
         :param str name: the name of the column object
@@ -55,7 +68,7 @@ class Column(SeriesSchemaBase):
         self._name = name
         return self
 
-    def coerce_dtype(self, series):
+    def _coerce_dtype(self, series):
         """Coerce the type of a pd.Series by the type specified in the Column
             object's self._pandas_dtype
 
@@ -86,11 +99,43 @@ class Index(SeriesSchemaBase):
 
     def __init__(
             self,
-            pandas_dtype,
-            checks: callable = None,
+            pandas_dtype: Union[str, PandasDtype],
+            checks: Union[Check, List[Check]] = None,
             nullable: bool = False,
             allow_duplicates: bool = True,
             name: str = None):
+        """Create Index validator.
+
+        :param pandas_dtype: datatype of the column. If a string is specified,
+            then assumes one of the valid pandas string values:
+            http://pandas.pydata.org/pandas-docs/stable/basics.html#dtypes
+        :param checks: checks to verify validity of the index.
+        :param nullable: Whether or not column can contain null values.
+        :param allow_duplicates: Whether or not to coerce the column to the
[…]
