### 2019-09-29 ae9c735
fix dataframeschema docstring
