### 2019-09-29 bb5811c
bump to v0.2.2: improve documentation
