### 2019-09-29 #12 @cosmicBboy
thanks @lwasser @xmnlab @mbjoseph!

I've cut a new pandera release 0.2.2 that adds [example docstrings](https://github.com/pandera-dev/pandera/commit/4eeeb16541e5d9e74748fe9d1307065a8902a6b1) to all public-facing classes and methods. The commit also:

- docstring examples should be reflected in the [docs](https://pandera.readthedocs.io/en/stable/API.html)
- changes README with updated development installation instructions.
- adds more test coverage in `schema.py`
- fixes unit test pandas FutureDeprecation warnings

Please let me know if you have any questions.
