### 2019-09-30 b00da64
Raise exception on invalid __init__ parameters
