### 2019-09-30 #12 @lwasser
thank you @cosmicBboy  !! @mbjoseph @xmnlab will you please have a look at the latest release? let me know if the changes are acceptable given your review! if so, you can check the. "the author has responded to my review" box at the bottom of your review submission. If you see anything that wasn't addressed to your satisfaction please let me know!!

thank you all for such a smooth review process!
