### 2019-09-30 #12 @mbjoseph
@cosmicBboy thanks for addressing my suggestions - v0.2.2 looks good to me!
