### 2019-10-01 #12 @xmnlab
@cosmicBboy good job with the examples in the documentation!

I have checked the my "Final approval" checkbox.

Just some minors observations:

1. There are some public methods without type hinting and/or docstring for `return`, if you could open an issue for that and address that it would be great.
2. And there are some private methods without any docstring or type hinting.

thanks for working on the suggestions we made!
