### 2019-10-01 #12 @lwasser
this is great. @cosmicBboy @xmnlab are these small things that could be done quickly? it would be nice to have that done and wrapped up before the final approval. pyopensci does value documentation highly!
