### 2019-10-02 d219d9a
Make dtype optional (#114)
* Make dtype optional

* Add assert for series

* Raise exception on invalid __init__ parameters
