### 2019-10-02 e2173c7
clean up docstring and typing of functions/methods
