+++ pandera/checks.py
@@ -3,18 +3,23 @@
 import pandas as pd
 
 from functools import partial
-from typing import Any, Union, Optional, List, Dict, Callable
+from typing import Union, Optional, List, Dict, Callable
 
 from . import errors, constants
 from .dtypes import PandasDtype
 
 
-CheckCallable = Callable[
-    [Union[pd.Series, Any]], Union[pd.Series, bool]
+GroupbyObject = Union[
+    pd.core.groupby.SeriesGroupBy,
+    pd.core.groupby.DataFrameGroupBy
 ]
+
+SeriesCheckObj = Union[pd.Series, Dict[str, pd.Series]]
+DataFrameCheckObj = Union[pd.DataFrame, Dict[str, pd.DataFrame]]
 
 
 class Check(object):
+    """Check a pandas Series or DataFrame for certain properties."""
 
     def __init__(
             self,
@@ -142,6 +147,7 @@ class Check(object):
 
     @property
     def _error_message(self):
+        """Check error message."""
         name = getattr(self.fn, '__name__', self.fn.__class__.__name__)
         if self.error:
             return "%s: %s" % (name, self.error)
@@ -149,13 +155,12 @@ class Check(object):
 
     def _vectorized_error_message(
             self,
-            parent_schema,
+            parent_schema: type,
             check_index: int,
             failure_cases: Union[pd.DataFrame, pd.Series]) -> str:
-        """Constructs an error message when an element-wise validator fails.
+        """Construct an error message when an element-wise validator fails.
 
-        :param parent_schema: The schema object that is being checked and that
-            was inherited from the parent class.
+        :param parent_schema: class of schema being validated.
         :param check_index: The validator that failed.
         :param failure_cases: The failure cases encountered by the element-wise
             validator.
@@ -170,12 +175,11 @@ class Check(object):
 
     def _generic_error_message(
             self,
-            parent_schema,
+            parent_schema: type,
             check_index: int) -> str:
-        """Constructs an error message when a check validator fails.
+        """Construct an error message when a check validator fails.
 
-        :param parent_schema: The schema object that is being checked and that
-            was inherited from the parent class.
+        :param parent_schema: class of schema being validated.
         :param check_index: The validator that failed.
 
         """
@@ -185,10 +189,14 @@ class Check(object):
     def _format_failure_cases(
             self,
             failure_cases: Union[pd.DataFrame, pd.Series]) -> pd.DataFrame:
-        """Constructs readable error messages for vectorized_error_message.
+        """Construct readable error messages for vectorized_error_message.
 
         :param failure_cases: The failure cases encountered by the element-wise
             validator.
+        :returns: DataFrame where index contains failure cases, the "index"
+            column contains a list of integer indexes in the validation
+            DataFrame that caused the failure, and a "count" column
+            representing how many failures of that case occurred.
 
         """
         if isinstance(failure_cases.index, pd.MultiIndex):
@@ -227,10 +235,17 @@ class Check(object):
         self.failure_cases = failure_cases
         return failure_cases.head(self.n_failure_cases)
 
-    def _format_input(
+    def _format_groupby_input(
             self,
-            groupby_obj,
-            groups) -> Union[Dict[str, Union[pd.Series, pd.DataFrame]]]:
+            groupby_obj: GroupbyObject,
+            groups: List[str]
+            ) -> Union[Dict[str, Union[pd.Series, pd.DataFrame]]]:
+        """Format groupby object into dict of groups to Series or DataFrame.
+
+        :param groupby_obj: a pandas groupby object.
+        :param groups: only include these groups in the output.
+        :returns: dictionary mapping group names to Series or DataFrame.
+        """
[…]
