### 2019-10-04 e4aa07a
add coerce logic to SeriesSchema, modularize tests
