+++ tests/test_pandera.py
@@ -1,1107 +1,0 @@
-"""Some unit tests."""
-
-import numpy as np
-import pandas as pd
-import pytest
-
-from pandera import errors
-from pandera import Column, DataFrameSchema, Index, MultiIndex, \
-    SeriesSchema, Bool, Category, Check, DateTime, Float, Int, Object, \
-    String, Timedelta, check_input, check_output, Hypothesis
-from pandera import dtypes
-from scipy import stats
-
-
-def test_column():
-    schema = DataFrameSchema({
-        "a": Column(Int, Check(lambda x: x > 0, element_wise=True))
-    })
-    data = pd.DataFrame({"a": [1, 2, 3]})
-    assert isinstance(schema.validate(data), pd.DataFrame)
-
-
-def test_series_schema():
-    schema = SeriesSchema(
-        Int, Check(lambda x: 0 <= x <= 100, element_wise=True))
-    validated_series = schema.validate(pd.Series([0, 30, 50, 100]))
-    assert isinstance(validated_series, pd.Series)
-
-    # error cases
-    for data in [-1, 101, 50.1, "foo"]:
-        with pytest.raises(errors.SchemaError):
-            schema.validate(pd.Series([data]))
-
-    for data in [-1, {"a": 1}, -1.0]:
-        with pytest.raises(TypeError):
-            schema.validate(TypeError)
-
-    non_duplicate_schema = SeriesSchema(
-        Int, allow_duplicates=False)
-    with pytest.raises(errors.SchemaError):
-        non_duplicate_schema.validate(pd.Series([0, 1, 2, 3, 4, 1]))
-
-    # when series name doesn't match schema
-    named_schema = SeriesSchema(Int, name="my_series")
-    with pytest.raises(
-            errors.SchemaError,
-            match=r"^Expected .+ to have name"):
-        named_schema.validate(pd.Series(range(5), name="your_series"))
-
-    # when series floats are declared to be integer
-    with pytest.raises(
-            errors.SchemaError,
-            match=r"^after dropping null values, expected values in series"):
-        SeriesSchema(Int, nullable=True).validate(
-            pd.Series([1.1, 2.3, 5.5, np.nan]))
-
-    # when series contains null values when schema is not nullable
-    with pytest.raises(
-            errors.SchemaError,
-            match=r"^non-nullable series .+ contains null values"):
-        SeriesSchema(Float, nullable=False).validate(
-            pd.Series([1.1, 2.3, 5.5, np.nan]))
-
-    # when series contains null values when schema is not nullable in addition
-    # to having the wrong data type
-    with pytest.raises(
-            errors.SchemaError,
-            match=(
-                r"^expected series '.+' to have type .+, got .+ and "
-                "non-nullable series contains null values")):
-        SeriesSchema(Int, nullable=False).validate(
-            pd.Series([1.1, 2.3, 5.5, np.nan]))
-
-
-def test_vectorized_checks():
-    schema = SeriesSchema(
-        Int, Check(
-            lambda s: s.value_counts() == 2, element_wise=False))
-    validated_series = schema.validate(pd.Series([1, 1, 2, 2, 3, 3]))
-    assert isinstance(validated_series, pd.Series)
-
-    # error case
-    with pytest.raises(errors.SchemaError):
-        schema.validate(pd.Series([1, 2, 3]))
-
-
-def test_series_schema_multiple_validators():
-    schema = SeriesSchema(
-        Int, [
-            Check(lambda x: 0 <= x <= 50, element_wise=True),
-            Check(lambda s: (s == 21).any())])
-    validated_series = schema.validate(pd.Series([1, 5, 21, 50]))
-    assert isinstance(validated_series, pd.Series)
-
-    # raise error if any of the validators fails
[…]
