### 2019-10-04 95b6ce5
add coerce logic to SeriesSchema, modularize tests (#117)
* add coerce logic to SeriesSchema, modularize tests

* fix flake8
