### 2019-10-04 d5816a3
vump to v0.2.3: coerce bugfix and modularize tests
