+++ pandera/__init__.py
@@ -17,4 +17,4 @@ from .schemas import DataFrameSchema, SeriesSchema
 from .schema_components import Column, Index, MultiIndex
 
 
-__version__ = "0.2.2"
+__version__ = "0.2.3"
