+++ conda.recipe/meta.yaml
@@ -1,9 +1,9 @@
 package:
   name: pandera
-  version: "0.2.2"
+  version: "0.2.3"
 
 source:
-  git_rev: 0.2.2
+  git_rev: 0.2.3
   git_url: https://github.com/pandera-dev/pandera.git
 
 build:
