### 2019-10-04 #12 @cosmicBboy
@lwasser I just released [v0.2.3](https://github.com/pandera-dev/pandera/releases/tag/v0.2.3) with improvements to documentation, unit test maintainability, and bugfix in coercion + nullable logic.
