### 2019-10-10 #12 @lwasser
awesome @cosmicBboy !! I believe this is done!! congratulations as you are the second package to successfully go through our review process!

Next steps.

1. I am redoing the website package list so i'll go ahead and add you to that and will provide the link when it's done!!
2. I will also add you as a contributor on our contributors page as i am working on that now.
3. we'd love for you to add a pyopensci badge to your readme file. Can you please do that?
```
[![pyOpenSci](https://tinyurl.com/y22nb8up)](https://github.com/pyOpenSci/software-review/issues/12)
```
Looks like this:

[![pyOpenSci](https://tinyurl.com/y22nb8up)](https://github.com/pyOpenSci/software-review/issues/12)
Congratulations all for pushing another package through the pyopensci review process.

@cosmicBboy because we are just building out our packages page, I may ask you for a bit more info on Pandera in the future. maybe a few sentence description. it won't be a big ask!!
