### 2019-10-14 e7d29ec
Add pyOpenSci badge
