### 2019-10-14 #12 @cosmicBboy
@lwasser badge added! Yes please let me know when you need more info on the package.

Out of curiosity, what are the requirements for submitting this package to JOSS?
