### 2019-10-15 #12 @lwasser
hi @cosmicBboy please check out this page

https://www.pyopensci.org/dev_guide/appendices/templates.html
 in our dev guide and scroll down to the section on joss!! (it's actually also in the review template) ... if you are interested in pursuing joss and meet their requirements, please do let us know. it would be the first submission that also went through JOSS!! but we do have an established partnership tih them.

thank you so much for also adding a badge!! i will close this submission now UNLESS you decide you'd like to submit to joss as well!! if so we'd need you to add the write up that they require  and then we'd push things up to their review process!!
please let me know what you'd like to do!!
