### 2019-11-09 a81d4b9
fix hypothesis docstrings: typos
