### 2019-11-09 033b814
make sure category dtype can be coerced, fix string coerce logic
- make sure that coercion logic can handle non-comparable
  series.dtype and pandera dtype.
- add unit tests for category and datetime datatypes
- bugfix: handle string coercian with nans and None
