### 2019-11-09 pandera-dev/pandera#110 @cosmicBboy
fixed by d5816a33b54793d3e191802c3bd9863d1efcd414
