### 2019-11-09 c679e46
bump to v0.2.4: testing for category and datetime dtypes
