+++ conda.recipe/meta.yaml
@@ -1,9 +1,9 @@
 package:
   name: pandera
-  version: "0.2.3"
+  version: "0.2.4"
 
 source:
-  git_rev: 0.2.3
+  git_rev: 0.2.4
   git_url: https://github.com/pandera-dev/pandera.git
 
 build:
