### 2019-11-11 e6bd065
add license to pypi distribution
