### 2019-11-11 pandera-dev/pandera#90 @cosmicBboy
PR to conda-forge https://github.com/conda-forge/staged-recipes/pull/10114
