### 2019-11-11 09c0780
bugfix: fix `index` missing bug in error formatting for named index
