### 2019-11-11 f06a3c3
fix failure case formatting logic
