### 2019-11-11 6fdd740
bugfix: fix `index` missing bug in error formatting for named index (#120)
* bugfix: fix `index` missing bug in error formatting for named index

* fix failure case formatting logic
