+++ pandera/checks.py
@@ -191,16 +191,19 @@ class Check(object):
         """Construct readable error messages for vectorized_error_message.
 
         :param failure_cases: The failure cases encountered by the element-wise
-            validator.
+            or vectorized validator.
         :returns: DataFrame where index contains failure cases, the "index"
             column contains a list of integer indexes in the validation
             DataFrame that caused the failure, and a "count" column
             representing how many failures of that case occurred.
 
         """
-        if isinstance(failure_cases.index, pd.MultiIndex):
+        # reset index so that index is just 0-indexed integers
+        if hasattr(failure_cases, "index") and \
+                isinstance(failure_cases.index, pd.MultiIndex):
             failure_cases = (
                 failure_cases
+                .reset_index(drop=True)
                 .rename("failure_case")
                 .reset_index()
                 .assign(
@@ -212,17 +215,23 @@ class Check(object):
         elif isinstance(failure_cases, pd.DataFrame):
             failure_cases = (
                 failure_cases
+                .reset_index(drop=True)
                 .pipe(lambda df: pd.Series(
                     df.itertuples()).map(lambda x: x.__repr__()))
                 .rename("failure_case")
                 .reset_index()
             )
-        else:
+        elif isinstance(failure_cases, pd.Series):
             failure_cases = (
                 failure_cases
+                .reset_index(drop=True)
                 .rename("failure_case")
                 .reset_index()
             )
+        else:
+            raise TypeError(
+                "type of failure_cases argument not understood: %s" %
+                type(failure_cases))
 
         failure_cases = (
             failure_cases
