### 2019-11-11 d915e00
bump to v0.2.6: named index bugfix
