### 2019-11-11 9dd4610
bugfix: preserve original index values in failure case formatting (#121)
