### 2019-11-22 6a973a8
adds multi-OS support to travis CI
