### 2019-11-22 0a28e30
add multi-OS support to travis CI
