+++ .travis.yml
@@ -1,30 +1,88 @@
-language: python
-dist: xenial
-python:
-  - "3.5"
-  - "3.6"
-  - "3.7"
+language: bash
+
+cache:
+  directories:
+    - $HOME/miniconda3
+
+before_cache:
+  - rm -rf $CONDA_DIR/pkgs/cache
+  - rm -rf $CONDA_DIR/envs/hosts
+  - rm -rf $CONDA_DIR/conda-meta/history
+  - touch $CONDA_DIR/conda-meta/history
+
+os:
+  - linux
+  - osx
+  - windows
+
+env:
+  - PYTHON_VERSION="3.5"
+  - PYTHON_VERSION="3.6"
+  - PYTHON_VERSION="3.7"
+
+before_install:
+  - |
+    export CONDA_DIR="$HOME/miniconda3"
+    if [ "${TRAVIS_OS_NAME}" == "windows" ]; then
+        export CONDA_DIR=`cygpath -w $CONDA_DIR`
+        export CONDA_BIN_DIR="$CONDA_DIR\scripts"
+    else
+        export CONDA_BIN_DIR="$CONDA_DIR/bin"
+    fi
+    export PATH="$CONDA_BIN_DIR:$PATH"
+
 install:
-  - sudo apt-get update
-  # We do this conditionally because it saves us some downloading if the
-  # version is the same.
-  - wget https://repo.continuum.io/miniconda/Miniconda3-latest-Linux-x86_64.sh -O miniconda.sh;
-  - bash miniconda.sh -b -p $HOME/miniconda
-  - export PATH="$HOME/miniconda/bin:$PATH"
-  - hash -r
-  - conda config --set always_yes yes --set changeps1 no
-  - conda update -q --yes conda
-  # Useful for debugging any issues with conda
-  - conda info -a
-  - conda create -q -n ci-env python=$TRAVIS_PYTHON_VERSION
-  - source activate ci-env
-  - conda install -y -c conda-forge numpy scipy pandas codecov
-  - pip install -r requirements.txt
-  - python setup.py install
+  # Install Conda
+  - |
+    if [ "${TRAVIS_OS_NAME}" == "windows" ]; then
+        choco install openssl.light
+    fi
+
+    if [ -d "$CONDA_DIR" ] && [ -e "$CONDA_BIN_DIR/conda" ]; then
+        echo "Miniconda install already present from cache: $CONDA_DIR"
+        rm -rf $CONDA_DIR/envs/hosts  # Just in case...
+    else
+        echo "Installing Miniconda..."
+        rm -rf $CONDA_DIR  # Just in case...
+
+        if [ "${TRAVIS_OS_NAME}" == "windows" ]; then
+            wget https://repo.continuum.io/miniconda/Miniconda3-latest-Windows-x86_64.exe -O miniconda.exe || exit 1
+            cmd //c "start /wait "" miniconda.exe /S /D=$CONDA_DIR"
+        else
+            if [ "${TRAVIS_OS_NAME}" == "osx" ]; then
+                wget https://repo.continuum.io/miniconda/Miniconda3-latest-MacOSX-x86_64.sh -O miniconda.sh || exit 1
+            elif [ "${TRAVIS_OS_NAME}" == "linux" ]; then
+                wget https://repo.continuum.io/miniconda/Miniconda3-latest-Linux-x86_64.sh -O miniconda.sh || exit 1
+            fi
+            bash miniconda.sh -b -p "$CONDA_DIR" || exit 1
+        fi
+    fi
+
+    echo "Configuring Miniconda..."
+    conda config --set ssl_verify false || exit 1
+    conda config --set always_yes true --set changeps1 false || exit 1
+
+    echo "Updating Miniconda..."
+    conda update conda
+    conda update --all
+    conda info -a || exit 1
[…]
