### 2019-11-23 b09ed0c
change language to bash to try and get windows VMs to build
