+++ .travis.yml
@@ -1,4 +1,4 @@
-language: minimal
+language: bash
 
 cache:
   directories:
