### 2019-11-23 pandera-dev/pandera#90 @cosmicBboy
https://github.com/pandera-dev/staged-recipes
