### 2019-11-26 3d9f26f
re-integrate shell scripts into travis.yml because no bash on windows
