+++ .travis.yml
@@ -25,8 +25,42 @@ before_install:
   - chmod +x ci/install_conda.sh ci/setup_conda_env.sh
 
 install:
-  - ci/install_conda.sh
-  - ci/setup_conda_env.sh
+  # Install Conda
+  - |
+    if [ -d "$HOME/miniconda3" ] && [ -e "$HOME/miniconda3/bin/conda" ]; then
+        echo "Miniconda install already present from cache: $HOME/miniconda3"
+        rm -rf $HOME/miniconda3/envs/hosts  # Just in case...
+    else
+        echo "Installing Miniconda..."
+        rm -rf $HOME/miniconda3  # Just in case...
+
+        if [ "${TRAVIS_OS_NAME}" == "osx" ]; then
+            wget http://repo.continuum.io/miniconda/Miniconda3-latest-MacOSX-x86_64.sh -O miniconda.sh || exit 1
+        else
+            wget http://repo.continuum.io/miniconda/Miniconda3-latest-Linux-x86_64.sh -O miniconda.sh || exit 1
+        fi
+
+        bash miniconda.sh -b -p "$HOME/miniconda3" || exit 1
+    fi
+
+    echo "Configuring Miniconda..."
+    conda config --set ssl_verify false || exit 1
+    conda config --set always_yes true --set changeps1 false || exit 1
+
+    echo "Updating Miniconda"
+    conda update conda
+    conda update --all
+    conda info -a || exit 1
+  # Setup Conda Env
+  - |
+    echo "Creating a Python $PYTHON_VERSION environment"
+    conda create -n hosts python=$PYTHON_VERSION || exit 1
+    source activate hosts
+
+    echo "Installing packages..."
+    conda install numpy scipy pandas codecov
+    pip install -r requirements.txt
+    python setup.py install
 
 script:
   - source activate hosts
