### 2019-11-26 1608b2b
remove reference to bash scripts
