### 2019-12-17 185d967
Feature/ci enhancements (#125)
* fix windows ci

* fix osx ci
