+++ .travis.yml
@@ -5,10 +5,10 @@ cache:
     - $HOME/miniconda3
 
 before_cache:
-  - rm -rf $HOME/miniconda3/pkgs/cache
-  - rm -rf $HOME/miniconda3/envs/hosts
-  - rm -rf $HOME/miniconda3/conda-meta/history
-  - touch $HOME/miniconda3/conda-meta/history
+  - rm -rf $CONDA_DIR/pkgs/cache
+  - rm -rf $CONDA_DIR/envs/hosts
+  - rm -rf $CONDA_DIR/conda-meta/history
+  - touch $CONDA_DIR/conda-meta/history
 
 os:
   - linux
@@ -21,38 +21,54 @@ env:
   - PYTHON_VERSION="3.7"
 
 before_install:
-  - export PATH="$HOME/miniconda3/bin:$PATH"
+  - |
+    export CONDA_DIR="$HOME/miniconda3"
+    if [ "${TRAVIS_OS_NAME}" == "windows" ]; then
+        export CONDA_DIR=`cygpath -w $CONDA_DIR`
+        export CONDA_BIN_DIR="$CONDA_DIR\scripts"
+    else
+        export CONDA_BIN_DIR="$CONDA_DIR/bin"
+    fi
+    export PATH="$CONDA_BIN_DIR:$PATH"
 
 install:
   # Install Conda
   - |
-    if [ -d "$HOME/miniconda3" ] && [ -e "$HOME/miniconda3/bin/conda" ]; then
-        echo "Miniconda install already present from cache: $HOME/miniconda3"
-        rm -rf $HOME/miniconda3/envs/hosts  # Just in case...
+    if [ "${TRAVIS_OS_NAME}" == "windows" ]; then
+        choco install openssl.light
+    fi
+
+    if [ -d "$CONDA_DIR" ] && [ -e "$CONDA_BIN_DIR/conda" ]; then
+        echo "Miniconda install already present from cache: $CONDA_DIR"
+        rm -rf $CONDA_DIR/envs/hosts  # Just in case...
     else
         echo "Installing Miniconda..."
-        rm -rf $HOME/miniconda3  # Just in case...
+        rm -rf $CONDA_DIR  # Just in case...
 
-        if [ "${TRAVIS_OS_NAME}" == "osx" ]; then
-            wget http://repo.continuum.io/miniconda/Miniconda3-latest-MacOSX-x86_64.sh -O miniconda.sh || exit 1
+        if [ "${TRAVIS_OS_NAME}" == "windows" ]; then
+            wget https://repo.continuum.io/miniconda/Miniconda3-latest-Windows-x86_64.exe -O miniconda.exe || exit 1
+            cmd //c "start /wait "" miniconda.exe /S /D=$CONDA_DIR"
         else
-            wget http://repo.continuum.io/miniconda/Miniconda3-latest-Linux-x86_64.sh -O miniconda.sh || exit 1
+            if [ "${TRAVIS_OS_NAME}" == "osx" ]; then
+                wget https://repo.continuum.io/miniconda/Miniconda3-latest-MacOSX-x86_64.sh -O miniconda.sh || exit 1
+            elif [ "${TRAVIS_OS_NAME}" == "linux" ]; then
+                wget https://repo.continuum.io/miniconda/Miniconda3-latest-Linux-x86_64.sh -O miniconda.sh || exit 1
+            fi
+            bash miniconda.sh -b -p "$CONDA_DIR" || exit 1
         fi
-
-        bash miniconda.sh -b -p "$HOME/miniconda3" || exit 1
     fi
 
     echo "Configuring Miniconda..."
     conda config --set ssl_verify false || exit 1
[…]
