### 2019-12-17 3dbaf05
simplifies multi-OS make step into a single-line of python
