### 2019-12-17 842e033
adds a test that the newly added pandas condig for doctest executes as expected
