### 2019-12-17 738938f
flake8 compliance
