### 2019-12-18 7f43f29
implement pylint including pre-commit hook, remove flake8, make code compliant
