### 2019-12-18 cb39fd0
implement pylint including pre-commit hook, remove flake8, make code compliant
