+++ pandera/constants.py
@@ -1,1 +1,2 @@
-N_FAILURE_CASES = 10
+"""Constants for use across the library"""
+N_FAILURE_CASES = 10 #The number of failure cases to report.
