+++ requirements.txt
@@ -1,7 +1,8 @@
 codecov
-flake8
 numpy >= 1.9.0
 pandas >= 0.23.0
+pre_commit
+pylint
 pytest
 pytest-cov
 scipy >= 1.2.0
