### 2019-12-30 fadc962
update conda installation in README
