### 2019-12-30 fa0c1ab
fix documentation DataFrame and Series
