### 2020-01-03 d5627e6
support pandas style dtype property on DataFrameSchema, SeriesSchema and Column
