### 2020-01-03 174fc95
better docstrings for dtype property and documentation section in DataFrame schema docs for dtype property
