### 2020-01-03 14f76a5
removing some test permutations for dtype property to get under pylint duplicate code threshold
