+++ tests/test_schemas.py
@@ -417,15 +417,12 @@ def test_dataframe_schema_dtype_property():
     (Bool, "bool"),
     (DateTime, "datetime64[ns]"),
     (Category, "category"),
-    (Float, "float64"),
     (Int, "int64"),
-    (Object, "object"),
     (String, "object"),
     (Timedelta, "timedelta64[ns]"),
     ("bool", "bool"),
     ("datetime64[ns]", "datetime64[ns]"),
     ("category", "category"),
-    ("float64", "float64"),
     ("float64", "float64"),
 ])
 def test_series_schema_dtype_property(pandas_dtype, expected):
