+++ tests/test_schema_components.py
@@ -96,15 +96,12 @@ def test_multi_index_index():
     (Bool, "bool"),
     (DateTime, "datetime64[ns]"),
     (Category, "category"),
-    (Float, "float64"),
     (Int, "int64"),
-    (Object, "object"),
     (String, "object"),
     (Timedelta, "timedelta64[ns]"),
     ("bool", "bool"),
     ("datetime64[ns]", "datetime64[ns]"),
     ("category", "category"),
-    ("float64", "float64"),
     ("float64", "float64"),
 ])
 def test_column_dtype_property(pandas_dtype, expected):
