### 2020-01-03 e786126
reverting removal of test permuations and adding R0801 to pylint ignore codes
