+++ tests/test_schemas.py
@@ -417,7 +417,9 @@ def test_dataframe_schema_dtype_property():
     (Bool, "bool"),
     (DateTime, "datetime64[ns]"),
     (Category, "category"),
+    (Float, "float64"),
     (Int, "int64"),
+    (Object, "object"),
     (String, "object"),
     (Timedelta, "timedelta64[ns]"),
     ("bool", "bool"),
