+++ tests/test_schema_components.py
@@ -96,7 +96,9 @@ def test_multi_index_index():
     (Bool, "bool"),
     (DateTime, "datetime64[ns]"),
     (Category, "category"),
+    (Float, "float64"),
     (Int, "int64"),
+    (Object, "object"),
     (String, "object"),
     (Timedelta, "timedelta64[ns]"),
     ("bool", "bool"),
