### 2020-01-03 a4afb3e
adjust indent of testoutput for dataframe dtype example to match other examples
