### 2020-01-03 a83a8e9
fix expected output for dataframe dtype documentation example
