### 2020-01-03 2b3da6f
move dataframe dtype documentation to top level section... it did not belong in the multi index section
