### 2020-01-04 7aecf16
support pandas style dtype property on DataFrameSchema, SeriesSchema and Column (#134)
* support pandas style dtype property on DataFrameSchema, SeriesSchema and Column

* better docstrings for dtype property and documentation section in DataFrame schema docs for dtype property

* removing some test permutations for dtype property to get under pylint duplicate code threshold

* reverting removal of test permuations and adding R0801 to pylint ignore codes

* adjust indent of testoutput for dataframe dtype example to match other examples

* fix expected output for dataframe dtype documentation e
[…]
