### 2020-01-06 a18b467
Fixes all C0114, C0116 pylint errors and re-enables CI tests for them.
