### 2020-01-07 cbc93f8
Fixes W0122
