### 2020-01-07 c2cbe3f
Fix R1705
