### 2020-01-07 aef4a33
Fix R1720
