### 2020-01-07 b52e18d
Fixes some pylint errors and re-enables CI tests for them. (#135)
* Fixes all C0114, C0116 pylint errors and re-enables CI tests for them.

* Fix C0103

* Fixes C0115

* Fixes W0122

* Fix W0212

* Fix R1705

* Fix R1720
