### 2020-01-07 efb6638
Fixes W0613, R0201, R0205, R0801.
