### 2020-01-07 pandera-dev/pandera#108 @mastersplinter
Solved by #127. Changes for pylint are tracked in #129.
