### 2020-01-07 7665b16
initial test of new multi-environment specific CI tests
