+++ requirements.txt
@@ -1,12 +1,0 @@
-codecov
-numpy >= 1.9.0
-pandas >= 0.23.0
-pre_commit
-pylint
-pytest
-pytest-cov
-scipy >= 1.2.0
-sphinx
-sphinx_rtd_theme
-sphinx_autodoc_typehints
-wrapt
