### 2020-01-08 1445b4a
fixes 3.5 CI and other pylint errors
