### 2020-01-08 c726a26
essential ci fixes and pylint changes (#138)
* Fixes all C0114, C0116 pylint errors and re-enables CI tests for them.

* Fix C0103

* Fixes C0115

* Fixes W0122

* Fix W0212

* Fix R1705

* Fix R1720

* Fixes W0613, R0201, R0205, R0801.

* initial test of new multi-environment specific CI tests

* make ENV_FILE a var name

* updated pylint dependency

* fixes 3.5 CI and other pylint errors
