+++ environment.yml
@@ -1,0 +1,31 @@
+name: pandera-dev-latest
+channels:
+  - defaults
+  - conda-forge
+dependencies:
+  - python
+
+  # environment management
+  - pip
+
+  # documentation
+  - sphinx
+  - sphinx_rtd_theme
+
+  # testing
+  - codecov
+  - pylint>=2.4.4
+  - pytest
+  - pytest-cov
+
+  # optional
+  - pre_commit
+
+  # pandera dependencies
+  - numpy
+  - pandas
+  - scipy
+  - wrapt
+
+  - pip:
+    - sphinx_autodoc_typehints
