### 2020-01-08 8df2995
adds equality operators for Checks, Columns, Schemas etc.
