+++ tests/test_checks.py
@@ -1,5 +1,6 @@
 """Tests the way Columns are Checked"""
 
+import copy
 import pandas as pd
 import pytest
 
@@ -265,3 +266,21 @@ def test_format_failure_case_exceptions():
     for data in [1, "foobar", 1.0, {"key": "value"}, list(range(10))]:
         with pytest.raises(TypeError):
             check._format_failure_cases(data)
+
+
+def test_check_equality_operators():
+    """Test the usage of == between a Check and an entirely different Check."""
+    check = Check(lambda g: g["foo"]["col1"].iat[0] == 1, groupby="col3")
+
+    not_equal_check = Check(lambda x: x.isna().sum() == 0)
+    assert check == copy.deepcopy(check)
+    assert check != not_equal_check
+
+
+def test_equality_operators_functional_equivalence():
+    """Test the usage of == for Checks where the Check callable object has
+    the same implementation."""
+    main_check = Check(lambda g: g["foo"]["col1"].iat[0] == 1, groupby="col3")
+    same_check = Check(lambda h: h["foo"]["col1"].iat[0] == 1, groupby="col3")
+
+    assert main_check == same_check
