### 2020-01-08 a5276fc
adds equality operators for Checks, Columns, Schemas etc.
