### 2020-01-09 01278d8
enable basic schema transformations
