+++ Makefile
@@ -20,7 +20,7 @@ upload-pypi:
 		rm -rf dist
 
 requirements:
-	pip install -r requirements.txt
+	pip install -r requirements-dev.txt
 
 docs:
 	make -C docs doctest && python -m sphinx -E -W "docs/source" "docs/_build"
