+++ .readthedocs.yml
@@ -20,6 +20,6 @@ formats: all
 python:
   version: 3.7
   install:
-    - requirements: requirements.txt
+    - requirements: requirements-dev.txt
     - method: pip
       path: .
