+++ tests/test_dtypes.py
@@ -33,13 +33,21 @@ def test_numeric_dtypes():
             dtypes.Float,
             dtypes.Float16,
             dtypes.Float32,
-            dtypes.Float64]:
-        schema = DataFrameSchema({"col": Column(dtype, nullable=False)})
-        validated_df = schema.validate(
-            pd.DataFrame(
-                {"col": [-123.1, -7654.321, 1.0, 1.1, 1199.51, 5.1, 4.6]},
-                dtype=dtype.value))
-        assert isinstance(validated_df, pd.DataFrame)
+            dtypes.Float64
+            ]:
+        assert all(
+            isinstance(
+                schema.validate(
+                    pd.DataFrame(
+                        {"col": [-123.1, -7654.321, 1.0, 1.1, 1199.51, 5.1]},
+                        dtype=dtype.value)),
+                pd.DataFrame
+            )
+            for schema in [
+                DataFrameSchema({"col": Column(dtype, nullable=False)}),
+                DataFrameSchema({"col": Column(dtype.value, nullable=False)})
+            ]
+        )
 
     for dtype in [
             dtypes.Int,
@@ -47,23 +55,38 @@ def test_numeric_dtypes():
             dtypes.Int16,
             dtypes.Int32,
             dtypes.Int64]:
-        schema = DataFrameSchema({"col": Column(dtype, nullable=False)})
-        validated_df = schema.validate(
-            pd.DataFrame(
-                {"col": [-712, -4, -321, 0, 1, 777, 5, 123, 9000]},
-                dtype=dtype.value))
-        assert isinstance(validated_df, pd.DataFrame)
+        assert all(
+            isinstance(
+                schema.validate(
+                    pd.DataFrame(
+                        {"col": [-712, -4, -321, 0, 1, 777, 5, 123, 9000]},
+                        dtype=dtype.value)),
+                pd.DataFrame
+            )
+            for schema in [
+                DataFrameSchema({"col": Column(dtype, nullable=False)}),
+                DataFrameSchema({"col": Column(dtype.value, nullable=False)})
+            ]
+        )
 
     for dtype in [
             dtypes.UInt8,
             dtypes.UInt16,
             dtypes.UInt32,
             dtypes.UInt64]:
-        schema = DataFrameSchema({"col": Column(dtype, nullable=False)})
-        validated_df = schema.validate(
-            pd.DataFrame(
-                {"col": [1, 777, 5, 123, 9000]}, dtype=dtype.value))
-        assert isinstance(validated_df, pd.DataFrame)
+        assert all(
+            isinstance(
+                schema.validate(
+                    pd.DataFrame(
+                        {"col": [1, 777, 5, 123, 9000]},
+                        dtype=dtype.value)),
+                pd.DataFrame
+            )
+            for schema in [
+                DataFrameSchema({"col": Column(dtype, nullable=False)}),
+                DataFrameSchema({"col": Column(dtype.value, nullable=False)})
+            ]
+        )
 
 
 def test_category_dtype():
