### 2020-01-10 df364f1
coerce=True should work when supplying string dtypes (#149)
* coerce=True should work when supplying string dtypes

this diff fixes a bug where coerce=True would not work if
a legal pandas string argument is passed into the `pandas_dtype`.
Update the unit tests to enforce this

* fix pylint
