### 2020-01-10 e71f05c
fix reqts-dev.txt generator script
