### 2020-01-10 771d281
fix reqts-dev.txt generator script (#150)
