### 2020-01-10 8b40b48
remove f strings from generator script
