### 2020-01-10 b49a4ab
remove f strings from generator script
