### 2020-01-10 2ecf476
add airspeed velocity benchmarks
