### 2020-01-12 b62fac4
adds equality operators for Checks, Columns, Schemas etc. (#143)
