### 2020-01-12 7bf2fe8
clean up Check.__call__ and SeriesSchemaBase.__call__  API (#153)
This diff makes the Check API cleaner by moving out the
error handling into the schema doing the data validation
instead of within the __call__ subroutine.

This also standardizes the Check.__call__ signature such that
it takes the dataframe or series object to be checked.

This also cleans up the SeriesSchemaBase API and its subclasses,
where the __call__ method takes a single argument df_or_series,
which is the object to be checked.
