+++ pandera/schemas.py
@@ -5,7 +5,7 @@ from typing import List, Optional, Union, Dict
 
 import pandas as pd
 
-from . import errors, constants, dtypes
+from . import errors, constants, dtypes, error_formatters
 from .checks import Check
 
 
@@ -154,14 +154,12 @@ class DataFrameSchema():
             pd.concat(dataframe_subsample).drop_duplicates()
 
     def _check_dataframe(self, dataframe):
-        val_results = []
+        check_results = []
         for check_index, check in enumerate(self.checks):
-            val_results.append(
-                check(
-                    self,
-                    check_index,
-                    check.prepare_dataframe_input(dataframe)))
-        return all(val_results)
+            check_results.append(
+                _handle_check_results(self, check_index, check, dataframe)
+            )
+        return all(check_results)
 
     @property
     def dtype(self) -> Dict[str, str]:
@@ -239,10 +237,12 @@ class DataFrameSchema():
 
         dataframe_to_validate = self._dataframe_to_validate(
             dataframe, head, tail, sample, random_state)
+
         assert (
             all(schema_component(dataframe_to_validate)
                 for schema_component in schema_components)
-            and self._check_dataframe(dataframe))
+            and self._check_dataframe(dataframe_to_validate))
+
         if self.transformer is not None:
             dataframe = self.transformer(dataframe)
 
@@ -337,6 +337,11 @@ class SeriesSchemaBase():
         return self._coerce
 
     @property
+    def name(self) -> str:
+        """Get SeriesSchema name."""
+        return self._name
+
+    @property
     def dtype(self) -> str:
         """String representation of the dtype."""
         return self._pandas_dtype if (
@@ -362,12 +367,17 @@ class SeriesSchemaBase():
             "The _allow_groupby property must be implemented by subclasses "
             "of SeriesSchemaBase")
 
-    def __call__(
-            self,
-            series: pd.Series,
-            dataframe_context: pd.DataFrame = None) -> bool:
+    def __call__(self, df_or_series: Union[pd.DataFrame, pd.Series]) -> bool:
         # pylint: disable=too-many-branches,W0212
-        """Validate a series."""
+        """Validate a series.
+
+        :df_or_series: pandas DataFrame of Series to validate.
+        :returns: True if validation checks pass.
+
+        """
+        series = df_or_series if isinstance(df_or_series, pd.Series) \
+            else df_or_series[self.name]
+
         if series.name != self._name:
             raise errors.SchemaError(
                 "Expected %s to have name '%s', found '%s'" %
@@ -377,8 +387,6 @@ class SeriesSchemaBase():
 
         if self._nullable:
             series = series.dropna()
-            if dataframe_context is not None:
-                dataframe_context = dataframe_context.loc[series.index]
             if _dtype in ["int_", "int8", "int16", "int32", "int64", "uint8",
                           "uint16", "uint32", "uint64"]:
                 _series = series.astype(_dtype)
@@ -427,14 +435,20 @@ class SeriesSchemaBase():
                 "expected series '%s' to have type %s, got %s" %
                 (series.name, _dtype, series.dtype))
 
-        val_results = []
+        check_results = []
+
+        if isinstance(df_or_series, pd.Series):
+            check_args = (series, )
+        else:
+            df_or_series = df_or_series.loc[series.index]
+            df_or_series[self.name] = series
+            check_args = (df_or_series, self.name)
+
         for check_index, check in enumerate(self.checks):
-            val_results.append(
-                check(
-                    self,
-                    check_index,
-                    check._prepare_series_input(series, dataframe_context)))
-        return all(val_results)
+            check_results.append(
+                _handle_check_results(self, check_index, check, *check_args)
+            )
+        return all(check_results)
 
[…]
