+++ pandera/schema_components.py
@@ -77,13 +77,12 @@ class Column(SeriesSchemaBase):
         self._name = name
         return self
 
-    def __call__(self, df: pd.DataFrame) -> bool:
+    def __call__(self, df_or_series: Union[pd.DataFrame, pd.Series]) -> bool:
         """Validate DataFrameSchema Column."""
         if self._name is None:
             raise RuntimeError(
                 "need to `set_name` of column before calling it.")
-        return super(Column, self).__call__(
-            df[self._name], dataframe_context=df.drop(self._name, axis=1))
+        return super(Column, self).__call__(df_or_series)
 
     def __repr__(self):
         if isinstance(self._pandas_dtype, PandasDtype):
@@ -150,9 +149,9 @@ class Index(SeriesSchemaBase):
         """Whether the schema or schema component allows groupby operations."""
         return False
 
-    def __call__(self, df: pd.DataFrame) -> bool:
+    def __call__(self, df_or_series: Union[pd.DataFrame, pd.Series]) -> bool:
         """Validate DataFrameSchema Index."""
-        return super(Index, self).__call__(pd.Series(df.index))
+        return super(Index, self).__call__(pd.Series(df_or_series.index))
 
     def __repr__(self):
         if self._name is None:
@@ -231,10 +230,14 @@ class MultiIndex(DataFrameSchema):
             strict=strict,
         )
 
-    def __call__(self, df: pd.DataFrame) -> bool:
+    def __call__(self, df_or_series: Union[pd.DataFrame, pd.Series]) -> bool:
+        # pylint: disable=signature-differs,W0222
+        # false positive warning is raised here, even though method signature
+        # is exactly the same. Will need to investigate why this is being
+        # raised.
         """Validate DataFrameSchema MultiIndex."""
         return isinstance(
-            super(MultiIndex, self).__call__(df.index.to_frame()),
+            super(MultiIndex, self).__call__(df_or_series.index.to_frame()),
             pd.DataFrame
         )
 
