+++ pandera/schemas.py
@@ -1,6 +1,7 @@
 """Core pandera schema class definitions."""
 
 import json
+import warnings
 from typing import List, Optional, Union, Dict
 
 import pandas as pd
@@ -95,27 +96,6 @@ class DataFrameSchema():
         self._validate_schema()
         self._set_column_names()
 
-    def __call__(
-            self,
-            dataframe: pd.DataFrame,
-            head: Optional[int] = None,
-            tail: Optional[int] = None,
-            sample: Optional[int] = None,
-            random_state: Optional[int] = None):
-        """Delegate to `validate` method.
-
-        :param pd.DataFrame dataframe: the dataframe to be validated.
-        :param head: validate the first n rows. Rows overlapping with `tail` or
-            `sample` are de-duplicated.
-        :type head: int
-        :param tail: validate the last n rows. Rows overlapping with `head` or
-            `sample` are de-duplicated.
-        :type tail: int
-        :param sample: validate a random sample of n rows. Rows overlapping
-            with `head` or `tail` are de-duplicated.
-        """
-        return self.validate(dataframe)
-
     def _validate_schema(self):
         for column_name, column in self.columns.items():
             for check in column.checks:
@@ -130,8 +110,17 @@ class DataFrameSchema():
                         (nonexistent_groupby_columns, column_name))
 
     def _set_column_names(self):
+
+        def _set_column_handler(column, column_name):
+            if column.name is not None and column.name != column_name:
+                warnings.warn(
+                    "resetting column for %s to '%s'." % (column, column_name))
+            elif column.name == column_name:
+                return column
+            return column.set_name(column_name)
+        
         self.columns = {
-            column_name: column.set_name(column_name)
+            column_name: _set_column_handler(column, column_name)
             for column_name, column in self.columns.items()
         }
 
@@ -211,6 +200,8 @@ class DataFrameSchema():
         4         0.80      dog
         5         0.76      dog
         """
+        dataframe = dataframe.copy()
+
         if self.strict:
             for column in dataframe:
                 if column not in self.columns:
@@ -229,24 +220,47 @@ class DataFrameSchema():
                 dataframe[colname] = col.coerce_dtype(dataframe[colname])
 
         schema_components = [
-            col.set_name(col_name) for col_name, col in self.columns.items()
+            col for col_name, col in self.columns.items()
             if col.required or col_name in dataframe
         ]
         if self.index is not None:
+            if self.index.coerce or self.coerce:
+                dataframe.index = self.index.coerce_dtype(dataframe.index)
             schema_components.append(self.index)
 
         dataframe_to_validate = self._dataframe_to_validate(
             dataframe, head, tail, sample, random_state)
 
         assert (
-            all(schema_component(dataframe_to_validate)
-                for schema_component in schema_components)
+            all(isinstance(component(dataframe_to_validate), pd.DataFrame)
+                for component in schema_components)
             and self._check_dataframe(dataframe_to_validate))
 
         if self.transformer is not None:
             dataframe = self.transformer(dataframe)
 
         return dataframe
+
+    def __call__(
+            self,
+            dataframe: pd.DataFrame,
+            head: Optional[int] = None,
+            tail: Optional[int] = None,
+            sample: Optional[int] = None,
+            random_state: Optional[int] = None):
+        """Delegate to `validate` method.
+
+        :param pd.DataFrame dataframe: the dataframe to be validated.
+        :param head: validate the first n rows. Rows overlapping with `tail` or
+            `sample` are de-duplicated.
+        :type head: int
[…]
