### 2020-01-13 7fc9582
more explicit hinting
