### 2020-01-13 8d09adc
Revert "more explicit hinting"
This reverts commit 7fc95829a220ae0ec70a4cc2a23f9be42a0f240a.
