### 2020-01-13 dcc40f4
enables simple schema-transformations with add/remove columns (#144)
* enables simple schema transformations

Co-authored-by: Niels Bantilan <niels.bantilan@gmail.com>
