+++ tests/test_schemas.py
@@ -486,3 +486,54 @@ def test_schema_equality_operators():
     assert series_schema != not_equal_schema
     assert series_schema_base == copy.deepcopy(series_schema_base)
     assert series_schema_base != not_equal_schema
+
+
+def test_add_and_remove_columns():
+    """Check that adding and removing columns works as expected and doesn't
+    modify the original underlying DataFrameSchema."""
+    schema1 = DataFrameSchema({
+        "col1": Column(Int, Check(lambda s: s >= 0)),
+        }, strict=True)
+
+    schema1_exact_copy = copy.deepcopy(schema1)
+
+    # test that add_columns doesn't modify schema1 after add_columns:
+    schema2 = schema1.add_columns({
+        "col2": Column(String, Check(lambda x: x <= 0)),
+        "col3": Column(Object, Check(lambda x: x == 0))
+        })
+
+    schema2_exact_copy = copy.deepcopy(schema2)
+
+    assert schema1 == schema1_exact_copy
+
+    # test that add_columns changed schema1 into schema2:
+    expected_schema_2 = DataFrameSchema({
+        "col1": Column(Int, Check(lambda s: s >= 0)),
+        "col2": Column(String, Check(lambda x: x <= 0)),
+        "col3": Column(Object, Check(lambda x: x == 0))
+        }, strict=True)
+
+    assert schema2 == expected_schema_2
+
+    # test that remove_columns doesn't modify schema2:
+    schema3 = schema2.remove_columns(["col2"])
+
+    assert schema2 == schema2_exact_copy
+
+    # test that remove_columns has removed the changes as expected:
+    expected_schema_3 = DataFrameSchema({
+        "col1": Column(Int, Check(lambda s: s >= 0)),
+        "col3": Column(Object, Check(lambda x: x == 0))
+        }, strict=True)
+
+    assert schema3 == expected_schema_3
+
+    # test that remove_columns can remove two columns:
+    schema4 = schema2.remove_columns(["col2", "col3"])
+
+    expected_schema_4 = DataFrameSchema({
+        "col1": Column(Int, Check(lambda s: s >= 0))
+        }, strict=True)
+
+    assert schema4 == expected_schema_4 == schema1
