### 2020-01-13 79a0c8a
fix and enhance multi index docstring
