### 2020-01-13 1bb3e66
improve code markup in docstring
