### 2020-01-13 dfa3079
fix and enhance multi index docstring (#158)
