### 2020-01-14 0de799f
clean up underlying interface of schema coercion
- implement Index and MultiIndex schema_component coercion logic,
  previously this functionaly was not operational.
- add documentation for stand-alone Column checks
