### 2020-01-14 f2a9157
fix docs typo
