### 2020-01-15 aece1d1
clean up API further, fix docs @mastersplinter
