+++ docs/source/dataframe_schemas.rst
@@ -263,7 +263,7 @@ objects can also be used to validate columns in a dataframe on its own:
     validated_df = column2_schema.validate(df)
 
     # use the DataFrame.pipe method to validate two columns
-    validated_df = df.pipe(column1).pipe(column2)
+    validated_df = df.pipe(column1_schema).pipe(column2_schema)
 
 
 For multi-column use cases, the ``DataFrameSchema`` is still recommended, but
