### 2020-01-15 10e32c6
feature/clean-schema-api (#160)
* make schema API methods more consistent, coerce indexes

This diff makes the schema API much cleaner by standardizing
the __call__ and validate method implementation. The __call__
method delegates to the validate method and should handle
type coercion, while the validate method handles executing
the validation checks.

This diff also implements type coercion logic for Index and
MultiIndex schema components, which was previously not in
effect.

* clean up underlying interface of schema coercion

- implement Index and MultiIndex schema_component coercion logic,

[…]
