+++ pandera/schemas.py
@@ -2,8 +2,8 @@
 
 import json
 import copy
+import warnings
 from typing import List, Optional, Union, Dict, Any
-
 
 import pandas as pd
 
@@ -19,7 +19,7 @@ class DataFrameSchema():
 
     def __init__(
             self,
-            columns,
+            columns: Dict[str, Any] = None,
             checks: Optional[List[Check]] = None,
             index=None,
             transformer: callable = None,
@@ -30,7 +30,7 @@ class DataFrameSchema():
         :param columns: a dict where keys are column names and values are
             Column objects specifying the datatypes and properties of a
             particular column.
-        :type columns: Dict[str, Column]
+        :type columns: mapping of column names and column schema component.
         :param checks: dataframe-wide checks.
         :param index: specify the datatypes and properties of the index.
         :param transformer: a callable with signature:
@@ -83,40 +83,25 @@ class DataFrameSchema():
         if isinstance(checks, Check):
             checks = [checks]
 
-        if coerce and None in [c.pandas_dtype for c in columns.values()]:
+        self.columns = {} if columns is None else columns
+
+        if coerce and None in [c.pandas_dtype for c in self.columns.values()]:
             raise errors.SchemaInitError(
                 "Must specify dtype in all Columns if coercing "
                 "DataFrameSchema")
 
         self.checks = checks
         self.index = index
-        self.columns = columns
         self.transformer = transformer
-        self.coerce = coerce
         self.strict = strict
+        self._coerce = coerce
         self._validate_schema()
         self._set_column_names()
 
-    def __call__(
-            self,
-            dataframe: pd.DataFrame,
-            head: Optional[int] = None,
-            tail: Optional[int] = None,
-            sample: Optional[int] = None,
-            random_state: Optional[int] = None):
-        """Delegate to `validate` method.
-
-        :param pd.DataFrame dataframe: the dataframe to be validated.
-        :param head: validate the first n rows. Rows overlapping with `tail` or
-            `sample` are de-duplicated.
-        :type head: int
-        :param tail: validate the last n rows. Rows overlapping with `head` or
-            `sample` are de-duplicated.
-        :type tail: int
-        :param sample: validate a random sample of n rows. Rows overlapping
-            with `head` or `tail` are de-duplicated.
-        """
-        return self.validate(dataframe)
+    @property
+    def coerce(self):
+        """Whether to coerce series to specified type."""
+        return self._coerce
 
     def _validate_schema(self):
         for column_name, column in self.columns.items():
@@ -132,8 +117,17 @@ class DataFrameSchema():
                         (nonexistent_groupby_columns, column_name))
 
     def _set_column_names(self):
+
+        def _set_column_handler(column, column_name):
+            if column.name is not None and column.name != column_name:
+                warnings.warn(
+                    "resetting column for %s to '%s'." % (column, column_name))
+            elif column.name == column_name:
+                return column
+            return column.set_name(column_name)
+
         self.columns = {
-            column_name: column.set_name(column_name)
+            column_name: _set_column_handler(column, column_name)
             for column_name, column in self.columns.items()
         }
 
@@ -213,6 +207,8 @@ class DataFrameSchema():
         4         0.80      dog
         5         0.76      dog
         """
+        dataframe = dataframe.copy()
+
         if self.strict:
             for column in dataframe:
                 if column not in self.columns:
@@ -231,24 +227,47 @@ class DataFrameSchema():
                 dataframe[colname] = col.coerce_dtype(dataframe[colname])
 
         schema_components = [
[…]
