### 2020-01-17 ee21cdb
add airspeed velocity benchmarks (#152)
