### 2020-01-18 b2b69eb
Added checks.greater_than and unit tests
