### 2020-01-18 73a22ac
Added error message to greater_than
