### 2020-01-18 b7353e6
bugfix: string and object coercion work correctly
