### 2020-01-19 f63e249
Refactored tests to narrow down their scope to the unit under test
