### 2020-01-19 d10c4a7
Moved greater_than inside Check class
