### 2020-01-19 8f71684
Added greater_or_equal, less_than, less_or_equal
