### 2020-01-19 872bddd
Corrected docstrings
