### 2020-01-20 d3d3615
Added equal_to and not_equal_to
