### 2020-01-20 a28c142
Added isin and notin
