### 2020-01-21 b5c3212
Added str_matches and str_contains
