### 2020-01-21 0802ed5
Added str_startswith and str_endswith
