### 2020-01-21 db5c26f
Added str_length, cleaning up
