+++ pandera/checks.py
@@ -291,7 +291,7 @@ class Check():
 
     @staticmethod
     def greater_than(min_value) -> 'Check':
-        """Get a :class:`Check` ensuring all values of a series are strictly greater
+        """Create a :class:`Check` ensuring all values of a series are strictly greater
             than a certain value.
 
         :param min_value: Lower bound to be exceeded. Must be a type comparable to
@@ -311,7 +311,7 @@ class Check():
 
     @staticmethod
     def greater_or_equal(min_value) -> 'Check':
-        """Get a :class:`Check` ensuring all values are greater or equal a certain value.
+        """Create a :class:`Check` ensuring all values are greater or equal a certain value.
 
         :param min_value: Allowed minimum value for values of a series. Must be a type
             comparable to the dtype of the :class:`pandas.Series` to be validated.
@@ -330,9 +330,9 @@ class Check():
 
     @staticmethod
     def less_than(max_value) -> 'Check':
-        """Get a :class:`Check` ensuring all values are strictly below a certain value.
+        """Create a :class:`Check` ensuring all values are strictly below a certain value.
 
-        :param min_value: All elements of a series must be strictly smaller than this.
+        :param max_value: All elements of a series must be strictly smaller than this.
             Must be a type comparable to the dtype of the :class:`pandas.Series` to be
             validated.
 
@@ -349,7 +349,7 @@ class Check():
 
     @staticmethod
     def less_or_equal(max_value) -> 'Check':
-        """Get a :class:`Check` ensuring no value of a series exceeds a certain value.
+        """Create a :class:`Check` ensuring no value of a series exceeds a certain value.
 
         :param max_value: Upper bound not to be exceeded. Must be a type comparable to
             the dtype of the :class:`pandas.Series` to be validated.
@@ -367,7 +367,7 @@ class Check():
 
     @staticmethod
     def in_range(min_value, max_value, min_included=True, max_included=True) -> 'Check':
-        """Get a :class:`Check` ensuring all values of a series are within an interval.
+        """Create a :class:`Check` ensuring all values of a series are within an interval.
 
         :param min_value: Left / lower endpoint of the interval.
         :param max_value: Right / upper endpoint of the interval. Must not be smaller
@@ -403,7 +403,7 @@ class Check():
 
     @staticmethod
     def equal_to(value) -> 'Check':
-        """Get a :class:`Check` ensuring all elements of a series equal a certain value.
+        """Create a :class:`Check` ensuring all elements of a series equal a certain value.
 
         :param value: This value all elements of a given :class:`pandas.Series` must have.
 
@@ -417,7 +417,7 @@ class Check():
 
     @staticmethod
     def not_equal_to(value) -> 'Check':
-        """Get a :class:`Check` ensuring no elements of a series equals a certain value.
+        """Create a :class:`Check` ensuring no elements of a series equals a certain value.
 
         :param value: This value must not occur in a :class:`pandas.Series` to check.
 
@@ -431,7 +431,7 @@ class Check():
 
     @staticmethod
     def isin(allowed_values: Iterable) -> 'Check':
-        """Get a :class:`Check` to ensure only allowed values occur within a series.
+        """Create a :class:`Check` to ensure only allowed values occur within a series.
 
         :param allowed_values: The set of allowed values. May be any iterable.
 
@@ -459,7 +459,7 @@ class Check():
 
     @staticmethod
     def notin(forbidden_values: Iterable) -> 'Check':
-        """Get a :class:`Check` to ensure some defined values don't occur within a series.
+        """Create a :class:`Check` to ensure some defined values don't occur within a series.
 
         :param forbidden_values: The set of values which should not occur. May be any iterable.
 
@@ -486,7 +486,7 @@ class Check():
 
     @staticmethod
[…]
