### 2020-01-21 f0344e5
Fixed linter warning
