### 2020-01-22 c3f9e7b
Doc section for built-in checks, renaming of greater/less_or_equal, and include_min/max
