+++ pandera/checks.py
@@ -310,7 +310,7 @@ class Check():
         return Check(fn=_greater_than, error="greater_than(%s)" % min_value)
 
     @staticmethod
-    def greater_or_equal(min_value) -> 'Check':
+    def greater_than_or_equal_to(min_value) -> 'Check':
         """Create a :class:`Check` ensuring all values are greater or equal a certain value.
 
         :param min_value: Allowed minimum value for values of a series. Must be a type
@@ -326,7 +326,7 @@ class Check():
             return series >= min_value
 
         return Check(fn=_greater_or_equal,
-                     error="greater_or_equal(%s)" % min_value)
+                     error="greater_than_or_equal_to(%s)" % min_value)
 
     @staticmethod
     def less_than(max_value) -> 'Check':
@@ -348,7 +348,7 @@ class Check():
         return Check(fn=_less_than, error="less_than(%s)" % max_value)
 
     @staticmethod
-    def less_or_equal(max_value) -> 'Check':
+    def less_than_or_equal_to(max_value) -> 'Check':
         """Create a :class:`Check` ensuring no value of a series exceeds a certain value.
 
         :param max_value: Upper bound not to be exceeded. Must be a type comparable to
@@ -363,18 +363,18 @@ class Check():
             """Comparison function for check"""
             return series <= max_value
 
-        return Check(fn=_less_or_equal, error="less_or_equal(%s)" % max_value)
+        return Check(fn=_less_or_equal, error="less_than_or_equal_to(%s)" % max_value)
 
     @staticmethod
-    def in_range(min_value, max_value, min_included=True, max_included=True) -> 'Check':
+    def in_range(min_value, max_value, include_min=True, include_max=True) -> 'Check':
         """Create a :class:`Check` ensuring all values of a series are within an interval.
 
         :param min_value: Left / lower endpoint of the interval.
         :param max_value: Right / upper endpoint of the interval. Must not be smaller
             than min_value.
-        :param min_included: Defines whether min_value is also an allowed value
+        :param include_min: Defines whether min_value is also an allowed value
             (the default) or whether all values must be strictly greater than min_value.
-        :param max_included: Defines whether min_value is also an allowed value
+        :param include_max: Defines whether min_value is also an allowed value
             (the default) or whether all values must be strictly smaller than max_value.
 
         Both endpoints must be a type comparable to the dtype of the
@@ -387,12 +387,12 @@ class Check():
         if max_value is None:
             raise ValueError("max_value must not be None")
         if max_value < min_value or (min_value == max_value
-                                     and (not min_included or not max_included)):
+                                     and (not include_min or not include_max)):
             raise ValueError("The combination of min_value = %s and max_value = %s "
                              "defines an empty interval!" % (min_value, max_value))
         # Using functions from operator module to keep conditions out of the closure
-        left_op = operator.le if min_included else operator.lt
-        right_op = operator.ge if max_included else operator.gt
+        left_op = operator.le if include_min else operator.lt
+        right_op = operator.ge if include_max else operator.gt
 
         def _in_range(series: pd.Series) -> pd.Series:
             """Comparison function for check"""
