### 2020-01-22 1be3bcc
feature/builtin-checks (#164)
* Added checks.greater_than and unit tests

* Added error message to greater_than

* Refactored tests to narrow down their scope to the unit under test

* Moved greater_than inside Check class

* Added greater_or_equal, less_than, less_or_equal

* Corrected docstrings

* Added in_range

* Added equal_to and not_equal_to

* Added isin and notin

* Added str_matches and str_contains

* Added str_startswith and str_endswith

* Added str_length, cleaning up

* Fixed linter warning

* Doc section for built-in checks, renaming of greater/less_or_equal, and include_min/m
[…]
