+++ pandera/checks.py
@@ -1,7 +1,9 @@
 """Data validation checks."""
 
 from collections import namedtuple
-from typing import Dict, Union, Optional, List, Callable
+import operator
+import re
+from typing import Dict, Union, Optional, List, Callable, Iterable
 
 import pandas as pd
 
@@ -286,3 +288,298 @@ class Check():
         name = getattr(self.fn, '__name__', self.fn.__class__.__name__)
         return "<Check %s: %s>" % (name, self.error) \
             if self.error is not None else "<Check %s>" % name
+
+    @staticmethod
+    def greater_than(min_value) -> 'Check':
+        """Create a :class:`Check` ensuring all values of a series are strictly greater
+            than a certain value.
+
+        :param min_value: Lower bound to be exceeded. Must be a type comparable to
+            the dtype of the :class:`pandas.Series` to be validated (e.g. a numerical
+            type for float or int and a datetime for datetime).
+
+        :returns :class:`Check` object
+        """
+        if min_value is None:
+            raise ValueError("min_value must not be None")
+
+        def _greater_than(series: pd.Series) -> pd.Series:
+            """Comparison function for check"""
+            return series > min_value
+
+        return Check(fn=_greater_than, error="greater_than(%s)" % min_value)
+
+    @staticmethod
+    def greater_than_or_equal_to(min_value) -> 'Check':
+        """Create a :class:`Check` ensuring all values are greater or equal a certain value.
+
+        :param min_value: Allowed minimum value for values of a series. Must be a type
+            comparable to the dtype of the :class:`pandas.Series` to be validated.
+
+        :returns :class:`Check` object
+        """
+        if min_value is None:
+            raise ValueError("min_value must not be None")
+
+        def _greater_or_equal(series: pd.Series) -> pd.Series:
+            """Comparison function for check"""
+            return series >= min_value
+
+        return Check(fn=_greater_or_equal,
+                     error="greater_than_or_equal_to(%s)" % min_value)
+
+    @staticmethod
+    def less_than(max_value) -> 'Check':
+        """Create a :class:`Check` ensuring all values are strictly below a certain value.
+
+        :param max_value: All elements of a series must be strictly smaller than this.
+            Must be a type comparable to the dtype of the :class:`pandas.Series` to be
+            validated.
+
+        :returns :class:`Check` object
+        """
+        if max_value is None:
+            raise ValueError("max_value must not be None")
+
+        def _less_than(series: pd.Series) -> pd.Series:
+            """Comparison function for check"""
+            return series < max_value
+
+        return Check(fn=_less_than, error="less_than(%s)" % max_value)
+
+    @staticmethod
+    def less_than_or_equal_to(max_value) -> 'Check':
+        """Create a :class:`Check` ensuring no value of a series exceeds a certain value.
+
+        :param max_value: Upper bound not to be exceeded. Must be a type comparable to
+            the dtype of the :class:`pandas.Series` to be validated.
+
+        :returns :class:`Check` object
+        """
+        if max_value is None:
+            raise ValueError("max_value must not be None")
+
+        def _less_or_equal(series: pd.Series) -> pd.Series:
+            """Comparison function for check"""
+            return series <= max_value
+
+        return Check(fn=_less_or_equal, error="less_than_or_equal_to(%s)" % max_value)
+
+    @staticmethod
+    def in_range(min_value, max_value, include_min=True, include_max=True) -> 'Check':
+        """Create a :class:`Check` ensuring all values of a series are within an interval.
+
+        :param min_value: Left / lower endpoint of the interval.
+        :param max_value: Right / upper endpoint of the interval. Must not be smaller
+            than min_value.
+        :param include_min: Defines whether min_value is also an allowed value
+            (the default) or whether all values must be strictly greater than min_value.
[…]
