### 2020-01-23 bcf87bd
version 0.3.0 release (#167)
