### 2020-02-21 f853e9c
add mypy to pre-commit config
