+++ .pre-commit-config.yaml
@@ -3,3 +3,7 @@ repos:
     rev: v2.4.4
     hooks:
     -   id: pylint
+  - repo: https://github.com/pre-commit/mirrors-mypy
+    rev: v0.761
+    hooks:
+    -   id: mypy
