+++ environment.yml
@@ -15,6 +15,7 @@ dependencies:
 
   # testing
   - codecov
+  - mypy
   - pylint>=2.4.4
   - pytest
   - pytest-cov
