### 2020-02-22 fcd5d2d
fix mypy errors, update pylint config, mypy in ci
