+++ pandera/schema_components.py
@@ -1,6 +1,6 @@
 """Components used in pandera schemas."""
 
-from typing import Union
+from typing import Union, Optional
 
 import pandas as pd
 
@@ -79,10 +79,25 @@ class Column(SeriesSchemaBase):
         self._name = name
         return self
 
-    def validate(self, check_obj: pd.DataFrame) -> pd.DataFrame:
+    def validate(
+            self,
+            check_obj: pd.DataFrame,
+            head: Optional[int] = None,
+            tail: Optional[int] = None,
+            sample: Optional[int] = None,
+            random_state: Optional[int] = None,
+    ) -> pd.DataFrame:
+        # pylint: disable=duplicate-code
         """Validate a Column in a DataFrame object.
 
         :param check_obj: pandas DataFrame to validate.
+        :param head: validate the first n rows. Rows overlapping with `tail` or
+            `sample` are de-duplicated.
+        :param tail: validate the last n rows. Rows overlapping with `head` or
+            `sample` are de-duplicated.
+        :param sample: validate a random sample of n rows. Rows overlapping
+            with `head` or `tail` are de-duplicated.
+        :param random_state: random seed for the ``sample`` argument.
         :returns: validated DataFrame.
         """
         if self._name is None:
@@ -177,11 +192,23 @@ class Index(SeriesSchemaBase):
 
     def validate(
             self,
-            check_obj: Union[pd.DataFrame, pd.Series]
+            check_obj: Union[pd.DataFrame, pd.Series],
+            head: Optional[int] = None,
+            tail: Optional[int] = None,
+            sample: Optional[int] = None,
+            random_state: Optional[int] = None,
     ) -> Union[pd.DataFrame, pd.Series]:
+        # pylint: disable=duplicate-code
         """Validate DataFrameSchema or SeriesSchema Index.
 
         :check_obj: pandas DataFrame of Series containing index to validate.
+        :param head: validate the first n rows. Rows overlapping with `tail` or
+            `sample` are de-duplicated.
+        :param tail: validate the last n rows. Rows overlapping with `head` or
+            `sample` are de-duplicated.
+        :param sample: validate a random sample of n rows. Rows overlapping
+            with `head` or `tail` are de-duplicated.
+        :param random_state: random seed for the ``sample`` argument.
         :returns: validated DataFrame or Series.
         """
 
@@ -306,15 +333,26 @@ class MultiIndex(DataFrameSchema):
 
     def validate(
             self,
-            check_obj: Union[pd.DataFrame, pd.Series]
+            check_obj: Union[pd.DataFrame, pd.Series],
+            head: Optional[int] = None,
+            tail: Optional[int] = None,
+            sample: Optional[int] = None,
+            random_state: Optional[int] = None,
     ) -> Union[pd.DataFrame, pd.Series]:
-        # pylint: disable=signature-differs,arguments-differ
+        # pylint: disable=signature-differs,arguments-differ,duplicate-code
         # will need to clean up the class structure of this module since
         # this MultiIndex subclasses DataFrameSchema, which has a different
         # signature
         """Validate DataFrame or Series MultiIndex.
 
-        :check_obj: pandas DataFrame of Series to validate.
+        :param check_obj: pandas DataFrame of Series to validate.
+        :param head: validate the first n rows. Rows overlapping with `tail` or
+            `sample` are de-duplicated.
+        :param tail: validate the last n rows. Rows overlapping with `head` or
+            `sample` are de-duplicated.
+        :param sample: validate a random sample of n rows. Rows overlapping
+            with `head` or `tail` are de-duplicated.
+        :param random_state: random seed for the ``sample`` argument.
         :returns: validated DataFrame or Series.
         """
 
[…]
