+++ setup.cfg
@@ -1,0 +1,3 @@
+[mypy]
+files = pandera/**/*.py,tests/**/*.py
+ignore_missing_imports = True

[…]
