### 2020-02-22 8c8639e
change mypy config
