+++ .travis.yml
@@ -91,4 +91,4 @@ script:
   # Linting
   - pylint pandera tests
   # Type checking
-  - mypy
+  - mypy pandera tests
