+++ setup.cfg
@@ -1,3 +1,2 @@
 [mypy]
-files = pandera/**/*.py,tests/**/*.py
 ignore_missing_imports = True
