### 2020-02-23 be1fa70
and -> None to __init__ methods, fix tests
