+++ .pre-commit-config.yaml
@@ -8,5 +8,7 @@ repos:
     rev: v0.761
     hooks:
     -   id: mypy
+        entry: mypy pandera tests
+        files: (^pandera/|^tests)
         exclude: (^docs/|^scripts)
-        args: ["pandera", "tests"]
+        pass_filenames: false
