### 2020-02-23 be3edc2
feature/171: fix mypy errors (#176)
* add mypy to pre-commit config

* fix mypy errors, update pylint config, mypy in ci

* regenerate requirements-dev.txt

* add mypy to ci deps

* change mypy config

* and -> None to __init__ methods, fix tests

* fix lint
