+++ pandera/schemas.py
@@ -3,12 +3,12 @@
 import json
 import copy
 import warnings
-from typing import List, Optional, Union, Dict, Any
+from typing import Callable, List, Optional, Union, Dict, Any
 
 import pandas as pd
 
 from . import errors, constants, dtypes, error_formatters
-from .checks import Check
+from .checks import Check, CheckResult
 
 
 N_INDENT_SPACES = 4
@@ -19,12 +19,12 @@ class DataFrameSchema():
 
     def __init__(
             self,
-            columns: Dict[str, Any] = None,
+            columns: Dict[Any, Any] = None,
             checks: Optional[List[Check]] = None,
             index=None,
-            transformer: callable = None,
+            transformer: Callable = None,
             coerce: bool = False,
-            strict=False):
+            strict=False) -> None:
         """Initialize DataFrameSchema validator.
 
         :param columns: a dict where keys are column names and values are
@@ -134,9 +134,9 @@ class DataFrameSchema():
     @staticmethod
     def _dataframe_to_validate(
             dataframe: pd.DataFrame,
-            head: int,
-            tail: int,
-            sample: int,
+            head: Optional[int],
+            tail: Optional[int],
+            sample: Optional[int],
             random_state: Optional[int]) -> pd.DataFrame:
         dataframe_subsample = []
         if head is not None:
@@ -153,7 +153,8 @@ class DataFrameSchema():
         check_results = []
         for check_index, check in enumerate(self.checks):
             check_results.append(
-                _handle_check_results(self, check_index, check, dataframe)
+                _handle_check_results(
+                    self, check_index, check, check(dataframe))
             )
         return all(check_results)
 
@@ -167,21 +168,22 @@ class DataFrameSchema():
     def validate(
             self,
             dataframe: pd.DataFrame,
-            head: int = None,
-            tail: int = None,
-            sample: int = None,
-            random_state: Optional[int] = None) -> pd.DataFrame:
+            head: Optional[int] = None,
+            tail: Optional[int] = None,
+            sample: Optional[int] = None,
+            random_state: Optional[int] = None
+    ) -> pd.DataFrame:
+        # pylint: disable=duplicate-code
         """Check if all columns in a dataframe have a column in the Schema.
 
         :param pd.DataFrame dataframe: the dataframe to be validated.
         :param head: validate the first n rows. Rows overlapping with `tail` or
             `sample` are de-duplicated.
-        :type head: int
         :param tail: validate the last n rows. Rows overlapping with `head` or
             `sample` are de-duplicated.
-        :type tail: int
         :param sample: validate a random sample of n rows. Rows overlapping
             with `head` or `tail` are de-duplicated.
+        :param random_state: random seed for the ``sample`` argument.
         :returns: validated ``DataFrame``
 
         :raises SchemaError: when ``DataFrame`` violates built-in or custom
@@ -345,11 +347,11 @@ class SeriesSchemaBase():
     def __init__(
             self,
             pandas_dtype: Union[str, dtypes.PandasDtype] = None,
-            checks: callable = None,
+            checks: Optional[Union[Check, List[Check]]] = None,
             nullable: bool = False,
             allow_duplicates: bool = True,
             coerce: bool = False,
-            name: str = None):
+            name: str = None) -> None:
         """Initialize series schema base object.
 
         :param pandas_dtype: datatype of the column. If a string is specified,
@@ -387,12 +389,12 @@ class SeriesSchemaBase():
         return self._coerce
 
     @property
-    def name(self) -> str:
+    def name(self) -> Union[str, None]:
         """Get SeriesSchema name."""
         return self._name
 
     @property
-    def dtype(self) -> str:
[…]
