### 2020-02-23 f8de6f7
update pandera docs intro text
