### 2020-03-02 db991ac
update documentation, logo, better API reference
