+++ docs/source/conf.py
@@ -74,9 +74,11 @@ html_theme = 'sphinx_rtd_theme'
 # Theme options are theme-specific and customize the look and feel of a theme
 # further.  For a list of options available for each theme, see the
 # documentation.
-#
+
+html_logo = "_static/pandera-banner-white.png"
 html_theme_options = {
     "collapse_navigation": False,
+    "logo_only": True,
 }
 
 # Add any paths that contain custom static files (such as style sheets) here,
