### 2020-03-03 6643c5e
minor corrections to typing
