### 2020-03-03 8eba610
feature/updated-docs (#180)
* update documentation, logo, better API reference

* add banner, update dev requirements

* minor corrections to typing

* fix travis.yml
