### 2020-03-03 3b1e5c1
make README logo larger
