### 2020-03-04 834d0e0
re-style documentation, update intro copy and title
