+++ docs/source/checks.rst
@@ -8,13 +8,13 @@ Checks
 Checking column properties
 --------------------------
 
-:class:`~pandera.checks.Check` objects accept a function as a required argument, which
-is expected to have the following signature:
+:class:`~pandera.checks.Check` objects accept a function as a required
+argument, which is expected to have the following signature:
 
-``pd.Series -> bool|pd.Series[bool]``.
+``pd.Series -> Union[bool, pd.Series]``
 
-For the :class:`~pandera.checks.Check` to pass, all of the elements in the boolean
-series must evaluate to ``True``.
+For the :class:`~pandera.checks.Check` to pass, all of the elements in the
+boolean series must evaluate to ``True``.
 
 
 .. testcode:: checks
@@ -41,10 +41,11 @@ Built-in Checks
 ---------------
 
 For common validation tasks built-in checks are available in ``pandera``.
-They are provided as factory methods in the :class:`~pandera.checks.Check` class.
+They are provided as factory methods in the :class:`~pandera.checks.Check`
+class.
 
-This way comparison operations, string validations and whitelisting or blacklisting
-of allowed values can be done more easily:
+This way comparison operations, string validations and whitelisting or
+blacklisting of allowed values can be done more easily:
 
 .. testcode:: builtin_checks
 
@@ -57,12 +58,13 @@ of allowed values can be done more easily:
       "phone_number": Column(pa.String, [Check.str_matches(r'^[a-z0-9-]+$')]),
   })
 
+
 Vectorized vs. Element-wise Checks
 ----------------------------------
 
-By default, :class:`~pandera.checks.Check` objects operate on ``pd.Series`` objects.
-If you want to make atomic checks for each element in the Column, then you can provide
-the ``element_wise=True`` keyword argument:
+By default, :class:`~pandera.checks.Check` objects operate on ``pd.Series``
+objects. If you want to make atomic checks for each element in the Column, then
+you can provide the ``element_wise=True`` keyword argument:
 
 .. testcode:: vectorized_element_wise_checks
 
@@ -97,16 +99,18 @@ checks.
 Column Check Groups
 -------------------
 
-:class:`~pandera.schema_components.Column` checks support grouping by a different column so that
-you can make assertions about subsets of the :class:`~pandera.schema_components.Column` of interest.
-This changes the function signature of the :class:`~pandera.checks.Check` function so that
-its input is a dict where keys are the group names and values are subsets
-of the :class:`~pandera.schema_components.Column` series.
+:class:`~pandera.schema_components.Column` checks support grouping by a
+different column so that you can make assertions about subsets of the
+:class:`~pandera.schema_components.Column` of interest. This changes the
+function signature of the :class:`~pandera.checks.Check` function so that its
+input is a dict where keys are the group names and values are subsets of the
+:class:`~pandera.schema_components.Column` series.
 
 Specifying ``groupby`` as a column name, list of column names, or
-callable changes the expected signature of the :class:`~pandera.checks.Check` function
-argument to ``Dict[Union[Any, Tuple[Any]], pd.Series] -> Union[bool, pd.Series]`` where
-the dict keys are the discrete keys in the ``groupby`` columns.
+callable changes the expected signature of the :class:`~pandera.checks.Check`
+function argument to
+``Dict[Any, pd.Series] -> Union[bool, pd.Series]``
+where the dict keys are the discrete keys in the ``groupby`` columns.
 
 .. testcode:: column_check_groups
 
@@ -148,9 +152,9 @@ the dict keys are the discrete keys in the ``groupby`` columns.
 
     schema.validate(df)
 
-In the above example we define a :class:`~pandera.schemas.DataFrameSchema` with column
[…]
