### 2020-03-04 bd2d507
fix readme intro snippet
