### 2020-03-08 0bc680e
add raise_warning keyword arg to Checks
