+++ tests/test_checks_builtin.py
@@ -5,6 +5,9 @@ import pandas as pd
 import pytest
 
 from pandera.checks import Check
+from pandera.schemas import DataFrameSchema, SeriesSchema
+from pandera.schema_components import Column
+from pandera.errors import SchemaError
 
 
 def check_values(values, check, expected_failure_cases):
@@ -39,6 +42,28 @@ def check_none_failures(values, check):
     assert not check_result.check_passed, "Check should fail due to None value"
     assert check_result.checked_object is series, "Wrong checked_object returned"
     assert check_result.failure_cases.isnull().all(), "Only null values should be failure cases"
+
+
+def check_raise_error_or_warning(failure_values, check):
+    """
+    Check that Series and DataFrameSchemas raise warnings instead of exceptions
+
+    NOTE: it's not ideal that we have to import schema and schemaa_components
+    modules into this test module to test this functionality, this doesn't
+    separate the units under test very well.
+    """
+    failure_series = pd.Series(failure_values)
+    failure_df = pd.DataFrame({"failure_column": failure_values})
+    check.raise_warning = False
+
+    with pytest.raises(SchemaError):
+        SeriesSchema(checks=check)(failure_series)
+        DataFrameSchema({"failure_column": Column(checks=check)})(failure_df)
+
+    check.raise_warning = True
+    with pytest.warns(UserWarning):
+        SeriesSchema(checks=check)(failure_series)
+        DataFrameSchema({"failure_column": Column(checks=check)})(failure_df)
 
 
 class TestGreaterThan:
@@ -78,6 +103,7 @@ class TestGreaterThan:
     def test_failing(values, min_val, failure_cases):
         """Run checks which should fail"""
         check_values(values, Check.greater_than(min_val), failure_cases)
+        check_raise_error_or_warning(values, Check.greater_than(min_val))
 
     @staticmethod
     @pytest.mark.parametrize('values, min_val', [
@@ -127,6 +153,8 @@ class TestGreaterThanOrEqualTo:
     def test_failing(values, min_val, failure_cases):
         """Run checks which should fail"""
         check_values(values, Check.greater_than_or_equal_to(min_val), failure_cases)
+        check_raise_error_or_warning(
+            values, Check.greater_than_or_equal_to(min_val))
 
     @staticmethod
     @pytest.mark.parametrize('values, min_val', [
@@ -176,6 +204,7 @@ class TestLessThan:
     def test_failing(values, max_value, failure_cases):
         """Run checks which should fail"""
         check_values(values, Check.less_than(max_value), failure_cases)
+        check_raise_error_or_warning(values, Check.less_than(max_value))
 
     @staticmethod
     @pytest.mark.parametrize('values, max_value', [
@@ -224,7 +253,10 @@ class TestLessThanOrEqualTo:
     ])
     def test_failing(values, max_value, failure_cases):
         """Run checks which should fail"""
-        check_values(values, Check.less_than_or_equal_to(max_value), failure_cases)
+        check_values(
+            values, Check.less_than_or_equal_to(max_value), failure_cases)
+        check_raise_error_or_warning(
+            values, Check.less_than_or_equal_to(max_value))
 
     @staticmethod
     @pytest.mark.parametrize('values, max_value', [
@@ -290,6 +322,7 @@ class TestInRange:
     def test_failing(values, check_args, failure_cases):
         """Run checks which should fail"""
         check_values(values, Check.in_range(*check_args), failure_cases)
+        check_raise_error_or_warning(values, Check.in_range(*check_args))
 
     @staticmethod
[…]
