### 2020-03-08 b3e2942
fix doctest docs output for raise warning
