### 2020-03-08 6185210
update pylint settings
