+++ .pre-commit-config.yaml
@@ -3,7 +3,7 @@ repos:
     rev: pylint-2.4.4
     hooks:
     -   id: pylint
-        args: ["--disable=import-error,too-many-instance-attributes"]
+        args: ["--disable=import-error"]
         exclude: (^docs/|^scripts)
   - repo: https://github.com/pre-commit/mirrors-mypy
     rev: v0.761
