### 2020-03-08 1a98c4e
fix warning doctest for py3.5
