### 2020-03-08 a87c71d
add raise_warning keyword arg to Checks (#182)
* add raise_warning keyword arg to Checks

* fix doctest docs output for raise warning

* update pylint settings

* fix warning doctest for py3.5

* fix check
