+++ pandera/checks.py
@@ -33,7 +33,8 @@ class Check():
             groupby: Optional[Union[str, List[str], Callable]] = None,
             element_wise: bool = False,
             error: Optional[str] = None,
-            n_failure_cases: Optional[int] = constants.N_FAILURE_CASES
+            raise_warning: bool = False,
+            n_failure_cases: Union[int, None] = constants.N_FAILURE_CASES,
     ) -> None:
         """Apply a validation function to each element, Series, or DataFrame.
 
@@ -77,6 +78,10 @@ class Check():
             number of elements as checks.
         :param error: custom error message if series fails validation
             check.
+        :param raise_warning: if True, raise a UserWarning and do not throw
+            exception instead of raising a SchemaError for a specific check.
+            This option should be used carefully in cases where a failing
+            check is informational and shouldn't stop execution of the program.
         :param n_failure_cases: report the top n failure cases. If None, then
             report all failure cases.
 
@@ -137,6 +142,7 @@ class Check():
         self.fn = fn
         self.element_wise = element_wise
         self.error = error
+        self.raise_warning = raise_warning
         self.n_failure_cases = n_failure_cases
 
         if groupby is None and groups is not None:
@@ -293,13 +299,15 @@ class Check():
             if self.error is not None else "<Check %s>" % name
 
     @staticmethod
-    def greater_than(min_value) -> 'Check':
+    def greater_than(min_value, raise_warning: bool = False) -> 'Check':
         """Create a :class:`Check` ensuring all values of a series are strictly greater
             than a certain value.
 
         :param min_value: Lower bound to be exceeded. Must be a type comparable to
             the dtype of the :class:`pandas.Series` to be validated (e.g. a numerical
             type for float or int and a datetime for datetime).
+        :param raise_warning: if True, check raises UserWarning instead of
+            SchemaError on validation.
 
         :returns: :class:`Check` object
         """
@@ -310,14 +318,21 @@ class Check():
             """Comparison function for check"""
             return series > min_value
 
-        return Check(fn=_greater_than, error="greater_than(%s)" % min_value)
+        return Check(
+            fn=_greater_than,
+            error="greater_than(%s)" % min_value,
+            raise_warning=raise_warning,
+        )
 
     @staticmethod
-    def greater_than_or_equal_to(min_value) -> 'Check':
+    def greater_than_or_equal_to(
+            min_value, raise_warning: bool = False) -> 'Check':
         """Create a :class:`Check` ensuring all values are greater or equal a certain value.
 
         :param min_value: Allowed minimum value for values of a series. Must be a type
             comparable to the dtype of the :class:`pandas.Series` to be validated.
+        :param raise_warning: if True, check raises UserWarning instead of
+            SchemaError on validation.
 
         :returns: :class:`Check` object
         """
@@ -328,16 +343,21 @@ class Check():
             """Comparison function for check"""
             return series >= min_value
 
-        return Check(fn=_greater_or_equal,
-                     error="greater_than_or_equal_to(%s)" % min_value)
+        return Check(
+            fn=_greater_or_equal,
+            error="greater_than_or_equal_to(%s)" % min_value,
+            raise_warning=raise_warning,
+        )
 
     @staticmethod
-    def less_than(max_value) -> 'Check':
+    def less_than(max_value, raise_warning: bool = False) -> 'Check':
         """Create a :class:`Check` ensuring all values are strictly below a certain value.
 
         :param max_value: All elements of a series must be strictly smaller than this.
             Must be a type comparable to the dtype of the :class:`pandas.Series` to be
             validated.
+        :param raise_warning: if True, check raises UserWarning instead of
[…]
