### 2020-03-08 dc0e372
add support for Column(..., regex=True)
