+++ tests/test_schema_components.py
@@ -233,3 +233,126 @@ def test_schema_component_equality_operators():
     assert index != not_equal_schema
     assert multi_index == copy.deepcopy(multi_index)
     assert multi_index != not_equal_schema
+
+
+def test_column_regex():
+    """Test that column regex work on single-level column index."""
+    column_schema = Column(
+        Int, Check(lambda s: s >= 0), name="foo_*", regex=True)
+    data = pd.DataFrame({
+        "foo_1": range(10),
+        "foo_2": range(10, 20),
+        "foo_3": range(20, 30),
+        "bar_1": range(10),
+        "bar_2": range(10, 20),
+        "bar_3": range(20, 30),
+    })
+    column_schema.validate(data)
+
+    # Raise an error on multi-index column case
+    data.columns = pd.MultiIndex.from_tuples(
+        (
+            ("foo_1", "biz_1"),
+            ("foo_2", "baz_1"),
+            ("foo_3", "baz_2"),
+            ("bar_1", "biz_2"),
+            ("bar_2", "biz_3"),
+            ("bar_3", "biz_3"),
+        )
+    )
+    with pytest.raises(IndexError):
+        column_schema.validate(data)
+
+
+
+def test_column_regex_multiindex():
+    """Text that column regex works on multi-index column."""
+    column_schema = Column(
+        Int, Check(lambda s: s >= 0), name=("foo_*", "baz_*"), regex=True,
+    )
+    data = pd.DataFrame({
+        ("foo_1", "biz_1"): range(10),
+        ("foo_2", "baz_1"): range(10, 20),
+        ("foo_3", "baz_2"): range(20, 30),
+        ("bar_1", "biz_2"): range(10),
+        ("bar_2", "biz_3"): range(10, 20),
+        ("bar_3", "biz_3"): range(20, 30),
+    })
+    column_schema.validate(data)
+
+    # Raise an error if tuple column name is applied to a dataframe with a
+    # flat pd.Index object.
+    failure_column_cases = (
+        ["foo_%s" % i for i in range(6)],
+        pd.MultiIndex.from_tuples([
+            ("foo_%s" % i, "bar_%s" % i, "baz_%s" % i) for i in range(6)
+        ])
+    )
+    for columns in failure_column_cases:
+        data.columns = columns
+        with pytest.raises(IndexError):
+            column_schema.validate(data)
+
+
+@pytest.mark.parametrize("column_name_regex, expected_matches, error", (
+    # match all values in first level, only baz_* for second level
+    (
+        (".", "baz_*"),
+        [("foo_2", "baz_1"), ("foo_3", "baz_2")],
+        None,
+    ),
+
+    # match bar_* in first level, all values in second level
+    (
+        ("bar_*", "."),
+        [("bar_1", "biz_2"), ("bar_2", "biz_3"), ("bar_3", "biz_3")],
+        None,
+    ),
+
+    # match specific columns in both levels
+    (
+        ("foo_*", "baz_*"),
+        [("foo_2", "baz_1"), ("foo_3", "baz_2")],
+        None,
+    ),
+    (("foo_*", "^biz_1$"), [("foo_1", "biz_1")], None),
+    (("^foo_3$", "^baz_2$"), [("foo_3", "baz_2")], None),
+
+    # no matches should raise a SchemaError
+    (("fiz", "."), None, errors.SchemaError),
+
+    # using a string name for a multi-index column raises IndexError
+    ("foo_1", None, IndexError),
+
+    # mis-matching number of elements in a tuple column name raises IndexError
+    (("foo_*", ), None, IndexError),
+    (("foo_*", ".", "."), None, IndexError),
+    (("foo_*", ".", ".", "."), None, IndexError),
+))
+def test_column_regex_matching(
[…]
