### 2020-03-09 4588520
add more unit tests for regex functionality, works in DataFrameSchema
