+++ tests/test_schema_components.py
@@ -239,6 +239,11 @@ def test_column_regex():
     """Test that column regex work on single-level column index."""
     column_schema = Column(
         Int, Check(lambda s: s >= 0), name="foo_*", regex=True)
+
+    dataframe_schema = DataFrameSchema({
+        "foo_*": Column(Int, Check(lambda s: s >= 0), regex=True),
+    })
+
     data = pd.DataFrame({
         "foo_1": range(10),
         "foo_2": range(10, 20),
@@ -247,7 +252,8 @@ def test_column_regex():
         "bar_2": range(10, 20),
         "bar_3": range(20, 30),
     })
-    column_schema.validate(data)
+    assert isinstance(column_schema.validate(data), pd.DataFrame)
+    assert isinstance(dataframe_schema.validate(data), pd.DataFrame)
 
     # Raise an error on multi-index column case
     data.columns = pd.MultiIndex.from_tuples(
@@ -262,6 +268,8 @@ def test_column_regex():
     )
     with pytest.raises(IndexError):
         column_schema.validate(data)
+    with pytest.raises(IndexError):
+        dataframe_schema.validate(data)
 
 
 
@@ -270,6 +278,10 @@ def test_column_regex_multiindex():
     column_schema = Column(
         Int, Check(lambda s: s >= 0), name=("foo_*", "baz_*"), regex=True,
     )
+    dataframe_schema = DataFrameSchema({
+        ("foo_*", "baz_*"): Column(Int, Check(lambda s: s >= 0), regex=True),
+    })
+
     data = pd.DataFrame({
         ("foo_1", "biz_1"): range(10),
         ("foo_2", "baz_1"): range(10, 20),
@@ -278,7 +290,8 @@ def test_column_regex_multiindex():
         ("bar_2", "biz_3"): range(10, 20),
         ("bar_3", "biz_3"): range(20, 30),
     })
-    column_schema.validate(data)
+    assert isinstance(column_schema.validate(data), pd.DataFrame)
+    assert isinstance(dataframe_schema.validate(data), pd.DataFrame)
 
     # Raise an error if tuple column name is applied to a dataframe with a
     # flat pd.Index object.
@@ -292,29 +305,21 @@ def test_column_regex_multiindex():
         data.columns = columns
         with pytest.raises(IndexError):
             column_schema.validate(data)
+        with pytest.raises(IndexError):
+            dataframe_schema.validate(data)
 
 
 @pytest.mark.parametrize("column_name_regex, expected_matches, error", (
     # match all values in first level, only baz_* for second level
-    (
-        (".", "baz_*"),
-        [("foo_2", "baz_1"), ("foo_3", "baz_2")],
-        None,
-    ),
+    ((".", "baz_*"), [("foo_2", "baz_1"), ("foo_3", "baz_2")], None),
 
     # match bar_* in first level, all values in second level
-    (
-        ("bar_*", "."),
-        [("bar_1", "biz_2"), ("bar_2", "biz_3"), ("bar_3", "biz_3")],
-        None,
-    ),
+    (("bar_*", "."),
+     [("bar_1", "biz_2"), ("bar_2", "biz_3"), ("bar_3", "biz_3")],
+     None),
 
     # match specific columns in both levels
-    (
-        ("foo_*", "baz_*"),
-        [("foo_2", "baz_1"), ("foo_3", "baz_2")],
-        None,
-    ),
+    (("foo_*", "baz_*"), [("foo_2", "baz_1"), ("foo_3", "baz_2")], None),
     (("foo_*", "^biz_1$"), [("foo_1", "biz_1")], None),
     (("^foo_3$", "^baz_2$"), [("foo_3", "baz_2")], None),
 
[…]
