### 2020-03-09 27dab4d
fix docs on column regex
