### 2020-03-10 1b94b34
make checks/regex docs deterministic
