### 2020-03-10 d06c561
fix pylint complaints
