### 2020-03-11 36c3aff
add test for strict regex matching
