### 2020-03-11 2feb0c9
feature/column-regex (#184)
This PR implements regex pattern matching on Columns. This works
only on string columns, and can be applied to single-level column
indexes and multi-index columns.

* add support for Column(..., regex=True)

* add more unit tests for regex functionality, works in DataFrameSchema

* fix docs on column regex

* make checks/regex docs deterministic

* fix pylint complaints

* add test for strict regex matching
