+++ tests/test_schemas.py
@@ -70,13 +70,37 @@ def test_dataframe_schema():
 
 
 def test_dataframe_schema_strict():
-    """checks if strict=True whether a schema error is raised because 'a' is
-    not present in the dataframe."""
-    schema = DataFrameSchema({"a": Column(Int, nullable=True)},
-                             strict=True)
+    """
+    Checks if strict=True whether a schema error is raised because 'a' is
+    not present in the dataframe.
+    """
+    schema = DataFrameSchema(
+        {"a": Column(Int, nullable=True)},
+        strict=True
+    )
     df = pd.DataFrame({"b": [1, 2, 3]})
     with pytest.raises(errors.SchemaError):
         schema.validate(df)
+
+
+def test_dataframe_schema_strict_regex():
+    """Test that strict dataframe schema checks for regex matches."""
+    schema = DataFrameSchema(
+        {"foo_*": Column(Int, regex=True)},
+        strict=True,
+    )
+    df = pd.DataFrame({
+        "foo_%d" % i: range(10) for i in range(5)
+    })
+
+    assert isinstance(schema.validate(df), pd.DataFrame)
+
+    # Raise a SchemaError if schema is strict and a regex pattern yields
+    # no matches
+    with pytest.raises(errors.SchemaError):
+        schema.validate(
+            pd.DataFrame({"bar_%d" % i: range(10) for i in range(5)})
+        )
 
 
 def test_series_schema():
