+++ .pre-commit-config.yaml
@@ -4,7 +4,7 @@ repos:
     hooks:
     -   id: pylint
         args: ["--disable=import-error"]
-        exclude: (^docs/|^scripts)
+        exclude: (^docs/|^scripts|setup.py)
   - repo: https://github.com/pre-commit/mirrors-mypy
     rev: v0.761
     hooks:
