### 2020-03-13 ff5eb06
make pandera PandasDtype enum more robust, more docs
this diff makes the PandasDtype enum more robust by mapping
Int to the "int" numpy dtype alias and Float to the "float"
numpy dtype alias.
