+++ pandera/dtypes.py
@@ -1,52 +1,80 @@
 """Schema datatypes."""
-# pylint: disable=C0103
 
 from enum import Enum
 
 
 class PandasDtype(Enum):
-    """Enumerate all valid pandas data types."""
+    """Enumerate all valid pandas data types.
 
-    Bool = "bool"
-    DateTime = "datetime64[ns]"
-    Category = "category"
-    Float = "float64"
-    Float16 = "float16"
-    Float32 = "float32"
-    Float64 = "float64"
-    Int = "int64"
-    Int8 = "int8"
-    Int16 = "int16"
-    Int32 = "int32"
-    Int64 = "int64"
-    UInt8 = "uint8"
-    UInt16 = "uint16"
-    UInt32 = "uint32"
-    UInt64 = "uint64"
-    Object = "object"
-    # the string datatype doesn't map to a unique string representation and is
-    # representated as a numpy object array. This will change after pandas 1.0,
-    # but for now will need to handle this as a special case.
+    ``pandera`` follows the
+    `numpy data types <https://docs.scipy.org/doc/numpy/reference/arrays.dtypes.html#arrays-dtypes>`_  # pylint: disable=line-too-long
+    subscribed to by ``pandas`` and by default supports using the numpy data
+    type string aliases to validate DataFrame or Series dtypes.
+
+    This class simply enumerates the valid numpy dtypes for pandas arrays.
+    For convenience ``PandasDtype`` enums can all be accessed in the top-level
+    ``pandera`` name space via the same enum name.
+
+    .. warning::
+        ``pandera`` also offers limited support for
+        `pandas extension types <https://pandas.pydata.org/pandas-docs/stable/getting_started/basics.html#dtypes>`_,  # pylint: disable=line-too-long
+        however since the release of pandas 1.0.0 there are backwards
+        incompatible extension types like the nullable ``Integer`` array and
+        the dedicated ``String`` array. In theory the string aliases for these
+        extension types should work when supplied to the ``pandas_dtype``
+        argument when initializing ``pa.SeriesSchemaBase`` objects, but this
+        is not currently tested.
+
+    :examples:
+
+    >>> import pandera as pa
+    >>>
+    >>> pa.Int
+    <PandasDtype.Int: 'int'>
+    >>> pa.Float
+    <PandasDtype.Float: 'float'>
+    >>> pa.String
+    <PandasDtype.String: 'string'>
+    >>> # etc...
+
+    You can also directly use the string alias for each data-type in the
+    schema definition:
+
+    >>> series_schema = pa.SeriesSchema("int")
+    >>> series_schema.validate(pd.Series([1, 2, 3]))
+    0    1
+    1    2
+    2    3
+    dtype: int64
+    """
+
+    Bool = "bool"  #: ``"bool"`` numpy dtype
+    DateTime = "datetime64[ns]" #: ``"datetime64[ns]"`` numpy dtype
+    Category = "category" #: pandas ``"categorical"`` datatype
+    Float = "float"  #: ``"float"`` numpy dtype
+    Float16 = "float16"  #: ``"float16"`` numpy dtype
+    Float32 = "float32"  #: ``"float32"`` numpy dtype
+    Float64 = "float64"  #: ``"float64"`` numpy dtype
+    Int = "int"  #: ``"int"`` numpy dtype
+    Int8 = "int8"  #: ``"int8"`` numpy dtype
+    Int16 = "int16"  #: ``"int16"`` numpy dtype
[…]
