### 2020-03-14 4bc2047
update dtype docs
