### 2020-03-14 7456f07
support dataframe bool output for Check.
- update documentation
