+++ docs/source/dataframe_schemas.rst
@@ -1,12 +1,14 @@
 .. pandera documentation for DataFrameSchemas
 
+.. currentmodule:: pandera
+
 .. _DataFrameSchemas:
 
 DataFrame Schemas
 =================
 
-The ``DataFrameSchema`` class enables the specification of a schema that
-verifies the columns and index of a ``pd.DataFrame`` object.
+The :py:class:`DataFrameSchema` class enables the specification of a schema
+that verifies the columns and index of a pandas ``DataFrame`` object.
 
 The ``DataFrameSchema`` object consists of |column|_\s and an |index|_.
 
@@ -41,16 +43,16 @@ The ``DataFrameSchema`` object consists of |column|_\s and an |index|_.
 Column Validation
 -----------------
 
-A ``Column`` must specify the properties of a column in a dataframe object.
-It can be optionally verified for its data type, `null values`_ or duplicate
-values. The column can be coerced_ into the specified type, and the
+A :py:class:`Column` must specify the properties of a column in a dataframe
+object. It can be optionally verified for its data type, `null values`_ or
+duplicate values. The column can be coerced_ into the specified type, and the
 required_ parameter allows control over whether or not the column is allowed to
 be missing.
 
 :ref:`Column checks<checks>` allow for the DataFrame's values to be
 checked against a user provided function. ``Check`` objects also support
 :ref:`grouping<grouping>` by a different column so that the user can make
-assertions about subsets of the ``Column`` of interest.
+assertions about subsets of the column of interest.
 
 Column Hypotheses enable you to perform statistical hypothesis tests on a
 DataFrame in either wide or tidy format. See
@@ -89,9 +91,12 @@ nullable. In order to accept null values, you need to explicitly specify
     SchemaError: non-nullable series contains null values: {2: nan}
 
 .. note:: Due to a known limitation in
-    `pandas <http://pandas.pydata.org/pandas-docs/stable/gotchas.html#support-for-integer-na>`__,
+    `pandas prior to version 0.24.0 <https://pandas.pydata.org/pandas-docs/stable/user_guide/integer_na.html>`_,
     integer arrays cannot contain ``NaN`` values, so this schema will return
     a DataFrame where ``column1`` is of type ``float``.
+    :py:class:`PandasDtype` does not currently support the nullable integer
+    array type, but you can still use the "Int64" string alias for nullable
+    integer arrays
 
 .. testcode:: null_values_in_columns
 
@@ -397,7 +402,7 @@ schema, specify ``strict=True``:
 Index Validation
 ----------------
 
-You can also specify an ``Index`` in the ``DataFrameSchema``.
+You can also specify an :py:class:`Index` in the :py:class:`DataFrameSchema`.
 
 .. testcode:: index_validation
 
@@ -494,8 +499,8 @@ tuples for each level in the index hierarchy:
 MultiIndex Indexes
 ~~~~~~~~~~~~~~~~~~
 
-The ``pandera.MultiIndex`` class allows you to define multi-index indexes by
-composing a list of ``pandera.Index`` objects.
+The :py:class:`pandera.MultiIndex` class allows you to define multi-index
+indexes by composing a list of ``pandera.Index`` objects.
 
 .. testcode:: multiindex_indexes
 
@@ -551,19 +556,18 @@ Some examples of where this can be provided to pandas are:
   import pandas as pd
   import pandera as pa
 
-  from pandera import Column, DataFrameSchema, Index, MultiIndex, Check
-
-  schema = DataFrameSchema(
[…]
