### 2020-03-14 f61a718
implement get_dtype method
