### 2020-03-14 a51eca8
update hypothesis and decorator docs
