### 2020-03-14 66b618f
update series schema docs
