### 2020-03-15 b9cf9dc
fix numpy type matching
