### 2020-03-15 0f1b43f
use numpy string alias
