+++ tests/test_schemas.py
@@ -469,7 +469,7 @@ def test_dataframe_schema_dtype_property():
         }
     )
     assert schema.dtype == {
-        "col1": "int",
+        "col1": "int64",
         "col2": "object",
         "col3": "datetime64[ns]",
         "col4": "uint16"
@@ -584,15 +584,15 @@ def test_schema_get_dtype():
     })
 
     with pytest.warns(UserWarning) as record:
-        assert schema.dtype == {"col1": "int"}
+        assert schema.dtype == {"col1": "int64"}
     assert len(record) == 1
     assert record[0].message.args[0].startswith(
         "Schema has columns specified as regex column names:"
     )
 
     assert schema.get_dtype(data) == {
-        "col1": "int",
-        "var1": "float",
-        "var2": "float",
-        "var3": "float",
+        "col1": "int64",
+        "var1": "float64",
+        "var2": "float64",
+        "var3": "float64",
     }
