+++ tests/test_dtypes.py
@@ -15,8 +15,8 @@ TESTABLE_DTYPES = [
     (Bool, "bool"),
     (DateTime, "datetime64[ns]"),
     (Category, "category"),
-    (Float, "float"),
-    (Int, "int"),
+    (Float, "float64"),
+    (Int, "int64"),
     (Object, "object"),
     (String, "object"),
     (Timedelta, "timedelta64[ns]"),
