### 2020-03-15 b2b4f85
use numpy str on test
