### 2020-03-15 4f52318
try defining array with actual data
