+++ pandera/dtypes.py
@@ -13,8 +13,8 @@ NUMPY_NONNULLABLE_INT_DTYPES = [
 # for int and float dtype, delegate string representation to the
 # default based on OS. In Windows, pandas defaults to int64 while numpy
 # defaults to int32.
-_DEFAULT_INT_TYPE = pd.Series(dtype="int").dtype
-_DEFAULT_FLOAT_TYPE = pd.Series(dtype="float").dtype
+_DEFAULT_INT_TYPE = pd.Series([1]).dtype
+_DEFAULT_FLOAT_TYPE = pd.Series([1.0]).dtype
 
 
 class PandasDtype(Enum):
@@ -98,8 +98,8 @@ class PandasDtype(Enum):
     Timedelta = "timedelta64[ns]"  #: ``"timedelta64[ns]"`` numpy dtype
 
     @property
-    def numpy_str(self):
-        """Get numpy datatype string alias."""
+    def str_alias(self):
+        """Get datatype string alias."""
         return {
             "int": str(_DEFAULT_INT_TYPE),
             "float": str(_DEFAULT_FLOAT_TYPE),
