+++ tests/test_dtypes.py
@@ -15,10 +15,10 @@ TESTABLE_DTYPES = [
     (Bool, "bool"),
     (DateTime, "datetime64[ns]"),
     (Category, "category"),
-    (Float, "float64"),
-    (Int, "int64"),
+    (Float, Float.str_alias),
+    (Int, Int.str_alias),
     (Object, "object"),
-    (String, "object"),
+    (String, String.str_alias),
     (Timedelta, "timedelta64[ns]"),
     ("bool", "bool"),
     ("datetime64[ns]", "datetime64[ns]"),
