### 2020-03-15 1b01f29
fix dtype tests
