### 2020-03-15 e18260a
dtype int/float default: use empty series
