### 2020-03-15 4f2ee8e
use series with data for default int/float
