### 2020-03-16 dc9be10
minor update to docs
