### 2020-03-16 c5db3c1
better handling of pandas/numpy dtypes with documentation (#187)
* make pandera PandasDtype enum more robust, more docs

this diff makes the PandasDtype enum more robust by mapping
Int to the "int" numpy dtype alias and Float to the "float"
numpy dtype alias. It also adds support for pandas extension
types as inputs to the `pandas_dtype` argument.
