+++ tests/test_schemas.py
@@ -271,8 +271,8 @@ def test_coerce_dtype_in_dataframe():
 
     for schema in [schema1, schema2]:
         result = schema.validate(df)
-        assert result.column1.dtype == Int.value
-        assert result.column2.dtype == DateTime.value
+        assert result.column1.dtype == Int.str_alias
+        assert result.column2.dtype == DateTime.str_alias
         for _, x in result.column3.iteritems():
             assert pd.isna(x) or isinstance(x, str)
 
@@ -567,3 +567,32 @@ def test_add_and_remove_columns():
         }, strict=True)
 
     assert schema4 == expected_schema_4 == schema1
+
+
+def test_schema_get_dtype():
+    """Test that schema dtype and get_dtype methods handle regex columns."""
+    schema = DataFrameSchema({
+        "col1": Column(Int),
+        "var*": Column(Float, regex=True),
+    })
+
+    data = pd.DataFrame({
+        "col1": [1, 2, 3],
+        "var1": [1.0, 1.1, 1.2],
+        "var2": [1.0, 1.1, 1.2],
+        "var3": [1.0, 1.1, 1.2],
+    })
+
+    with pytest.warns(UserWarning) as record:
+        assert schema.dtype == {"col1": Int.str_alias}
+    assert len(record) == 1
+    assert record[0].message.args[0].startswith(
+        "Schema has columns specified as regex column names:"
+    )
+
+    assert schema.get_dtype(data) == {
+        "col1": Int.str_alias,
+        "var1": Float.str_alias,
+        "var2": Float.str_alias,
+        "var3": Float.str_alias,
+    }
