### 2020-03-17 4f58420
fix doc dtype in typo
