### 2020-03-17 a9d13e7
add code-cov make target for local code coverage report
