### 2020-03-17 f473949
clean up Makefile
