### 2020-03-17 a71f1d3
bump to version 0.3.2
- delete conda recipe, conda is maintained via conda-forge
