+++ pandera/__init__.py
@@ -30,4 +30,4 @@ String = PandasDtype.String
 Timedelta = PandasDtype.Timedelta
 
 
-__version__ = "0.3.1"
+__version__ = "0.3.2"
