+++ conda.recipe/build.sh
@@ -1,3 +1,0 @@
-#!/bin/sh
-pip install schema
-$PYTHON setup.py install
