+++ conda.recipe/meta.yaml
@@ -1,38 +1,0 @@
-package:
-  name: pandera
-  version: "0.3.1"
-
-source:
-  git_rev: 0.3.1
-  git_url: https://github.com/pandera-dev/pandera.git
-
-build:
-  noarch: python
-
-requirements:
-  build:
-    - python >=3.5
-    - numpy
-    - pandas
-    - scipy
-    - wrapt
-  run:
-    - python >=3.5
-    - numpy
-    - pandas
-    - scipy
-    - wrapt
-
-test:
-  imports:
-    - pandera
-
-about:
-  home: https://github.com/pandera-dev/pandera
-  license: MIT
-  summary: 'Light-weight and flexible validation for pandas data structures.'
-  description: |
-    Pandas data structures contain information that pandera explicitly
-    validates at runtime, useful in production-critical or reproducible
-    research settings.
-  doc_url: https://pandera.readthedocs.io/
