### 2020-03-21 9e2f610
fix hypothesis documentation
