### 2020-03-25 b6108d0
fix series schema documentation
