+++ pandera/schemas.py
@@ -603,12 +603,11 @@ class SeriesSchemaBase():
                 (series.name, _dtype, series.dtype))
 
         check_results = []
-        if len(self.checks) != 0:
+        if self.checks:
             if isinstance(check_obj, pd.Series):
                 check_args = [series, None]
             else:
                 _check_obj = check_obj.loc[series.index.unique()].copy()
-                _check_obj[self.name] = series
                 check_args = [_check_obj, self.name]
 
             for check_index, check in enumerate(self.checks):
