### 2020-04-03 58dca9b
fix mypy error in builtin checks test
