### 2020-04-03 3fb4cd4
Fix #188 (#189)
* Update schemas.py

* Update schemas.py

* Update schemas.py

* fix mypy error in builtin checks test

* Update schemas.py

* Update schemas.py

* Update schemas.py

Co-authored-by: cosmicBboy <niels.bantilan@gmail.com>
