### 2020-04-03 dde9d00
Add test to cover #191
