### 2020-04-03 5e3f925
Make pandas_dtype a read only property.
Partialy addresses #191 by preventing assignment of Column.pandas_dtype
instead of exposing silently an inconsistent behaviour.
