### 2020-04-03 a249765
Add setter to Column=pandas_dtypes to fix #191
TODO: Double check whether any check of action is required on the value
being set.
