### 2020-04-03 0819ed9
More detailed error when coercing types
Report the name of the columns whose type is missing.
