### 2020-04-03 4534cdd
WIP More detailed error for unrecognized type
