### 2020-04-04 3aacbf4
update and clean up docs
- add __call__ method to docs
