### 2020-04-06 0191cdc
Add test coverage for issue #196.
Checks are applied after removal of missing values,
whicj prevent any checking of the rate or other properties
of missing values for nullable columns.
