### 2020-04-06 3dafbcf
simplify type-matching logic in SeriesSchemaBase (#194)
