### 2020-04-06 438da34
feature/define-check-base (#195)
* update and clean up docs

- add __call__ method to docs

* fix docs, update css
