+++ pandera/schemas.py
@@ -9,9 +9,17 @@ import pandas as pd
 
 from . import errors, constants, dtypes, error_formatters
 from .checks import Check, CheckResult
+from .hypotheses import Hypothesis
 
 
 N_INDENT_SPACES = 4
+
+CheckList = Optional[
+    Union[
+        Union[Check, Hypothesis],
+        List[Union[Check, Hypothesis]]
+    ]
+]
 
 
 class DataFrameSchema():
@@ -20,7 +28,7 @@ class DataFrameSchema():
     def __init__(
             self,
             columns: Dict[Any, Any] = None,
-            checks: Optional[List[Check]] = None,
+            checks: CheckList = None,
             index=None,
             transformer: Callable = None,
             coerce: bool = False,
@@ -79,7 +87,7 @@ class DataFrameSchema():
         """
         if checks is None:
             checks = []
-        if isinstance(checks, Check):
+        if isinstance(checks, (Check, Hypothesis)):
             checks = [checks]
 
         self.columns = {} if columns is None else columns
@@ -178,8 +186,8 @@ class DataFrameSchema():
 
     def get_dtype(self, dataframe: pd.DataFrame) -> Dict[str, str]:
         """
-        Same as the ``dtype`` property, but expands columns where regex == True
-        based on the supplied dataframe.
+        Same as the ``dtype`` property, but expands columns where
+        ``regex == True`` based on the supplied dataframe.
 
         :returns: dictionary of columns and their associated dtypes.
         """
@@ -333,7 +341,7 @@ class DataFrameSchema():
             tail: Optional[int] = None,
             sample: Optional[int] = None,
             random_state: Optional[int] = None):
-        """Delegate to `validate` method.
+        """Alias for :func:`DataFrameSchema.validate` method.
 
         :param pd.DataFrame dataframe: the dataframe to be validated.
         :param head: validate the first n rows. Rows overlapping with `tail` or
@@ -425,7 +433,7 @@ class SeriesSchemaBase():
             pandas_dtype: Union[
                 str, dtypes.PandasDtype, dtypes.PandasExtensionType
             ] = None,
-            checks: Optional[Union[Check, List[Check]]] = None,
+            checks: CheckList = None,
             nullable: bool = False,
             allow_duplicates: bool = True,
             coerce: bool = False,
@@ -437,8 +445,10 @@ class SeriesSchemaBase():
             http://pandas.pydata.org/pandas-docs/stable/basics.html#dtypes
         :param checks: If element_wise is True, then callable signature should
             be:
-            x -> x where x is a scalar element in the column. Otherwise,
-            x is assumed to be a pandas.Series object.
+
+            ``Callable[Any, bool]`` where the ``Any`` input is a scalar element
+            in the column. Otherwise, the input is assumed to be a
+            pandas.Series object.
         :type checks: callable
         :param nullable: Whether or not column can contain null values.
         :type nullable: bool
@@ -451,7 +461,7 @@ class SeriesSchemaBase():
         self._coerce = coerce
         if checks is None:
             checks = []
-        if isinstance(checks, Check):
+        if isinstance(checks, (Check, Hypothesis)):
             checks = [checks]
         self.checks = checks
         self._name = name
@@ -621,7 +631,7 @@ class SeriesSchemaBase():
             sample: Optional[int] = None,
             random_state: Optional[int] = None,
     ) -> Union[pd.DataFrame, pd.Series]:
-        """Validate a series or column in a dataframe."""
+        """Alias for ``validate`` method."""
         return self.validate(check_obj, head, tail, sample, random_state)
 
     def __eq__(self, other):
@@ -630,48 +640,6 @@ class SeriesSchemaBase():
 
[…]
