+++ pandera/checks.py
@@ -23,8 +23,8 @@ SeriesCheckObj = Union[pd.Series, Dict[str, pd.Series]]
 DataFrameCheckObj = Union[pd.DataFrame, Dict[str, pd.DataFrame]]
 
 
-class Check():
-    """Check a pandas Series or DataFrame for certain properties."""
+class _CheckBase():
+    """Check base class."""
 
     def __init__(
             self,
@@ -127,7 +127,7 @@ class Check():
         ... })
         >>>
         >>> schema.validate(df)[["measure_1", "measure_2", "group"]]
-           measure_1  measure_2 group
+            measure_1  measure_2 group
         0         10          2     B
         1         12          4     B
         2         14          6     A
@@ -136,6 +136,7 @@ class Check():
         See :ref:`here<checks>` for more usage details.
 
         """
+
         if element_wise and groupby is not None:
             raise errors.SchemaInitError(
                 "Cannot use groupby when element_wise=True.")
@@ -232,8 +233,8 @@ class Check():
     ) -> CheckResult:
         """Validate pandas DataFrame or Series.
 
-        :df_or_series: pandas DataFrame of Series to validate.
-        :column: apply the check function to this column.
+        :param df_or_series: pandas DataFrame of Series to validate.
+        :param column: apply the check function to this column.
         :returns: CheckResult tuple containing checked object,
             check validation result, and failure cases from the checked object.
         """
@@ -303,14 +304,17 @@ class Check():
         return "<Check %s: %s>" % (name, self.error) \
             if self.error is not None else "<Check %s>" % name
 
+
+class Check(_CheckBase):
+    """Check a pandas Series or DataFrame for certain properties."""
+
     @staticmethod
     def greater_than(min_value, raise_warning: bool = False) -> 'Check':
-        """Create a :class:`Check` ensuring all values of a series are strictly greater
-            than a certain value.
+        """Ensure values of a series are strictly greater than a minimum value.
 
-        :param min_value: Lower bound to be exceeded. Must be a type comparable to
-            the dtype of the :class:`pandas.Series` to be validated (e.g. a numerical
-            type for float or int and a datetime for datetime).
+        :param min_value: Lower bound to be exceeded. Must be a type comparable
+            to the dtype of the :class:`pandas.Series` to be validated (e.g. a
+            numerical type for float or int and a datetime for datetime).
         :param raise_warning: if True, check raises UserWarning instead of
             SchemaError on validation.
 
@@ -332,10 +336,11 @@ class Check():
     @staticmethod
     def greater_than_or_equal_to(
             min_value, raise_warning: bool = False) -> 'Check':
-        """Create a :class:`Check` ensuring all values are greater or equal a certain value.
+        """Ensure all values are greater or equal a certain value.
 
-        :param min_value: Allowed minimum value for values of a series. Must be a type
-            comparable to the dtype of the :class:`pandas.Series` to be validated.
+        :param min_value: Allowed minimum value for values of a series. Must be
+            a type comparable to the dtype of the :class:`pandas.Series` to be
+            validated.
         :param raise_warning: if True, check raises UserWarning instead of
             SchemaError on validation.
 
@@ -356,11 +361,11 @@ class Check():
 
     @staticmethod
     def less_than(max_value, raise_warning: bool = False) -> 'Check':
-        """Create a :class:`Check` ensuring all values are strictly below a certain value.
+        """Ensure values of a series are strictly below a maximum value.
 
-        :param max_value: All elements of a series must be strictly smaller than this.
-            Must be a type comparable to the dtype of the :class:`pandas.Series` to be
-            validated.
+        :param max_value: All elements of a series must be strictly smaller
[…]
