### 2020-04-10 2c9e62b
implement simple schema inference
- dataframe and series schemas can be inferred by providing
  `infer_schema` with the corresponding pandas object.
- extend `PandasDtype` enum class with `from_pandas_api_type`
  and `from_str_alias` methods. Also implemented `__eq__` and
  `__hash__` methods.
- implemented an `_inferred_schema_guard` function that wraps
  dataframe and series schema methods. This should raise a
  UserWarning if a user validates a dataframe/series using
  an inferred schema that hasn't been modified
