+++ pandera/schema_inference.py
@@ -1,0 +1,192 @@
+"""Module for inferring dataframe/series schema."""
+
+import warnings
+from typing import Any, Dict, Union
+
+import pandas as pd
+
+from .checks import Check
+from .dtypes import PandasDtype
+from .schemas import DataFrameSchema, SeriesSchema
+from .schema_components import Column
+
+
+NUMERIC_DTYPES = frozenset([
+    PandasDtype.Float,
+    PandasDtype.Float16,
+    PandasDtype.Float32,
+    PandasDtype.Float64,
+    PandasDtype.Int,
+    PandasDtype.Int8,
+    PandasDtype.Int16,
+    PandasDtype.Int32,
+    PandasDtype.Int64,
+    PandasDtype.UInt8,
+    PandasDtype.UInt16,
+    PandasDtype.UInt32,
+    PandasDtype.UInt64,
+])
+
+
+def infer_schema(pandas_obj: Union[pd.DataFrame, pd.Series]):
+    """Infer schema for pandas DataFrame or Series object.
+
+    :param pandas_obj: DataFrame or Series object to infer.
+    :returns: DataFrameSchema or SeriesSchema
+    :raises: TypeError if pandas_obj is not expected type.
+    """
+    if isinstance(pandas_obj, pd.DataFrame):
+        return infer_dataframe_schema(pandas_obj)
+    elif isinstance(pandas_obj, pd.Series):
+        return infer_series_schema(pandas_obj)
+    else:
+        raise TypeError(
+            "pandas_obj type not recognized. Expected a pandas DataFrame or "
+            "Series, found %s" % type(pandas_obj)
+        )
+
+
+def infer_dataframe_schema(df: pd.DataFrame) -> DataFrameSchema:
+    """Infer a DataFrameSchema from a pandas DataFrame.
+
+    :param df: DataFrame object to infer.
+    :returns: DataFrameSchema
+    """
+    df_statistics = infer_dataframe_statistics(df)
+    schema = DataFrameSchema(
+        columns={
+            colname: Column(
+                properties["pandas_dtype"],
+                checks=_parse_check_statistics(properties["checks"]),
+                nullable=properties["nullable"],
+            )
+            for colname, properties in df_statistics["columns"].items()
+        },
+        coerce=True,
+    )
+    schema._is_inferred = True  # pylint: disable=protected-access
+    return schema
+
+
+def infer_series_schema(series) -> SeriesSchema:
+    """Infer a SeriesSchema from a pandas DataFrame.
+
+    :param series: Series object to infer.
+    :returns: SeriesSchema
+    """
+    series_statistics = infer_series_statistics(series)
+    schema = SeriesSchema(
+        pandas_dtype=series_statistics["pandas_dtype"],
+        checks=_parse_check_statistics(series_statistics["checks"]),
+        nullable=series_statistics["nullable"],
+        name=series_statistics["name"],
+        coerce=True,
+    )
+    schema._is_inferred = True  # pylint: disable=protected-access
+    return schema
+
+
+def infer_dataframe_statistics(df: pd.DataFrame) -> Dict[str, Any]:
+    """Infer column and index statistics from a pandas DataFrame."""
+    nullable_columns = df.isna().any()
+    inferred_column_dtypes = pd.Series({
+        col: _get_array_type(df[col]) for col in df
+    })
+    column_statistics = {
+        col: {
+            "pandas_dtype": dtype,
+            "nullable": nullable_columns[col],
+            "checks": _get_array_check_statistics(df[col], dtype),
+        }
+        for col, dtype in inferred_column_dtypes.iteritems()
+    }
+    return {
+        "columns": column_statistics,
+        "index": infer_index_statistics(df.index),
[…]
