+++ tests/test_schema_inference.py
@@ -1,0 +1,333 @@
+# pylint: disable=W0212
+"""Unit tests for schema inference module."""
+
+import pandas as pd
+import pytest
+
+import pandera as pa
+from pandera import schema_inference
+from pandera import dtypes, PandasDtype
+
+
+DEFAULT_INT = PandasDtype.from_str_alias(dtypes._DEFAULT_INT_TYPE)
+DEFAULT_FLOAT = PandasDtype.from_str_alias(dtypes._DEFAULT_FLOAT_TYPE)
+
+
+def _create_dataframe(multi_index=False, nullable=False):
+    if multi_index:
+        index = pd.MultiIndex.from_arrays(
+            [[1, 1, 2], ["a", "b", "c"]],
+            names=["int_index", "str_index"],
+        )
+    else:
+        index = pd.Index([10, 11, 12], name="int_index")
+
+    df = pd.DataFrame(
+        data={
+            "int": [1, 2, 3],
+            "float": [1., 2., 3.],
+            "boolean": [True, False, True],
+            "string": ["a", "b", "c"],
+            "datetime": pd.to_datetime(["20180101", "20180102", "20180103"]),
+        },
+        index=index,
+    )
+
+    if nullable:
+        df.iloc[0, :] = None
+
+    return df
+
+
+@pytest.mark.parametrize("pandas_obj, expectation", [
+    [pd.DataFrame({"col": [1, 2, 3]}), pa.DataFrameSchema],
+    [pd.Series([1, 2, 3]), pa.SeriesSchema],
+
+    # error cases
+    [int, TypeError],
+    [pd.Index([1, 2, 3]), TypeError],
+    ["foobar", TypeError],
+    [1, TypeError],
+    [[1, 2, 3], TypeError],
+    [{"key": "value"}, TypeError],
+])
+def test_infer_schema(pandas_obj, expectation):
+    """Test that convenience function correctly infers dataframe or series."""
+    if expectation is TypeError:
+        with pytest.raises(TypeError, match="^pandas_obj type not recognized"):
+            schema_inference.infer_schema(pandas_obj)
+    else:
+        assert isinstance(
+            schema_inference.infer_schema(pandas_obj), expectation
+        )
+
+
+@pytest.mark.parametrize("multi_index, nullable", [
+    [False, False],
+    [True, False],
+    [False, True],
+    [True, True],
+])
+def test_infer_dataframe_statistics(multi_index, nullable):
+    """Test dataframe statistics are correctly inferred."""
+    dataframe = _create_dataframe(multi_index, nullable)
+    statistics = schema_inference.infer_dataframe_statistics(dataframe)
+    stat_columns = statistics["columns"]
+
+    if nullable:
+        # bool and int dtypes are cast to float in the nullable case
+        assert stat_columns["int"]["pandas_dtype"] is DEFAULT_FLOAT
+        assert stat_columns["boolean"]["pandas_dtype"] is DEFAULT_FLOAT
+    else:
+        assert stat_columns["int"]["pandas_dtype"] is DEFAULT_INT
+        assert stat_columns["boolean"]["pandas_dtype"] is pa.Bool
+
+    assert stat_columns["float"]["pandas_dtype"] is DEFAULT_FLOAT
+    assert stat_columns["string"]["pandas_dtype"] is pa.String
+    assert stat_columns["datetime"]["pandas_dtype"] is pa.DateTime
+
+    if multi_index:
+        stat_indices = statistics["index"]
+        for stat_index, name, dtype in zip(
+                stat_indices,
+                ["int_index", "str_index"],
+                [DEFAULT_INT, pa.String]):
+            assert stat_index["name"] == name
+            assert stat_index["pandas_dtype"] is dtype
+            assert not stat_index["nullable"]
+    else:
+        stat_index = statistics["index"][0]
[…]

[…]
