### 2020-04-14 a4864e6
add default numeric dtype test
