+++ tests/test_dtypes.py
@@ -11,6 +11,7 @@ from pandera import (
     Column, DataFrameSchema, SeriesSchema, Check, DateTime, Float, Int,
     String, Bool, Category, Object, Timedelta
 )
+from pandera.dtypes import _DEFAULT_INT_TYPE, _DEFAULT_FLOAT_TYPE
 from pandera.errors import SchemaError
 
 
@@ -30,6 +31,17 @@ TESTABLE_DTYPES = [
     ("category", "category"),
     ("float64", "float64"),
 ]
+
+
+def test_default_numeric_dtypes():
+    """Test that default numeric dtypes int and float are consistent."""
+    assert str(pd.Series([1], dtype=int).dtype) == _DEFAULT_INT_TYPE
+    assert str(pd.Series([1], dtype="int").dtype) == _DEFAULT_INT_TYPE
+    assert pa.Int.str_alias == _DEFAULT_INT_TYPE
+
+    assert str(pd.Series([1], dtype=float).dtype) == _DEFAULT_FLOAT_TYPE
+    assert str(pd.Series([1], dtype="float").dtype) == _DEFAULT_FLOAT_TYPE
+    assert pa.Float.str_alias == _DEFAULT_FLOAT_TYPE
 
 
 def test_numeric_dtypes():
