### 2020-04-14 b640268
testing numeric dtypes
