### 2020-04-14 cb65d04
more testing of numeric dtypes
