+++ .travis.yml
@@ -11,8 +11,8 @@ before_cache:
   - touch $CONDA_DIR/conda-meta/history
 
 os:
-  - linux
-  - osx
+  # - linux
+  # - osx
   - windows
 
 env:
