### 2020-04-18 9df61d5
add DataFrameSchema.update_column and SeriesSchema.set_checks methods
