### 2020-04-18 31cf731
add __hash__ to check object
