+++ pandera/checks.py
@@ -1,8 +1,10 @@
 """Data validation checks."""
 
-from collections import namedtuple
+import inspect
 import operator
 import re
+from collections import namedtuple
+from functools import partial, wraps
 from typing import Dict, Union, Optional, List, Callable, Iterable
 
 import pandas as pd
@@ -28,13 +30,15 @@ class _CheckBase():
 
     def __init__(
             self,
-            fn: Callable,
+            check_fn: Callable,
             groups: Optional[Union[str, List[str]]] = None,
             groupby: Optional[Union[str, List[str], Callable]] = None,
             element_wise: bool = False,
+            name: str = None,
             error: Optional[str] = None,
             raise_warning: bool = False,
             n_failure_cases: Union[int, None] = constants.N_FAILURE_CASES,
+            **check_kwargs
     ) -> None:
         """Apply a validation function to each element, Series, or DataFrame.
 
@@ -66,16 +70,20 @@ class _CheckBase():
             The the case of ``Column`` checks, this function has access to the
             entire dataframe, but ``Column.name`` is selected from this
             DataFrameGroupby object so that a SeriesGroupBy object is passed
-            into ``fn``.
+            into ``check_fn``.
 
-            Specifying the groupby argument changes the ``fn`` signature to: ``
-            Callable[[Dict[Union[str, Tuple[str]], pd.Series]],
-            Union[bool, pd.Series]]``, where the input is a dictionary mapping
+            Specifying the groupby argument changes the ``check_fn`` signature
+            to:
+            
+            ``Callable[[Dict[Union[str, Tuple[str]], pd.Series]], Union[bool, pd.Series]]``  # noqa
+            
+            where the input is a dictionary mapping
             keys to subsets of the column/dataframe.
         :param element_wise: Whether or not to apply validator in an
             element-wise fashion. If bool, assumes that all checks should be
             applied to the column element-wise. If list, should be the same
             number of elements as checks.
+        :param name: optional name for the check.
         :param error: custom error message if series fails validation
             check.
         :param raise_warning: if True, raise a UserWarning and do not throw
@@ -84,6 +92,7 @@ class _CheckBase():
             check is informational and shouldn't stop execution of the program.
         :param n_failure_cases: report the top n failure cases. If None, then
             report all failure cases.
+        :param check_kwargs: key-word arguments to pass into ``check_fn``
 
         :example:
 
@@ -140,9 +149,11 @@ class _CheckBase():
         if element_wise and groupby is not None:
             raise errors.SchemaInitError(
                 "Cannot use groupby when element_wise=True.")
-        self.fn = fn
+        self._check_fn = check_fn
+        self._check_kwargs = check_kwargs
         self.element_wise = element_wise
         self.error = error
+        self.name = name
         self.raise_warning = raise_warning
         self.n_failure_cases = n_failure_cases
 
@@ -158,6 +169,10 @@ class _CheckBase():
             groups = [groups]
         self.groups = groups
         self.failure_cases = None
+
+    @property
+    def check_fn(self):
+        return partial(self._check_fn, **self._check_kwargs)
 
     def _format_groupby_input(
             self,
@@ -260,13 +275,13 @@ class _CheckBase():
 
         # apply check function to check object
         if self.element_wise:
-            check_result = check_obj.apply(self.fn, axis=1) if \
+            check_result = check_obj.apply(self.check_fn, axis=1) if \
                 isinstance(check_obj, pd.DataFrame) else \
-                check_obj.map(self.fn) if isinstance(check_obj, pd.Series) \
-                else self.fn(check_obj)
[…]
