+++ tests/test_schema_statistics.py
@@ -1,0 +1,415 @@
+"""Unit tests for inferring statistics of pandas objects."""
+
+import pandas as pd
+import pytest
+
+import pandera as pa
+from pandera import schema_statistics
+from pandera import dtypes, PandasDtype
+
+
+DEFAULT_INT = PandasDtype.from_str_alias(dtypes._DEFAULT_PANDAS_INT_TYPE)
+DEFAULT_FLOAT = PandasDtype.from_str_alias(dtypes._DEFAULT_PANDAS_FLOAT_TYPE)
+
+
+def _create_dataframe(multi_index=False, nullable=False):
+    if multi_index:
+        index = pd.MultiIndex.from_arrays(
+            [[1, 1, 2], ["a", "b", "c"]],
+            names=["int_index", "str_index"],
+        )
+    else:
+        index = pd.Index([10, 11, 12], name="int_index")
+
+    df = pd.DataFrame(
+        data={
+            "int": [1, 2, 3],
+            "float": [1., 2., 3.],
+            "boolean": [True, False, True],
+            "string": ["a", "b", "c"],
+            "datetime": pd.to_datetime(["20180101", "20180102", "20180103"]),
+        },
+        index=index,
+    )
+
+    if nullable:
+        df.iloc[0, :] = None
+
+    return df
+
+
+@pytest.mark.parametrize("multi_index, nullable", [
+    [False, False],
+    [True, False],
+    [False, True],
+    [True, True],
+])
+def test_infer_dataframe_statistics(multi_index, nullable):
+    """Test dataframe statistics are correctly inferred."""
+    dataframe = _create_dataframe(multi_index, nullable)
+    statistics = schema_statistics.infer_dataframe_statistics(dataframe)
+    stat_columns = statistics["columns"]
+
+    if nullable:
+        # bool and int dtypes are cast to float in the nullable case
+        assert stat_columns["int"]["pandas_dtype"] is DEFAULT_FLOAT
+        assert stat_columns["boolean"]["pandas_dtype"] is DEFAULT_FLOAT
+    else:
+        assert stat_columns["int"]["pandas_dtype"] is DEFAULT_INT
+        assert stat_columns["boolean"]["pandas_dtype"] is pa.Bool
+
+    assert stat_columns["float"]["pandas_dtype"] is DEFAULT_FLOAT
+    assert stat_columns["string"]["pandas_dtype"] is pa.String
+    assert stat_columns["datetime"]["pandas_dtype"] is pa.DateTime
+
+    if multi_index:
+        stat_indices = statistics["index"]
+        for stat_index, name, dtype in zip(
+                stat_indices,
+                ["int_index", "str_index"],
+                [DEFAULT_INT, pa.String]):
+            assert stat_index["name"] == name
+            assert stat_index["pandas_dtype"] is dtype
+            assert not stat_index["nullable"]
+    else:
+        stat_index = statistics["index"][0]
+        assert stat_index["name"] == "int_index"
+        assert stat_index["pandas_dtype"] is DEFAULT_INT
+        assert not stat_index["nullable"]
+
+    for properties in stat_columns.values():
+        if nullable:
+            assert properties["nullable"]
+        else:
+            assert not properties["nullable"]
+
+
+@pytest.mark.parametrize("check_stats, expectation", [
+    [{"min": 1}, [pa.Check.greater_than_or_equal_to(1)]],
+    [{"max": 10}, [pa.Check.less_than_or_equal_to(10)]],
+    [{"levels": ["a", "b", "c"]}, [pa.Check.isin(["a", "b", "c"])]],
+    [
+        {"min": 1, "max": 10},
+        [pa.Check.greater_than_or_equal_to(1),
[…]
