### 2020-04-20 2292b3b
fix pylint errors
