+++ pandera/schema_statistics.py
@@ -117,6 +117,7 @@ def parse_check_statistics(check_stats: Union[Dict[str, Any], None]):
 
 def get_dataframe_schema_statistics(dataframe_schema):
     """Get statistical properties from dataframe schema."""
+    # pylint: disable=protected-access
     statistics = {
         "columns": {
             col_name: {
@@ -135,6 +136,7 @@ def get_dataframe_schema_statistics(dataframe_schema):
 
 
 def _get_series_base_schema_statistics(series_schema_base):
+    # pylint: disable=protected-access
     return {
         "pandas_dtype": series_schema_base._pandas_dtype,
         "nullable": series_schema_base.nullable,
@@ -144,6 +146,7 @@ def _get_series_base_schema_statistics(series_schema_base):
 
 
 def get_index_schema_statistics(index_schema_component):
+    """Get statistical properties of index schema component."""
     try:
         # get index components from MultiIndex
         index_components = index_schema_component.indexes
