+++ pandera/io.py
@@ -99,8 +99,15 @@ def to_yaml(dataframe_schema, stream=None):
     """
     statistics = _serialize_schema(
         get_dataframe_schema_statistics(dataframe_schema))
+
+    def _write_yaml(obj, stream):
+        try:
+            return yaml.safe_dump(obj, stream=stream, sort_keys=False)
+        except TypeError:
+            return yaml.safe_dump(obj, stream=stream)
+
     try:
         with Path(stream).open("w") as f:
-            yaml.safe_dump(statistics, stream=f, sort_keys=False)
+            _write_yaml(statistics, f)
     except (TypeError, OSError):
-        return yaml.safe_dump(statistics, stream=stream, sort_keys=False)
+        return _write_yaml(statistics, stream)
