+++ pandera/checks.py
@@ -310,12 +310,14 @@ class _CheckBase():
 
         # apply check function to check object
         if self.element_wise:
+            # pylint: disable=not-callable
             check_result = check_obj.apply(self.check_fn, axis=1) if \
                 isinstance(check_obj, pd.DataFrame) else \
                 check_obj.map(self.check_fn) if \
                 isinstance(check_obj, pd.Series) else self.check_fn(check_obj)
         else:
             # vectorized check function case
+            # pylint: disable=not-callable
             check_result = self.check_fn(check_obj)
 
         # failure cases only apply when the check function returns a boolean
