### 2020-04-21 e15a611
add yaml schema parsing logic
