+++ tests/test_io.py
@@ -10,7 +10,10 @@ import pandera as pa
 from pandera import io
 
 
-PYYAML_VERSION = version.parse(yaml.__version__)
+try:
+    PYYAML_VERSION = version.parse(yaml.__version__)
+except AttributeError:
+    PYYAML_VERSION = None  # type: ignore
 
 
 def _create_schema():
@@ -54,7 +57,8 @@ index:
 
 
 @pytest.mark.skipif(
-    PYYAML_VERSION.release < (5, 1, 0),  # type: ignore
+    (PYYAML_VERSION is None or
+     PYYAML_VERSION.release < (5, 1, 0)),  # type: ignore
     reason="pyyaml >= 5.1.0 required",
 )
 def test_to_yaml():
@@ -68,7 +72,8 @@ def test_to_yaml():
 
 
 @pytest.mark.skipif(
-    PYYAML_VERSION.release < (5, 1, 0),  # type: ignore
+    (PYYAML_VERSION is None or
+     PYYAML_VERSION.release < (5, 1, 0)),  # type: ignore
     reason="pyyaml >= 5.1.0 required",
 )
 def test_from_yaml():
