+++ pandera/io.py
@@ -42,8 +42,11 @@ def _deserialize_schema(serialized_schema):
     # pylint: disable-all
     from pandera import DataFrameSchema, Column, Index, MultiIndex
 
-    if isinstance(serialized_schema, list) and len(serialized_schema) == 1:
-        serialized_schema = serialized_schema[0]
+    try:
+        serialized_schema["columns"]
+    except TypeError:
+        print(serialized_schema)
+        raise
 
     columns, index = None, None
     if serialized_schema["columns"] is not None:
