### 2020-04-21 0c52bbf
windows debugging: add file encoding
