+++ pandera/io.py
@@ -45,7 +45,9 @@ def _deserialize_schema(serialized_schema):
     try:
         serialized_schema["columns"]
     except TypeError:
-        print("SCHEMA IS A LIST")
+        # TODO: in windows, the file string name is being passed into this
+        # function: https://travis-ci.org/github/pandera-dev/pandera/jobs/677523664#L449
+        print("SCHEMA IS A STRING")
         print(serialized_schema)
         serialized_schema = serialized_schema[0]
 
@@ -91,7 +93,8 @@ def from_yaml(yaml_schema):
     try:
         with open(yaml_schema, "r", encoding='utf8') as f:
             serialized_schema = yaml.safe_load(f)
-    except OSError:
+    except OSError as e:
+        print("OS ERROR, %s" % e)
         serialized_schema = yaml.safe_load(yaml_schema)
     return _deserialize_schema(serialized_schema)
 
