+++ pandera/io.py
@@ -49,7 +49,6 @@ def _deserialize_schema(serialized_schema):
         # function: https://travis-ci.org/github/pandera-dev/pandera/jobs/677523664#L449
         print("ERROR")
         print(serialized_schema)
-        serialized_schema = serialized_schema[0]
 
     columns, index = None, None
     if serialized_schema["columns"] is not None:
@@ -93,8 +92,7 @@ def from_yaml(yaml_schema):
     try:
         with Path(yaml_schema).open("r") as f:
             serialized_schema = yaml.safe_load(f)
-    except OSError as e:
-        print(e)
+    except (TypeError, OSError):
         serialized_schema = yaml.safe_load(yaml_schema)
     return _deserialize_schema(serialized_schema)
 
