+++ pandera/schema_statistics.py
@@ -1,0 +1,220 @@
+"""Module for inferring the statistics of pandas objects."""
+
+import warnings
+from typing import Any, Dict, Union
+
+import pandas as pd
+
+from .checks import Check
+from .dtypes import PandasDtype
+
+
+NUMERIC_DTYPES = frozenset([
+    PandasDtype.Float,
+    PandasDtype.Float16,
+    PandasDtype.Float32,
+    PandasDtype.Float64,
+    PandasDtype.Int,
+    PandasDtype.Int8,
+    PandasDtype.Int16,
+    PandasDtype.Int32,
+    PandasDtype.Int64,
+    PandasDtype.UInt8,
+    PandasDtype.UInt16,
+    PandasDtype.UInt32,
+    PandasDtype.UInt64,
+    PandasDtype.DateTime,
+])
+
+# pylint: disable=unnecessary-lambda
+STATISTICS_TO_CHECKS = {
+    "min": lambda x: Check.greater_than_or_equal_to(x),
+    "max": lambda x: Check.less_than_or_equal_to(x),
+    "levels": lambda x: Check.isin(x),
+}
+
+CHECKS_TO_STATISTICS = {
+    Check.greater_than_or_equal_to.__name__: (
+        "min", lambda x: x.statistics["min_value"]),
+    Check.less_than_or_equal_to.__name__: (
+        "max", lambda x: x.statistics["max_value"]),
+    Check.isin.__name__: (
+        "levels", lambda x: x.statistics["allowed_values"]),
+}
+
+
+def infer_dataframe_statistics(df: pd.DataFrame) -> Dict[str, Any]:
+    """Infer column and index statistics from a pandas DataFrame."""
+    nullable_columns = df.isna().any()
+    inferred_column_dtypes = {
+        col: _get_array_type(df[col]) for col in df
+    }
+    column_statistics = {
+        col: {
+            "pandas_dtype": dtype,
+            "nullable": nullable_columns[col],
+            "checks": _get_array_check_statistics(df[col], dtype),
+        }
+        for col, dtype in inferred_column_dtypes.items()
+    }
+    return {
+        "columns": column_statistics if column_statistics else None,
+        "index": infer_index_statistics(df.index),
+    }
+
+
+def infer_series_statistics(series: pd.Series) -> Dict[str, Any]:
+    """Infer column and index statistics from a pandas Series."""
+    dtype = _get_array_type(series)
+    return {
+        "pandas_dtype": dtype,
+        "nullable": series.isna().any(),
+        "checks": _get_array_check_statistics(series, dtype),
+        "name": series.name,
+    }
+
+
+def infer_index_statistics(index: Union[pd.Index, pd.MultiIndex]):
+    """Infer index statistics given a pandas Index object."""
+
+    def _index_stats(index_level):
+        dtype = _get_array_type(index_level)
+        return {
+            "pandas_dtype": dtype,
+            "nullable": index_level.isna().any(),
+            "checks": _get_array_check_statistics(index_level, dtype),
+            "name": index_level.name,
+        }
+
+    if isinstance(index, pd.MultiIndex):
+        index_statistics = [
+            _index_stats(index.get_level_values(i))
+            for i in range(index.nlevels)
+        ]
+    elif isinstance(index, pd.Index):
+        index_statistics = [_index_stats(index)]
+    else:
+        warnings.warn(
+            "index type %s not recognized, skipping index inference" %
+            type(index),
+            UserWarning
+        )
+        index_statistics = []
+    return index_statistics if index_statistics else None
+
+
+def parse_check_statistics(check_stats: Union[Dict[str, Any], None]):
+    """Convert check statistics to a list of Check objects."""
+    if check_stats is None:
+        return None
[…]
