+++ pandera/schema_inference.py
@@ -1,38 +1,16 @@
 """Module for inferring dataframe/series schema."""
 
-import warnings
-from typing import Any, Dict, Union
+from typing import Union
 
 import pandas as pd
 
-from .checks import Check
-from .dtypes import PandasDtype
 from .schemas import DataFrameSchema, SeriesSchema
 from .schema_components import Column
-
-
-NUMERIC_DTYPES = frozenset([
-    PandasDtype.Float,
-    PandasDtype.Float16,
-    PandasDtype.Float32,
-    PandasDtype.Float64,
-    PandasDtype.Int,
-    PandasDtype.Int8,
-    PandasDtype.Int16,
-    PandasDtype.Int32,
-    PandasDtype.Int64,
-    PandasDtype.UInt8,
-    PandasDtype.UInt16,
-    PandasDtype.UInt32,
-    PandasDtype.UInt64,
-])
-
-# pylint: disable=unnecessary-lambda
-STATISTICS_TO_CHECKS = {
-    "min": lambda x: Check.greater_than_or_equal_to(x),
-    "max": lambda x: Check.less_than_or_equal_to(x),
-    "levels": lambda x: Check.isin(x),
-}
+from .schema_statistics import (
+    infer_dataframe_statistics,
+    infer_series_statistics,
+    parse_check_statistics,
+)
 
 
 def infer_schema(
@@ -66,7 +44,7 @@ def infer_dataframe_schema(df: pd.DataFrame) -> DataFrameSchema:
         columns={
             colname: Column(
                 properties["pandas_dtype"],
-                checks=_parse_check_statistics(properties["checks"]),
+                checks=parse_check_statistics(properties["checks"]),
                 nullable=properties["nullable"],
             )
             for colname, properties in df_statistics["columns"].items()
@@ -86,114 +64,10 @@ def infer_series_schema(series) -> SeriesSchema:
     series_statistics = infer_series_statistics(series)
     schema = SeriesSchema(
         pandas_dtype=series_statistics["pandas_dtype"],
-        checks=_parse_check_statistics(series_statistics["checks"]),
+        checks=parse_check_statistics(series_statistics["checks"]),
         nullable=series_statistics["nullable"],
         name=series_statistics["name"],
         coerce=True,
     )
     schema._is_inferred = True  # pylint: disable=protected-access
     return schema
-
-
-def infer_dataframe_statistics(df: pd.DataFrame) -> Dict[str, Any]:
-    """Infer column and index statistics from a pandas DataFrame."""
-    nullable_columns = df.isna().any()
-    inferred_column_dtypes = pd.Series({
-        col: _get_array_type(df[col]) for col in df
-    })
-    column_statistics = {
-        col: {
-            "pandas_dtype": dtype,
-            "nullable": nullable_columns[col],
-            "checks": _get_array_check_statistics(df[col], dtype),
-        }
-        for col, dtype in inferred_column_dtypes.iteritems()
-    }
-    return {
-        "columns": column_statistics,
-        "index": infer_index_statistics(df.index),
-    }
-
-
-def infer_series_statistics(series: pd.Series) -> Dict[str, Any]:
-    """Infer column and index statistics from a pandas Series."""
-    dtype = _get_array_type(series)
-    return {
-        "pandas_dtype": dtype,
-        "nullable": series.isna().any(),
-        "checks": _get_array_check_statistics(series, dtype),
-        "name": series.name,
-    }
-
-
-def infer_index_statistics(index: Union[pd.Index, pd.MultiIndex]):
-    """Infer index statistics given a pandas Index object."""
-
-    def _index_stats(index_level):
-        dtype = _get_array_type(index_level)
-        return {
-            "name": index_level.name,
-            "pandas_dtype": dtype,
-            "nullable": index_level.isna().any(),
[…]
