### 2020-05-04 6a753f9
implement lazy schema validation
