+++ pandera/schemas.py
@@ -9,9 +9,14 @@ from typing import Callable, List, Optional, Union, Dict, Any
 
 import pandas as pd
 
-from . import errors, constants, dtypes, error_formatters
-from .checks import Check, CheckResult
+from . import errors, constants, dtypes
+from .checks import Check
 from .dtypes import PandasDtype
+from .error_formatters import (
+    format_generic_error_message, format_vectorized_error_message,
+    reshape_failure_cases, scalar_failure_case,
+)
+from .error_handlers import SchemaErrorHandler
 from .hypotheses import Hypothesis
 
 
@@ -56,7 +61,8 @@ class DataFrameSchema():
             index=None,
             transformer: Callable = None,
             coerce: bool = False,
-            strict=False) -> None:
+            strict=False,
+            name: str = None) -> None:
         """Initialize DataFrameSchema validator.
 
         :param columns: a dict where keys are column names and values are
@@ -73,6 +79,7 @@ class DataFrameSchema():
             validation.
         :param strict: whether or not to accept columns in the dataframe that
             aren't in the DataFrameSchema.
+        :param name: name of the schema.
 
         :raises SchemaInitError: if impossible to build schema from parameters
 
@@ -125,6 +132,7 @@ class DataFrameSchema():
         self.index = index
         self.transformer = transformer
         self.strict = strict
+        self.name = name
         self._coerce = coerce
         self._validate_schema()
         self._set_column_names()
@@ -198,7 +206,7 @@ class DataFrameSchema():
         for check_index, check in enumerate(self.checks):
             check_results.append(
                 _handle_check_results(
-                    self, check_index, check, check(dataframe))
+                    self, check_index, check, dataframe)
             )
         return all(check_results)
 
@@ -242,11 +250,12 @@ class DataFrameSchema():
 
     def validate(
             self,
-            dataframe: pd.DataFrame,
+            check_obj: pd.DataFrame,
             head: Optional[int] = None,
             tail: Optional[int] = None,
             sample: Optional[int] = None,
-            random_state: Optional[int] = None
+            random_state: Optional[int] = None,
+            lazy: bool = False,
     ) -> pd.DataFrame:
         # pylint: disable=duplicate-code,too-many-locals
         """Check if all columns in a dataframe have a column in the Schema.
@@ -259,6 +268,9 @@ class DataFrameSchema():
         :param sample: validate a random sample of n rows. Rows overlapping
             with `head` or `tail` are de-duplicated.
         :param random_state: random seed for the ``sample`` argument.
+        :param lazy: if True, lazily evaluates dataframe against all validation
+            checks and raises a ``SchemaErrorReport``. Otherwise, raise
+            ``SchemaError`` as soon as one occurs.
         :returns: validated ``DataFrame``
 
         :raises SchemaError: when ``DataFrame`` violates built-in or custom
@@ -309,7 +321,7 @@ class DataFrameSchema():
                 UserWarning
             )
 
-        dataframe = dataframe.copy()
+        error_handler = SchemaErrorHandler(lazy)
 
         # dataframe strictness check makes sure all columns in the dataframe
         # are specified in the dataframe schema
@@ -321,7 +333,7 @@ class DataFrameSchema():
                 if col_schema.regex:
                     try:
                         col_regex_matches.extend(
-                            col_schema.get_regex_columns(dataframe.columns))
+                            col_schema.get_regex_columns(check_obj.columns))
                     except errors.SchemaError:
                         pass
 
@@ -330,56 +342,98 @@ class DataFrameSchema():
                 col_regex_matches
             )
 
-            for column in dataframe:
+            for column in check_obj:
                 if column not in expanded_column_names:
-                    raise errors.SchemaError(
+                    msg = (
[…]
