### 2020-05-08 f06dcc1
fix test error difference in py3.5
