### 2020-05-08 6be0076
fix mypy __init__ return None type
