### 2020-05-08 5e0627f
ignore lazy validation doctest if python < 3.6
