### 2020-05-08 bccd522
skip lazy validation if pyversion < 3.6
