### 2020-05-08 1c0ec72
feature/lazy-validation (#207)
* implement lazy schema validation

* fix sphinx warning and dataframe schema doc

* fix test error difference in py3.5

* fix mypy __init__ return None type

* add lazy validation to docs

* ignore lazy validation doctest if python < 3.6

* skip lazy validation if pyversion < 3.6
