### 2020-05-10 85910ee
implement simplified schema inference and io format
