### 2020-05-10 723b822
fix comparison column __eq__ operator
