### 2020-05-10 d64e5b8
remove unused statistics keys
