### 2020-05-12 a9505fc
implement to_script io function, update docs
