### 2020-05-12 6956706
use autopep8 to format script code python3.5 black for python3.6+
