### 2020-05-12 8f4f614
skip schema inference for python3.5
