### 2020-05-12 d4f2af0
implement simplified schema inference and io format (#209)
* implement simplified schema inference and io format

* fix comparison column __eq__ operator

* remove unused statistics keys

* implement to_script io function, update docs

* use autopep8 to format script code python3.5 black for python3.6+

* skip schema inference for python3.5
