+++ tests/test_schema_statistics.py
@@ -86,11 +86,23 @@ def test_infer_dataframe_statistics(multi_index, nullable):
 
 
 @pytest.mark.parametrize("check_stats, expectation", [
-    [{"min": 1}, [pa.Check.greater_than_or_equal_to(1)]],
-    [{"max": 10}, [pa.Check.less_than_or_equal_to(10)]],
-    [{"levels": ["a", "b", "c"]}, [pa.Check.isin(["a", "b", "c"])]],
     [
-        {"min": 1, "max": 10},
+        {"greater_than_or_equal_to": {"min_value": 1}},
+        [pa.Check.greater_than_or_equal_to(1)]
+    ],
+    [
+        {"less_than_or_equal_to": {"max_value": 10}},
+        [pa.Check.less_than_or_equal_to(10)]
+    ],
+    [
+        {"isin": {"allowed_values": ["a", "b", "c"]}},
+        [pa.Check.isin(["a", "b", "c"])]
+    ],
+    [
+        {
+            "greater_than_or_equal_to": {"min_value": 1},
+            "less_than_or_equal_to": {"max_value": 10},
+        },
         [pa.Check.greater_than_or_equal_to(1),
          pa.Check.less_than_or_equal_to(10)]
     ],
@@ -111,7 +123,10 @@ def test_parse_check_statistics(check_stats, expectation):
         [
             pd.Series([1, 2, 3], dtype=dtype.str_alias), {
                 "pandas_dtype": dtype, "nullable": False,
-                "checks": {"min": 1, "max": 3},
+                "checks": {
+                    "greater_than_or_equal_to": 1,
+                    "less_than_or_equal_to": 3
+                },
                 "name": None,
             }
         ]
@@ -122,7 +137,7 @@ def test_parse_check_statistics(check_stats, expectation):
     [
         pd.Series(["a", "b", "c", "a"], dtype="category"), {
             "pandas_dtype": pa.Category, "nullable": False,
-            "checks": {"levels": ["a", "b", "c"]},
+            "checks": {"isin": ["a", "b", "c"]},
             "name": None,
         }
     ],
@@ -136,8 +151,8 @@ def test_parse_check_statistics(check_stats, expectation):
         pd.Series(pd.to_datetime(["20180101", "20180102", "20180103"])), {
             "pandas_dtype": pa.DateTime, "nullable": False,
             "checks": {
-                "min": pd.Timestamp("20180101"),
-                "max": pd.Timestamp("20180103")
+                "greater_than_or_equal_to": pd.Timestamp("20180101"),
+                "less_than_or_equal_to": pd.Timestamp("20180103")
             },
             "name": None,
         }
@@ -168,7 +183,10 @@ INTEGER_TYPES = [
             0, pd.Series([1, 2, 3], dtype=dtype.value), {
                 # introducing nans to integer arrays upcasts to float
                 "pandas_dtype": DEFAULT_FLOAT, "nullable": True,
-                "checks": {"min": 2, "max": 3},
+                "checks": {
+                    "greater_than_or_equal_to": 2,
+                    "less_than_or_equal_to": 3
+                },
                 "name": None,
             }
         ]
@@ -178,14 +196,17 @@ INTEGER_TYPES = [
         # introducing nans to integer arrays upcasts to float
         0, pd.Series([True, False, True, False]), {
             "pandas_dtype": DEFAULT_FLOAT, "nullable": True,
-            "checks": {"min": 0, "max": 1},
+            "checks": {
+                "greater_than_or_equal_to": 0,
+                "less_than_or_equal_to": 1
+            },
             "name": None,
         }
     ],
     [
[…]
