+++ pandera/__init__.py
@@ -32,4 +32,4 @@ String = PandasDtype.String
 Timedelta = PandasDtype.Timedelta
 
 
-__version__ = "0.3.2"
+__version__ = "0.4.0"
