### 2020-05-18 923a197
add lazy kwarg to check_{input, output} decorators (#210)
