### 2020-05-25 ed3b416
fix lazy validation with nullable columns
