### 2020-05-25 f5051fa
fix lazy validation with nullable columns (#213)
