### 2020-05-30 424651b
bugfix: samples attr empty implies one sample ttest
