### 2020-05-30 d69b64c
add tests for hypothesis bugfix
