### 2020-05-30 0b0d6cc
bugfix: samples attr empty implies one sample ttest (#214)
* bugfix: samples attr empty implies one sample ttest

* add tests for hypothesis bugfix
