### 2020-05-31 28ac334
implement ignore_na in Check objects
