+++ tests/test_checks_builtin.py
@@ -11,9 +11,13 @@ from pandera.errors import SchemaError
 
 
 def check_values(values, check, expected_failure_cases):
-    """Creates a pd.Series from the given values and validates it with the check"""
+    """
+    Creates a pd.Series from the given values and validates it with the check
+    """
     series = pd.Series(values)
     check_result = check(series)
+    expected_check_obj = series.dropna() if check.ignore_na else series
+
     n_failure_cases = len(expected_failure_cases)
 
     # Assert that the check only fails if we expect it to
@@ -23,7 +27,8 @@ def check_values(values, check, expected_failure_cases):
     )
 
     # Assert that the returned check object is what was passed in
-    assert check_result.checked_object is series, "Wrong checked_object returned"
+    assert (check_result.checked_object == expected_check_obj).all(), \
+        "Wrong checked_object returned"
 
     # Assert that the failure cases are correct
     assert set(check_result.failure_cases) == set(expected_failure_cases), (
@@ -34,14 +39,16 @@ def check_values(values, check, expected_failure_cases):
 def check_none_failures(values, check):
     """Like check_values but expects a failure and due to Null values.
 
-    Asserts that the check fails on the given values and that the only failures are
-    Null values.
+    Asserts that the check fails on the given values and that the only
+    failures are Null values.
     """
     series = pd.Series(values)
     check_result = check(series)
     assert not check_result.check_passed, "Check should fail due to None value"
-    assert check_result.checked_object is series, "Wrong checked_object returned"
-    assert check_result.failure_cases.isnull().all(), "Only null values should be failure cases"
+    assert check_result.checked_object is series, \
+        "Wrong checked_object returned"
+    assert check_result.failure_cases.isnull().all(), \
+        "Only null values should be failure cases"
 
 
 def check_raise_error_or_warning(failure_values, check):
@@ -113,7 +120,9 @@ class TestGreaterThan:
     ])
     def test_failing_with_none(values, min_val):
         """Validate the check works also on dataframes with None values"""
-        check_none_failures(values, Check.greater_than(min_val))
+        check_none_failures(
+            values, Check.greater_than(min_val, ignore_na=False)
+        )
 
 
 class TestGreaterThanOrEqualTo:
@@ -152,7 +161,8 @@ class TestGreaterThanOrEqualTo:
     ])
     def test_failing(values, min_val, failure_cases):
         """Run checks which should fail"""
-        check_values(values, Check.greater_than_or_equal_to(min_val), failure_cases)
+        check_values(
+            values, Check.greater_than_or_equal_to(min_val), failure_cases)
         check_raise_error_or_warning(
             values, Check.greater_than_or_equal_to(min_val))
 
@@ -164,7 +174,9 @@ class TestGreaterThanOrEqualTo:
     ])
     def test_failing_with_none(values, min_val):
         """Validate the check works also on dataframes with None values"""
-        check_none_failures(values, Check.greater_than_or_equal_to(min_val))
+        check_none_failures(
+            values, Check.greater_than_or_equal_to(min_val, ignore_na=False)
+        )
 
 
 class TestLessThan:
@@ -214,7 +226,9 @@ class TestLessThan:
     ])
     def test_failing_with_none(values, max_value):
         """Validate the check works also on dataframes with None values"""
-        check_none_failures(values, Check.less_than(max_value))
+        check_none_failures(
+            values, Check.less_than(max_value, ignore_na=False)
+        )
 
 
 class TestLessThanOrEqualTo:
@@ -266,7 +280,9 @@ class TestLessThanOrEqualTo:
     ])
[…]

[…]
