### 2020-06-02 a45b0a3
handle null check column and groupby columns
