### 2020-06-04 180072e
implement ignore_na in Check objects (#215)
* implement ignore_na in Check objects

* handle null check column and groupby columns

* add check null value handling in docs
