+++ pandera/checks.py
@@ -60,6 +60,7 @@ class _CheckBase():
             check_fn: Callable,
             groups: Optional[Union[str, List[str]]] = None,
             groupby: Optional[Union[str, List[str], Callable]] = None,
+            ignore_na: bool = True,
             element_wise: bool = False,
             name: str = None,
             error: Optional[str] = None,
@@ -69,7 +70,7 @@ class _CheckBase():
     ) -> None:
         """Apply a validation function to each element, Series, or DataFrame.
 
-        :param fn: A function to check pandas data structure. For Column
+        :param check_fn: A function to check pandas data structure. For Column
             or SeriesSchema checks, if element_wise is True, this function
             should have the signature: ``Callable[[pd.Series],
             Union[pd.Series, bool]]``, where the output series is a boolean
@@ -106,6 +107,9 @@ class _CheckBase():
 
             where the input is a dictionary mapping
             keys to subsets of the column/dataframe.
+        :param ignore_na: If True, drops null values on the checked series or
+            dataframe before passing into the ``check_fn``. For dataframes,
+            drops rows with any null value.
         :param element_wise: Whether or not to apply validator in an
             element-wise fashion. If bool, assumes that all checks should be
             applied to the column element-wise. If list, should be the same
@@ -181,6 +185,7 @@ class _CheckBase():
         self.element_wise = element_wise
         self.error = error
         self.name = name
+        self.ignore_na = ignore_na
         self.raise_warning = raise_warning
         self.n_failure_cases = n_failure_cases
 
@@ -276,6 +281,35 @@ class _CheckBase():
         groupby_obj = dataframe.groupby(self.groupby)
         return self._format_groupby_input(groupby_obj, self.groups)
 
+    def _handle_na(
+            self,
+            df_or_series: Union[pd.DataFrame, pd.Series],
+            column: Optional[str] = None):
+        """Handle nan values before passing object to check function."""
+        if not self.ignore_na:
+            return df_or_series
+
+        drop_na_columns = []
+        if column is not None:
+            drop_na_columns.append(column)
+        if self.groupby is not None and isinstance(self.groupby, list):
+            # if groupby is specified as a list of columns, include them in
+            # the columns to consider when dropping records
+            for col in self.groupby:
+                # raise schema definition error if column is not in the
+                # validated dataframe
+                if isinstance(df_or_series, pd.DataFrame) and \
+                        col not in df_or_series:
+                    raise errors.SchemaDefinitionError(
+                        "`groupby` column '%s' not found" % col)
+            drop_na_columns.extend(self.groupby)
+
+        if drop_na_columns:
+            return df_or_series.loc[
+                df_or_series[drop_na_columns].dropna().index
+            ]
+        return df_or_series.dropna()
+
     def __call__(
             self,
             df_or_series: Union[pd.DataFrame, pd.Series],
@@ -289,12 +323,13 @@ class _CheckBase():
         :returns: CheckResult tuple containing checked object,
             check validation result, and failure cases from the checked object.
         """
+        df_or_series = self._handle_na(df_or_series, column)
+
+        column_dataframe_context = None
         if column is not None and isinstance(df_or_series, pd.DataFrame):
             column_dataframe_context = df_or_series.drop(
                 column, axis="columns")
             df_or_series = df_or_series[column].copy()
-        else:
-            column_dataframe_context = None
 
         # prepare check object
         if isinstance(df_or_series, pd.Series):
@@ -338,7 +373,6 @@ class _CheckBase():
             check_passed = check_result.all().all()
[…]
