### 2020-06-05 98486bc
bump to version 0.4.2
use version.py file for single source of truth
for package version.
