+++ pandera/version.py
@@ -1,0 +1,3 @@
+"""Version file"""
+
+__version__ = "0.4.2"
