### 2020-06-12 bb155e5
bugfix: lazy validation handles check returning scalar False value
