### 2020-06-14 37af9ed
add unit tests for all schema and schema components
