### 2020-06-14 3bf8e72
bugfix: lazy validation handles check_fn returning scalar False value, io can handle null index (#217)
* bugfix: lazy validation handles check returning scalar False value

* add unit tests for all schema and schema components
