### 2020-06-15 ac76748
Adds strict keyword to yaml; addresses #218
