### 2020-06-15 7e51056
A friendly helper for reminding to build the docs
