### 2020-06-15 3bd2655
Adds strict keyword to yaml; addresses #218 (#219)
* Adds strict keyword to yaml; addresses #218

* Still for #218; fix sphinxdoc inference

* A friendly helper for reminding to build the docs

Co-authored-by: Steve Taylor <stta9820@colorado.edu>
