+++ docs/source/schema_inference.rst
@@ -168,7 +168,7 @@ is a convenience method for this functionality.
    :skipif: SKIP
 
     schema_type: dataframe
-    version: 0.4.2
+    version: 0.4.3
     columns:
       column1:
         pandas_dtype: int64
