+++ pandera/io.py
@@ -156,8 +156,8 @@ def _deserialize_schema(serialized_schema):
             for index_component in serialized_schema["index"]
         ]
 
-    if isinstance(index, type(None)):
-        index = None
+    if index is None:
+        pass
     elif len(index) == 1:
         index = Index(**index[0])
     else:
