### 2020-06-17 7f9294d
From yaml null index (#225)
* from_yaml bug fix (index:null)

* whitespace

* fix trailing whitespace

* Update io.py

Co-authored-by: Niels Bantilan <niels.bantilan@gmail.com>
