### 2020-06-17 527fe0f
docsting "bug" fix (sphinx complaining)
