### 2020-06-18 618f3f1
use copy.deep_copy
