+++ pandera/schemas.py
@@ -6,7 +6,6 @@ import warnings
 from functools import wraps
 from pathlib import Path
 from typing import Callable, List, Optional, Union, Dict, Any
-from copy import deepcopy
 
 import pandas as pd
 
@@ -593,7 +592,7 @@ class DataFrameSchema():
         
         # We iterate over the existing columns dict and replace those keys
         # that exist in the rename_dict
-        new_schema = deepcopy(self)
+        new_schema = copy.deepcopy(self)
         new_columns = {
             (rename_dict[col_name] if col_name in rename_dict else col_name): col_attrs
             for col_name, col_attrs in self.columns.items()
