### 2020-06-18 7433dc9
Implemented rename_columns() (#226)
fixes #201 

* from_yaml bug fix (index:null)

* whitespace

* implemented rename_columns() in DataFrameSchema

* bug fix

* docsting "bug" fix (sphinx complaining)

* fix trailing whitespace

* use copy.deep_copy

Co-authored-by: Niels Bantilan <niels.bantilan@gmail.com>
