### 2020-06-20 98e0bf2
use raise from pattern
