### 2020-06-20 619d860
fix column coerce warning, decorator Schema error failure cases
