### 2020-06-20 6947d08
Add details in errors reporting to ease data debugging. (#193)
* More detailed error when coercing types

Report the name of the columns whose type is missing.

* WIP More detailed error for unrecognized type

* Update schemas.py

* use raise from pattern

* Update schemas.py

Co-authored-by: Anthony Truchet <anthony.truchet@c3iot.com>
Co-authored-by: Niels Bantilan <niels.bantilan@gmail.com>
