### 2020-06-20 c5e6516
fix column coerce warning, decorator Schema error failure cases (#228)
