### 2020-06-23 91d37a4
pin sphinx_autodoc_typehints in environment.yml
