+++ environment.yml
@@ -38,4 +38,4 @@ dependencies:
   - pyyaml
 
   - pip:
-    - sphinx_autodoc_typehints
+    - sphinx_autodoc_typehints==1.10.3
