### 2020-06-25 a77c7fb
move reshaping logic to the error formatter
