### 2020-06-25 b0bbb6d
fix dataframe handling logic
