### 2020-06-25 8269ca5
fix pylint disable on Check.__call__
