### 2020-06-25 5036676
enhancement: make failure case error reporting more intuitive (#232)
* enhancement: make failure case error reporting more intuitive

the way `SchemaError`s report failure cases is not very intuitive.
this diff simplifies this by skipping the aggregation step and
formatting the raw `reshaped_failure_cases` dataframe, which
it turns out is easier to read.

Add logic in `checks.py` that handles the case where the output
of `check_fn` is a boolean `DataFrame`. This allows for failure_cases
to be reported as tidy data at the `index` and `column` level.

* move reshaping logic to the error formatte
[…]
