+++ pandera/checks.py
@@ -315,6 +315,7 @@ class _CheckBase():
             df_or_series: Union[pd.DataFrame, pd.Series],
             column: Optional[str] = None,
     ) -> CheckResult:
+        # pylint: disable=too-many-branches
         """Validate pandas DataFrame or Series.
 
         :param df_or_series: pandas DataFrame of Series to validate.
@@ -363,14 +364,28 @@ class _CheckBase():
                 check_obj.shape[0] != check_result.shape[0] or \
                 (check_obj.index != check_result.index).all():
             failure_cases = None
-        else:
+        elif isinstance(check_result, pd.Series):
             failure_cases = check_obj[~check_result]
+        elif isinstance(check_result, pd.DataFrame):
+            # check results consisting of a boolean dataframe should be
+            # reported at the most granular level.
+            failure_cases = (
+                check_obj.unstack()[~check_result.unstack()]
+                .rename("failure_case")
+                .rename_axis(["column", "index"])
+                .reset_index()
+            )
+        else:
+            raise TypeError(
+                "output type of check_fn not recognized: %s" %
+                type(check_result)
+            )
 
         # handle check_result return types
         if isinstance(check_result, pd.Series):
             check_passed = check_result.all()
         elif isinstance(check_result, pd.DataFrame):
-            check_passed = check_result.all().all()
+            check_passed = check_result.all(axis=None)
         else:
             check_passed = check_result
         return CheckResult(check_passed, check_obj, failure_cases)
