### 2020-06-26 9a08e92
renamed decorator as per issue #233
