### 2020-06-26 8efc536
renamed decorator as per issue #233 (#235)
* renamed decorator as per issue #233

* changed decorator name at line 446
