### 2020-07-05 bfebc9b
Add issue templates
Adding issue templates for feature requests, bug reports, documentation improvements, and general questions
