+++ .github/ISSUE_TEMPLATE/documentation-improvement.md
@@ -1,0 +1,22 @@
+---
+name: Documentation Improvement
+about: Report wrong or missing documentation
+title: ''
+labels: docs
+assignees: ''
+
+---
+
+#### Location of the documentation
+
+[this should provide the location of the documentation, e.g. "pandera.schemas.DataFrameSchema" or the URL of the documentation, e.g. "https://pandera.readthedocs.io/en/stable/dataframe_schemas.html#column-validation"]
+
+**Note**: You can check the latest versions of the docs on `master` [here](https://pandera.readthedocs.io/en/latest/).
+
+#### Documentation problem
+
+[this should provide a description of what documentation you believe needs to be fixed/improved]
+
+#### Suggested fix for documentation
+
+[this should explain the suggested fix and **why** it's better than the existing documentation]
