+++ .github/ISSUE_TEMPLATE/bug_report.md
@@ -1,0 +1,39 @@
+---
+name: Bug report
+about: Create a report to help us improve
+title: ''
+labels: bug
+assignees: ''
+
+---
+
+**Describe the bug**
+A clear and concise description of what the bug is.
+
+- [ ] I have checked that this issue has not already been reported.
+- [ ] I have confirmed this bug exists on the latest version of pandas.
+- [ ] (optional) I have confirmed this bug exists on the master branch of pandas.
+
+**Note**: Please read [this guide](https://matthewrocklin.com/blog/work/2018/02/28/minimal-bug-reports) detailing how to provide the necessary information for us to reproduce your bug.
+
+#### Code Sample, a copy-pastable example
+
+```python
+# Your code here
+
+```
+
+#### Expected behavior
+A clear and concise description of what you expected to happen.
+
+#### Desktop (please complete the following information):
+
+ - OS: [e.g. iOS]
+ - Browser [e.g. chrome, safari]
+ - Version [e.g. 22]
+
+#### Screenshots
+If applicable, add screenshots to help explain your problem.
+
+#### Additional context
+Add any other context about the problem here.
