+++ .github/ISSUE_TEMPLATE/submit-question.md
@@ -1,0 +1,21 @@
+---
+name: Submit Question
+about: Ask a general question about pandera
+title: ''
+labels: question
+assignees: ''
+
+---
+
+- [ ] I have searched the [[pandas] tag](https://stackoverflow.com/questions/tagged/pandas) on StackOverflow for similar questions.
+- [ ] I have asked my usage related question on [StackOverflow](https://stackoverflow.com).
+
+#### Question about pandera
+
+**Note**: If you'd still like to submit a question, please read [this guide](
+https://matthewrocklin.com/blog/work/2018/02/28/minimal-bug-reports) detailing how to provide the necessary information for us to reproduce your question.
+
+```python
+# Your code here, if applicable
+
+```
