### 2020-07-06 9f48a60
Additional and optional from_yaml column properties; attempting to address pandera-dev#239
