### 2020-07-07 aee5926
Simplifying from_yaml additions, changed to_yaml, to_script as well. Including some tests; continued work towards pandera-dev#239
