+++ tests/test_schema_statistics.py
@@ -347,6 +347,10 @@ def test_get_dataframe_schema_statistics():
                     "less_than_or_equal_to": {"max_value": 100},
                 },
                 "nullable": True,
+                "allow_duplicates": True,
+                "coerce": False,
+                "required": True,
+                "regex": False,
             },
             "float": {
                 "pandas_dtype": pa.Float,
@@ -355,11 +359,19 @@ def test_get_dataframe_schema_statistics():
                     "less_than_or_equal_to": {"max_value": 100},
                 },
                 "nullable": False,
+                "allow_duplicates": True,
+                "coerce": False,
+                "required": True,
+                "regex": False,
             },
             "str": {
                 "pandas_dtype": pa.String,
                 "checks": {"isin": {"allowed_values": ["foo", "bar", "baz"]}},
                 "nullable": False,
+                "allow_duplicates": True,
+                "coerce": False,
+                "required": True,
+                "regex": False,
             },
         },
         "index": [
