+++ pandera/schema_statistics.py
@@ -110,6 +110,10 @@ def get_dataframe_schema_statistics(dataframe_schema):
             col_name: {
                 "pandas_dtype": column._pandas_dtype,
                 "nullable": column.nullable,
+                "allow_duplicates": column.allow_duplicates,
+                "coerce": column.coerce,
+                "required": column.required,
+                "regex": column.regex,
                 "checks": parse_checks(column.checks),
             }
             for col_name, column in dataframe_schema.columns.items()
