### 2020-07-07 a66da58
removing whitespace; for pandera-dev#239
