### 2020-07-07 822ed6c
Adding new props to test yaml string; pandera-dev#239
