### 2020-07-08 d6208e1
Additional and optional from_yaml column properties; addresses pandera-dev#239 (#240)
* Additional and optional from_yaml column properties; attempting to address pandera-dev#239

* Simplifying from_yaml additions, changed to_yaml, to_script as well. Including some tests; continued work towards pandera-dev#239

* removing whitespace; for pandera-dev#239

* Adding new props to test yaml string; pandera-dev#239

* testoutput updates; pandera-dev#239

Co-authored-by: dabu5788 <dave.bunten@colorado.edu>
