+++ pandera/schemas.py
@@ -588,26 +588,29 @@ class DataFrameSchema():
         """
         import pandera.io  # pylint: disable-all
         return pandera.io.to_yaml(self, fp)
-    
+
     def rename_columns(self, rename_dict: dict):
         """Rename columns using a dictionary of key value pairs 
-        
-        :param rename_dict: Dictionary of 'old_name':'new_name' key-value pairs.
+
+        :param rename_dict: Dictionary of 'old_name':'new_name' key-value
+            pairs.
         :returns: dataframe schema (copy of original)
         """
-        
+
         # We iterate over the existing columns dict and replace those keys
         # that exist in the rename_dict
         new_schema = copy.deepcopy(self)
         new_columns = {
-            (rename_dict[col_name] if col_name in rename_dict else col_name): col_attrs
+            (
+                rename_dict[col_name]
+                if col_name in rename_dict else col_name
+            ): col_attrs
             for col_name, col_attrs in self.columns.items()
         }
-        
+
         new_schema.columns = new_columns
- 
+
         return new_schema
-        
 
 
 class SeriesSchemaBase():
