### 2020-07-21 fc13735
skip pandas int tests if pandas < 1.0.0
