+++ pandera/schemas.py
@@ -592,8 +592,7 @@ class DataFrameSchema():
     def rename_columns(self, rename_dict: dict):
         """Rename columns using a dictionary of key value pairs 
 
-        :param rename_dict: Dictionary of 'old_name':'new_name' key-value
-            pairs.
+        :param rename_dict: Dictionary of 'old_name':'new_name' key-value pairs.
         :returns: dataframe schema (copy of original)
         """
 
@@ -601,16 +600,14 @@ class DataFrameSchema():
         # that exist in the rename_dict
         new_schema = copy.deepcopy(self)
         new_columns = {
-            (
-                rename_dict[col_name]
-                if col_name in rename_dict else col_name
-            ): col_attrs
+            (rename_dict[col_name] if col_name in rename_dict else col_name): col_attrs
             for col_name, col_attrs in self.columns.items()
         }
 
         new_schema.columns = new_columns
 
         return new_schema
+
 
 
 class SeriesSchemaBase():
