### 2020-07-22 1371786
support pandas nullable integer string aliases (#244)
* support pandas nullable integer string aliases

* skip pandas int tests if pandas < 1.0.0

* revert changes in schemas.py
