+++ pandera/__init__.py
@@ -28,6 +28,14 @@ UInt8 = PandasDtype.UInt8
 UInt16 = PandasDtype.UInt16
 UInt32 = PandasDtype.UInt32
 UInt64 = PandasDtype.UInt64
+INT8 = PandasDtype.INT8
+INT16 = PandasDtype.INT16
+INT32 = PandasDtype.INT32
+INT64 = PandasDtype.INT64
+UINT8 = PandasDtype.UINT8
+UINT16 = PandasDtype.UINT16
+UINT32 = PandasDtype.UINT32
+UINT64 = PandasDtype.UINT64
 Object = PandasDtype.Object
 String = PandasDtype.String
 Timedelta = PandasDtype.Timedelta
