### 2020-07-22 868f9f1
add viewcode sphinx extension to docs
