### 2020-07-22 8ab40c8
bugfix: io serialization can handle Index.name = None
before this fix, the way Index.name = None was serialized
would be in the string form `"None"`. Now, it should serialize
currently to `null` in yaml and `None` in scripts
