### 2020-07-23 486f381
bugfix: io serialization can handle Index.name = None (#248)
* bugfix: io serialization can handle Index.name = None

before this fix, the way Index.name = None was serialized
would be in the string form `"None"`. Now, it should serialize
currently to `null` in yaml and `None` in scripts

* fix error in schema_inference.rst docs
