### 2020-07-25 369333e
add `check_output` to the `CheckResult` namedtuple
the reason for this is to fulfill the use case where the
user wants to use `Check` objects by themselves and access
the boolean scalar, Series, or DataFrame that the `check_fn`
function outputs for downstream purposes like visualization
or boolean indexing
