### 2020-07-26 a4247b4
fix Check.__call__ docstring
