+++ .pre-commit-config.yaml
@@ -1,12 +1,12 @@
 repos:
   - repo: https://github.com/pycqa/pylint
-    rev: pylint-2.4.4
+    rev: pylint-2.5.0
     hooks:
     -   id: pylint
         args: ["--disable=import-error"]
         exclude: (^docs/|^scripts|setup.py)
   - repo: https://github.com/pre-commit/mirrors-mypy
-    rev: v0.761
+    rev: v0.782
     hooks:
     -   id: mypy
         entry: mypy pandera tests
