### 2020-08-05 6e5b555
fix error in Column.pandas_dtype (#256)
