### 2020-08-06 6a559f4
Fix check_input decorator when df passed in kwargs
When we call a function/method wrapped in @check_input like so:

    @check_input(in_schema)
    def decorated_function(df):
        return df

    decorated_function(df)  # passes
    decorated_function(df=df)  # fails

then the decorator fails as it doesn't expect the keyword argument. It's
counterintuitive for the final user of this API.

This patch assumes that the first keyword argument is the dataframe to
validate.
