+++ pandera/decorators.py
@@ -140,7 +140,7 @@ def check_input(
                 args_dict[obj_getter] = schema.validate(
                     args_dict[obj_getter], *validate_args)
                 args = list(args_dict.values())
-        elif obj_getter is None:
+        elif obj_getter is None and args:
             try:
                 args[0] = schema.validate(args[0], *validate_args)
             except errors.SchemaError as e:
@@ -150,6 +150,23 @@ def check_input(
                 )
                 raise errors.SchemaError(
                     schema, args[0], msg,
+                    failure_cases=e.failure_cases,
+                    check=e.check,
+                    check_index=e.check_index,
+                )
+        elif obj_getter is None and kwargs:
+            try:
+                # assume that the first keyword argument is the dataframe we want to validate
+                # since the dicts are ordered in python 3.6+
+                key = next(iter(kwargs.keys()))
+                kwargs[key] = schema.validate(kwargs[key], *validate_args)
+            except errors.SchemaError as e:
+                msg = (
+                    "error in check_input decorator of function '%s': %s" %
+                    (fn.__name__, e)
+                )
+                raise errors.SchemaError(
+                    schema, kwargs[key], msg,
                     failure_cases=e.failure_cases,
                     check=e.check,
                     check_index=e.check_index,
