### 2020-08-06 e2d3988
Fix check_input to accept only one kwd arg
