### 2020-08-07 2b84735
Parse check_input args in the order they defined
