### 2020-08-07 5f1c992
Fix len(seq) pylint warning
