### 2020-08-07 97c11da
Change SchemaError handling when using kwargs for df
