### 2020-08-07 ea81c74
Fix wrong argument for SchemaError
