### 2020-08-07 ec67d57
Test check_input decorator argument order
