### 2020-08-07 0812c57
Add reusable schema error handler
