### 2020-08-12 91ba6d9
first working class-based api
