+++ pandera/dtypes.py
@@ -1,7 +1,6 @@
 """Schema datatypes."""
 
 from enum import Enum
-from typing import Union
 
 import numpy as np
 import pandas as pd
@@ -63,8 +62,16 @@ class PandasDtype(Enum):
     2    c
     dtype: object
 
-    You can also directly use the string alias for each data-type in the
-    schema definition:
+    Alternatively, you can use built-in python scalar types for integers,
+    floats, booleans, and strings:
+
+    >>> pa.SeriesSchema(int).validate(pd.Series([1, 2, 3]))
+    0    1
+    1    2
+    2    3
+    dtype: int64
+
+    You can also use the pandas string aliases in the schema definition:
 
     >>> pa.SeriesSchema("int").validate(pd.Series([1, 2, 3]))
     0    1
@@ -177,7 +184,7 @@ class PandasDtype(Enum):
 
     @classmethod
     def from_pandas_api_type(
-            cls, pandas_api_type: str) -> Union["PandasDtype", None]:
+            cls, pandas_api_type: str) -> "PandasDtype":
         """Get PandasDtype enum from pandas api type.
 
         :param pandas_api_type: string output from
@@ -187,7 +194,7 @@ class PandasDtype(Enum):
         if pandas_api_type.startswith("mixed"):
             return cls.Object
 
-        return {
+        pandas_dtype = {
             "string": cls.String,
             "floating": cls.Float,
             "integer": cls.Int,
@@ -199,14 +206,23 @@ class PandasDtype(Enum):
             "timedelta": cls.Timedelta,
         }.get(pandas_api_type)
 
+        if pandas_dtype is None:
+            raise TypeError(
+                "pandas api type '%s' not recognized" %
+                pandas_api_type
+            )
+
+        return pandas_dtype
+
     @classmethod
     def from_python_type(cls, python_type: type) -> "PandasDtype":
         """Get PandasDtype enum from built-in python type.
 
         :param python_type: built-in python type. Allowable types are:
-            str, int, float, list, dict
+            str, int, float, and bool.
         """
         pandas_dtype = {
+            bool: cls.Bool,
             str: cls.String,
             int: cls.Int,
             float: cls.Float,
