### 2020-08-18 3c85cd1
add test coverage for pandas_api_type conversion
