### 2020-08-19 c7dcef4
built-in python scalar types are supported: int, float, str, bool (#263)
* support built-in int, float, str python types

* update documentation

* add test coverage for pandas_api_type conversion
