### 2020-08-31 da97a8c
add pandera.typing.Field
Used to provide extra information about a field of a SchemaModel
