+++ tests/test_typing.py
@@ -1,12 +1,13 @@
 """Tests schema creation and validation from type annotations."""
 
+from typing import Any, Optional
+
 import pandas as pd
 import pytest
 
 import pandera as pa
-from pandera import Column, DataFrameSchema, check_types
-from pandera.errors import SchemaError
-from pandera.typing import DataFrame, Index, SchemaModel, Series
+from pandera.errors import SchemaError, SchemaInitError
+from pandera.typing import DataFrame, Field, Index, SchemaModel, Series
 
 
 def test_schemamodel_to_dataframeschema():
@@ -17,8 +18,9 @@ def test_schemamodel_to_dataframeschema():
         b: Series["string"]
         idx: Index["string"]
 
-    expected = DataFrameSchema(
-        columns={"a": Column("int"), "b": Column("string")}, index=pa.Index("string")
+    expected = pa.DataFrameSchema(
+        columns={"a": pa.Column(pa.Int64), "b": pa.Column("string")},
+        index=pa.Index("string"),
     )
 
     assert expected == Schema.get_schema()
@@ -28,7 +30,7 @@ def test_check_types():
     class A(SchemaModel):
         a: Series["int"]
 
-    @check_types()
+    @pa.check_types()
     def transform(df: DataFrame[A]) -> DataFrame[A]:
         return df
 
@@ -49,24 +51,112 @@ def test_check_types_errors():
 
     df = pd.DataFrame({"a": [1]}, index=["1"])
 
-    @check_types()
+    @pa.check_types()
     def transform_index(df) -> DataFrame[A]:
         return df.reset_index(drop=True)
 
     with pytest.raises(SchemaError):
         transform_index(df)
 
-    @check_types()
+    @pa.check_types()
     def to_b(df: DataFrame[A]) -> DataFrame[B]:
         return df
 
     with pytest.raises(SchemaError, match="column 'b' not in dataframe"):
         to_b(df)
 
-    @check_types()
+    @pa.check_types()
     def to_str(df: DataFrame[A]) -> DataFrame[A]:
         df["a"] = "1"
         return df
 
     with pytest.raises(SchemaError, match="expected series 'a' to have type int64"):
         to_str(df)
+
+
+def test_optional_column():
+    class Schema(SchemaModel):
+        a: Optional[Series["string"]]
+        b: Optional[Series["string"]] = Field(eq="b")
+
+    schema = Schema.get_schema()
+    assert not schema.columns["a"].required
+    assert not schema.columns["b"].required
+
+
+def test_optional_index():
+    class Schema(SchemaModel):
+        idx: Optional[Index["string"]]
+
+    with pytest.raises(SchemaInitError, match="Index 'idx' cannot be Optional."):
+        Schema.get_schema()
+
+
+def test_schemamodel_with_fields():
+    """Tests that a SchemaModel.get_schema() can produce the correct schema."""
+
+    class Schema(SchemaModel):
+        a: Series["int"] = Field(eq=9, neq=0)
+        b: Series["string"]
+        idx: Index["string"] = Field(str_length={"min_value": 1})
+
+    actual = Schema.get_schema()
+    expected = pa.DataFrameSchema(
+        columns={
+            "a": pa.Column(
+                "int", checks=[pa.Check.equal_to(9), pa.Check.not_equal_to(0)]
+            ),
+            "b": pa.Column("string"),
+        },
+        index=pa.Index("string", pa.Check.str_length(1)),
+    )
+
+    assert actual == expected
+
+
+def test_field_to_column():
+    for flag in ["nullable", "allow_duplicates", "coerce", "regex"]:
+        for value in [True, False]:
[…]
