+++ pandera/dtypes.py
@@ -240,7 +240,9 @@ class PandasDtype(Enum):
         # see https://github.com/PyCQA/pylint/issues/2306
         if other is None:
             return False
-        elif self.value == "string":
+        if isinstance(other, str):
+            other = self.from_str_alias(other)
+        if self.value == "string":
             return self.value == other.value
         return self.str_alias == other.str_alias
 
