### 2020-08-31 f0c39ff
add support for MultiIndex to SchemaModel
