### 2020-08-31 bbde17a
allow check_types without empty parentheses
