+++ pandera/schema_model.py
@@ -1,7 +1,21 @@
 """Define typing extensions."""
 import inspect
+import re
 import warnings
-from typing import Any, Dict, Generic, Iterable, List, Optional, Type, TypeVar
+from collections import namedtuple
+from typing import (
+    Any,
+    Callable,
+    Dict,
+    Generic,
+    Iterable,
+    List,
+    Optional,
+    Set,
+    Type,
+    TypeVar,
+    Union,
+)
 
 import pandas as pd
 from typing_inspect import get_args, get_forward_arg, get_origin, is_optional_type
@@ -10,9 +24,14 @@ from . import dtypes, schema_components
 from .checks import Check
 from .dtypes import PandasDtype
 from .errors import SchemaInitError
+from .hypotheses import Hypothesis
 from .schemas import CheckList, DataFrameSchema, PandasDtypeInputTypes, SeriesSchemaBase
 
 Dtype = TypeVar("Dtype", PandasDtype, dtypes.PandasExtensionType, bool, int, str, float)
+SchemaIndex = Union[schema_components.Index, schema_components.MultiIndex]
+
+_ValidatorConfig = namedtuple("_ValidatorConfig", ["fields", "regex", "check"])
+_VALIDATOR_CONFIG_KEY = "__validator_config__"
 
 
 def get_first_arg(annotation: Type) -> type:
@@ -41,11 +60,36 @@ class Index(pd.Index, Generic[Dtype]):
     """Representation of pandas.Index."""
 
 
+def _filter_regexes(seq: Iterable, regexes: List[str]) -> Set[str]:
+    matched: Set[str] = set()
+    for regex in regexes:
+        pattern = re.compile(regex)
+        matched.update(filter(pattern.match, seq))
+    return matched
+
+
 class SchemaModel:
-    _schema: Optional[DataFrameSchema] = None
+    __schema: Optional[DataFrameSchema] = None
+    __checks: Dict[str, Set[Check]] = {}
 
     def __new__(cls, *args, **kwargs):
         raise TypeError(f"{cls.__name__} may not be instantiated")
+
+    @classmethod
+    def _get_attrs(cls):
+        def not_callable(x):
+            return (
+                not inspect.ismethod(x)
+                and not inspect.isfunction(x)
+                and not inspect.isbuiltin(x)
+            )
+
+        private_attrs = {"_SchemaModel__checks", "_SchemaModel__schema"}
+        return [
+            attr
+            for attr, _ in inspect.getmembers(cls, not_callable)
+            if not attr.startswith("__") and attr not in private_attrs
+        ]
 
     @classmethod
     def _check_missing_annotations(cls):
@@ -53,31 +97,46 @@ class SchemaModel:
         if not annotations:
             raise SchemaInitError(f"{cls.__name__} is not annotated.")
 
-        missing = []
-        for name, value in inspect.getmembers(cls):
-            if (
-                not name.startswith("_")
-                and not inspect.ismethod(value)
-                and name not in annotations.keys()
-            ):
-                missing.append(name)
-
+        missing = [attr for attr in cls._get_attrs() if attr not in annotations.keys()]
         if missing:
             warnings.warn(
                 f"The following unannotated attributes will be ignored: {missing}"
             )
 
     @classmethod
+    def _extract_validators(cls):
+        model_fields = list(cls.__annotations__.keys())
+        for fn_name, fn in inspect.getmembers(cls, inspect.isfunction):
+            check_definition = getattr(fn, _VALIDATOR_CONFIG_KEY, None)
+            if isinstance(check_definition, _ValidatorConfig):
+                if check_definition.regex:
+                    matched = _filter_regexes(model_fields, check_definition.fields)
+                else:
+                    matched = check_definition.fields
+
+                for field in matched:
[…]
