### 2020-09-08 88f2abe
Use Check.name in repr
