### 2020-09-11 c9fff4b
fix pylint errors (#266)
