### 2020-09-11 bf7ada3
Fixed broken link to pandas dtypes (#264)
