### 2020-09-11 44a2083
Use Check.name in repr (#265)
* Use Check.name in repr

* Fix extra paren
