### 2020-09-13 d22858f
add support for SchemaModel inheritance
