+++ pandera/schema_model.py
@@ -12,9 +12,11 @@ from typing import (
     List,
     Optional,
     Set,
+    Tuple,
     Type,
     TypeVar,
     Union,
+    cast,
 )
 
 import pandas as pd
@@ -29,9 +31,10 @@ from .schemas import CheckList, DataFrameSchema, PandasDtypeInputTypes, SeriesSc
 
 Dtype = TypeVar("Dtype", PandasDtype, dtypes.PandasExtensionType, bool, int, str, float)
 SchemaIndex = Union[schema_components.Index, schema_components.MultiIndex]
+CheckOrHypothesis = Union[Check, Hypothesis]
 
 _ValidatorConfig = namedtuple("_ValidatorConfig", ["fields", "regex", "check"])
-_VALIDATOR_CONFIG_KEY = "__validator_config__"
+_VALIDATOR_KEY = "__validator_config__"
 
 
 def get_first_arg(annotation: Type) -> type:
@@ -60,83 +63,103 @@ class Index(pd.Index, Generic[Dtype]):
     """Representation of pandas.Index."""
 
 
-def _filter_regexes(seq: Iterable, regexes: List[str]) -> Set[str]:
+def _regex_filter(seq: Iterable, regexps: List[str]) -> Set[str]:
+    """Filter items matching at least one of the regexes."""
     matched: Set[str] = set()
-    for regex in regexes:
+    for regex in regexps:
         pattern = re.compile(regex)
         matched.update(filter(pattern.match, seq))
     return matched
 
 
+def _get_field_annotations(cls: Type["SchemaModel"]) -> Dict[str, Any]:
+    annotations = cls.__annotations__
+    if not annotations:
+        raise SchemaInitError(f"{cls.__name__} is not annotated.")
+    missing = []
+    for attr_name, _ in inspect.getmembers(cls, lambda x: not inspect.isroutine(x)):
+        if attr_name.startswith("_"):  # ignore protected attributes
+            annotations.pop(attr_name, None)
+        elif attr_name not in annotations:
+            missing.append(attr_name)
+
+    if missing:
+        warnings.warn(
+            f"The following unannotated attributes will be ignored: {missing}"
+        )
+    return annotations
+
+
+def _update_checks(
+    checks: Dict[str, List[CheckOrHypothesis]],
+    fn: Callable,
+    fields: List[str],
+) -> Dict[str, List[CheckOrHypothesis]]:
+    """Extract validator from function and append it to the checks list."""
+    field_validator = getattr(fn, _VALIDATOR_KEY, None)
+    if isinstance(field_validator, _ValidatorConfig):
+        if field_validator.regex:
+            matched = _regex_filter(fields, field_validator.fields)
+        else:
+            matched = field_validator.fields
+        for field in matched:
+            if field not in fields:
+                raise SchemaInitError(
+                    f"Validator {fn.__name__} is assigned to a non-existing "
+                    + f"field '{field}'."
+                )
+            if field not in checks:
+                checks[field] = []
+            checks[field].append(field_validator.check)
+
+    return checks
+
+
+def _build_schema_index(
+    indexes: List[schema_components.Index],
+) -> Optional[SchemaIndex]:
+    index: Optional[SchemaIndex] = None
+    if indexes:
+        if len(indexes) == 1:
+            index = indexes[0]
+            index._name = None  # don't force name on single index
+        else:
+            index = schema_components.MultiIndex(indexes)
+    return index
+
+
 class SchemaModel:
-    __schema: Optional[DataFrameSchema] = None
-    __checks: Dict[str, Set[Check]] = {}
+    __schema__: Optional[DataFrameSchema] = None
 
     def __new__(cls, *args, **kwargs):
[…]
