### 2020-09-13 bf00106
add support for schema-wide model options
