+++ tests/test_schema_model.py
@@ -1,5 +1,4 @@
 """Tests schema creation and validation from type annotations."""
-
 from typing import Any, Optional
 
 import pandas as pd
@@ -16,7 +15,6 @@ from pandera.schema_model import (
     validator,
     dataframe_validator,
     dataframe_transformer,
-    BaseConfig,
 )
 
 
@@ -24,45 +22,42 @@ def test_schemamodel_to_dataframeschema():
     """Tests that a SchemaModel.to_schema() can produce the correct schema."""
 
     class Schema(SchemaModel):
-        a: Series["int"]
-        b: Series["string"]
-        idx: Index["string"]
+        a: Series[int]
+        b: Series[str]
+        idx: Index[str]
 
     expected = pa.DataFrameSchema(
-        columns={"a": pa.Column(pa.Int64), "b": pa.Column("string")},
-        index=pa.Index("string"),
+        columns={"a": pa.Column(int), "b": pa.Column(str)},
+        index=pa.Index(str),
     )
 
     assert expected == Schema.to_schema()
 
+# required to be globals: see https://pydantic-docs.helpmanual.io/usage/postponed_annotations/
+class A(SchemaModel):
+    a: Series[int]
+    idx: Index[str]
+
+
+class B(SchemaModel):
+    b: Series[int]
+
 
 def test_check_types():
-    class A(SchemaModel):
-        a: Series["int"]
-
     @pa.check_types
     def transform(df: DataFrame[A]) -> DataFrame[A]:
         return df
 
-    df = pd.DataFrame({"a": [1]})
-    try:
-        transform(df)
-    except Exception as e:
-        pytest.fail(f"Unexpected Exception {e}")
+    df = pd.DataFrame({"a": [1]},  index=["1"])
+    pd.testing.assert_frame_equal(transform(df), df)
 
 
 def test_check_types_errors():
-    class A(SchemaModel):
-        a: Series["int"]
-        idx: Index["string"]
-
-    class B(SchemaModel):
-        b: Series["int"]
 
     df = pd.DataFrame({"a": [1]}, index=["1"])
 
     @pa.check_types
-    def transform_index(df) -> DataFrame[A]:
+    def transform_index(df: DataFrame[A]) -> DataFrame[A]:
         return df.reset_index(drop=True)
 
     with pytest.raises(SchemaError):
@@ -86,8 +81,8 @@ def test_check_types_errors():
 
 def test_optional_column():
     class Schema(SchemaModel):
-        a: Optional[Series["string"]]
-        b: Optional[Series["string"]] = Field(eq="b")
+        a: Optional[Series[str]]
+        b: Optional[Series[str]] = Field(eq="b")
 
     schema = Schema.to_schema()
     assert not schema.columns["a"].required
@@ -96,7 +91,7 @@ def test_optional_column():
 
 def test_optional_index():
     class Schema(SchemaModel):
-        idx: Optional[Index["string"]]
+        idx: Optional[Index[str]]
 
     with pytest.raises(SchemaInitError, match="Index 'idx' cannot be Optional."):
         Schema.to_schema()
@@ -106,19 +101,19 @@ def test_schemamodel_with_fields():
     """Tests that a SchemaModel.to_schema() can produce the correct schema."""
 
     class Schema(SchemaModel):
-        a: Series["int"] = Field(eq=9, neq=0)
-        b: Series["string"]
-        idx: Index["string"] = Field(str_length={"min_value": 1})
+        a: Series[int] = Field(eq=9, neq=0)
+        b: Series[str]
+        idx: Index[str] = Field(str_length={"min_value": 1})
 
     actual = Schema.to_schema()
     expected = pa.DataFrameSchema(
         columns={
             "a": pa.Column(
[…]
