### 2020-09-14 c4d556c
clean up class-api code
