+++ tests/test_schema_model.py
@@ -5,23 +5,13 @@ import pandas as pd
 import pytest
 
 import pandera as pa
-from pandera.errors import SchemaError, SchemaErrors, SchemaInitError
-from pandera.schema_model import (
-    DataFrame,
-    Field,
-    Index,
-    SchemaModel,
-    Series,
-    validator,
-    dataframe_validator,
-    dataframe_transformer,
-)
+from pandera.typing import Index, Series
 
 
-def test_schemamodel_to_dataframeschema():
+def test_simple_to_schema():
     """Tests that a SchemaModel.to_schema() can produce the correct schema."""
 
-    class Schema(SchemaModel):
+    class Schema(pa.SchemaModel):
         a: Series[int]
         b: Series[str]
         idx: Index[str]
@@ -33,56 +23,29 @@ def test_schemamodel_to_dataframeschema():
 
     assert expected == Schema.to_schema()
 
-# required to be globals: see https://pydantic-docs.helpmanual.io/usage/postponed_annotations/
-class A(SchemaModel):
-    a: Series[int]
-    idx: Index[str]
 
+def test_invalid_annotations():
+    """Tests that a SchemaModel.to_schema() can produce the correct schema."""
 
-class B(SchemaModel):
-    b: Series[int]
+    class IntSchema(pa.SchemaModel):
+        a: int
 
+    with pytest.raises(pa.errors.SchemaInitError, match="Invalid annotation"):
+        IntSchema.to_schema()
 
-def test_check_types():
-    @pa.check_types
-    def transform(df: DataFrame[A]) -> DataFrame[A]:
-        return df
+    from decimal import Decimal
 
-    df = pd.DataFrame({"a": [1]},  index=["1"])
-    pd.testing.assert_frame_equal(transform(df), df)
+    class InvalidDtypeSchema(pa.SchemaModel):
+        d: Series[Decimal]
 
-
-def test_check_types_errors():
-
-    df = pd.DataFrame({"a": [1]}, index=["1"])
-
-    @pa.check_types
-    def transform_index(df: DataFrame[A]) -> DataFrame[A]:
-        return df.reset_index(drop=True)
-
-    with pytest.raises(SchemaError):
-        transform_index(df)
-
-    @pa.check_types
-    def to_b(df: DataFrame[A]) -> DataFrame[B]:
-        return df
-
-    with pytest.raises(SchemaError, match="column 'b' not in dataframe"):
-        to_b(df)
-
-    @pa.check_types
-    def to_str(df: DataFrame[A]) -> DataFrame[A]:
-        df["a"] = "1"
-        return df
-
-    with pytest.raises(SchemaError, match="expected series 'a' to have type int64"):
-        to_str(df)
+    with pytest.raises(TypeError, match="python type '<class 'decimal.Decimal'>"):
+        InvalidDtypeSchema.to_schema()
 
 
 def test_optional_column():
-    class Schema(SchemaModel):
+    class Schema(pa.SchemaModel):
         a: Optional[Series[str]]
-        b: Optional[Series[str]] = Field(eq="b")
+        b: Optional[Series[str]] = pa.Field(eq="b")
 
     schema = Schema.to_schema()
     assert not schema.columns["a"].required
@@ -90,20 +53,22 @@ def test_optional_column():
 
 
 def test_optional_index():
-    class Schema(SchemaModel):
+    class Schema(pa.SchemaModel):
         idx: Optional[Index[str]]
 
-    with pytest.raises(SchemaInitError, match="Index 'idx' cannot be Optional."):
+    with pytest.raises(
+        pa.errors.SchemaInitError, match="Index 'idx' cannot be Optional."
+    ):
         Schema.to_schema()
 
 
 def test_schemamodel_with_fields():
[…]
