+++ pandera/schema_model.py
@@ -7,7 +7,6 @@ from typing import (
     Any,
     Callable,
     Dict,
-    Generic,
     Iterable,
     List,
     Optional,
@@ -21,288 +20,33 @@ from typing import (
 )
 
 import pandas as pd
-import typing_inspect
 
-from . import dtypes, schema_components
+from . import schema_components
 from .checks import Check
-from .dtypes import PandasDtype
 from .errors import SchemaInitError
 from .hypotheses import Hypothesis
-from .schemas import CheckList, DataFrameSchema, PandasDtypeInputTypes, SeriesSchemaBase
+from .schema_components import SeriesSchemaBase
+from .schemas import CheckList, DataFrameSchema, PandasDtypeInputTypes
+from .typing import Index, Series, parse_annotation
 
-Dtype = TypeVar("Dtype", PandasDtype, dtypes.PandasExtensionType, bool, int, str, float)
+__all__ = [
+    "Field",
+    "SchemaModel",
+    "BaseConfig",
+    "dataframe_validator",
+    "dataframe_transformer",
+    "dataframe_validator",
+]
+
 SchemaIndex = Union[schema_components.Index, schema_components.MultiIndex]
 CheckOrHypothesis = Union[Check, Hypothesis]
+SchemaComponent = TypeVar("SchemaComponent", bound=SeriesSchemaBase)
 
 _ValidatorConfig = namedtuple("_ValidatorConfig", ["fields", "regex", "check"])
 _VALIDATOR_KEY = "__validator_config__"
 _DATAFRAME_VALIDATOR_KEY = "__dataframe_validator_config__"
 _TRANSFORMER_KEY = "__dataframe_transformer_config__"
 _CONFIG_KEY = "Config"
-
-
-def get_first_arg(annotation: Type) -> type:
-    """Get first argument of subscripted type tp
-
-    :example:
-
-    >>> import numpy as np
-    >>> from pandera.schema_model import Series, get_first_arg
-    >>>
-    >>> assert get_first_arg(Series[np.int32]) == np.int32
-    >>> assert get_first_arg(Series["np.int32"]) == "np.int32"
-    """
-    arg = typing_inspect.get_args(annotation)[0]
-    # e.g get_args(Series["int32"])[0] gives ForwardRef('int32')
-    fwd = typing_inspect.get_forward_arg(arg)
-
-    return fwd if fwd is not None else arg
-
-
-class Series(pd.Series, Generic[Dtype]):
-    """Representation of pandas.Series."""
-
-
-class Index(pd.Index, Generic[Dtype]):
-    """Representation of pandas.Index."""
-
-
-class BaseConfig:
-    name: Optional[str] = None
-    coerce: bool = False
-    strict: bool = False
-    multiindex_coerce: bool = False
-    multiindex_strict: bool = False
-    multiindex_name: Optional[str] = None
-
-
-_config_options = [attr for attr in vars(BaseConfig) if not attr.startswith("_")]
-
-
-def _extract_config_options(config: Type) -> Dict[str, Any]:
-    return {
-        name: value for name, value in vars(config).items() if name in _config_options
-    }
-
-
-def _regex_filter(seq: Iterable, regexps: List[str]) -> Set[str]:
-    """Filter items matching at least one of the regexes."""
-    matched: Set[str] = set()
-    for regex in regexps:
-        pattern = re.compile(regex)
-        matched.update(filter(pattern.match, seq))
-    return matched
-
-
-def _get_field_annotations(cls: Type["SchemaModel"]) -> Dict[str, Any]:
-    annotations = get_type_hints(cls)
-    if not annotations:
-        raise SchemaInitError(f"{cls.__name__} is not annotated.")
-    missing = []
[…]

[…]
