### 2020-09-17 2527bc0
remove transformer from class-api
