### 2020-09-17 1351f1c
fix test_schema_model error regexps
