+++ pandera/__init__.py
@@ -9,7 +9,6 @@ from .schemas import DataFrameSchema, SeriesSchema
 from .schema_components import Column, Index, MultiIndex
 from .schema_inference import infer_schema
 from .schema_model import Field, SchemaModel, check, dateframe_check
-from .schemas import DataFrameSchema, SeriesSchema
 from .version import __version__
 
 # pylint: disable=invalid-name
