+++ tests/test_schema_model.py
@@ -1,4 +1,6 @@
 """Tests schema creation and validation from type annotations."""
+# pylint:disable=R0903,C0115,C0116
+
 from typing import Any, Optional
 
 import pandas as pd
@@ -9,7 +11,7 @@ from pandera.typing import Index, Series
 
 
 def test_simple_to_schema():
-    """Tests that a SchemaModel.to_schema() can produce the correct schema."""
+    """Test that SchemaModel.to_schema() can produce the correct schema."""
 
     class Schema(pa.SchemaModel):
         a: Series[int]
@@ -25,7 +27,9 @@ def test_simple_to_schema():
 
 
 def test_invalid_annotations():
-    """Tests that a SchemaModel.to_schema() can produce the correct schema."""
+    """Test that SchemaModel.to_schema() fails if annotations or types are not
+    recognized.
+    """
 
     class IntSchema(pa.SchemaModel):
         a: int
@@ -33,16 +37,18 @@ def test_invalid_annotations():
     with pytest.raises(pa.errors.SchemaInitError, match="Invalid annotation"):
         IntSchema.to_schema()
 
-    from decimal import Decimal
+    from decimal import Decimal # pylint:disable=C0415
 
     class InvalidDtypeSchema(pa.SchemaModel):
-        d: Series[Decimal]
+        d: Series[Decimal] # type: ignore
 
     with pytest.raises(TypeError, match="python type '<class 'decimal.Decimal'>"):
         InvalidDtypeSchema.to_schema()
 
 
 def test_optional_column():
+    """Test that optional columns are not required."""
+
     class Schema(pa.SchemaModel):
         a: Optional[Series[str]]
         b: Optional[Series[str]] = pa.Field(eq="b")
@@ -53,6 +59,8 @@ def test_optional_column():
 
 
 def test_optional_index():
+    """Test that optional indices are not required."""
+
     class Schema(pa.SchemaModel):
         idx: Optional[Index[str]]
 
@@ -63,7 +71,7 @@ def test_optional_index():
 
 
 def test_schemamodel_with_fields():
-    """Tests that a SchemaModel.to_schema() can produce the correct schema."""
+    """Test that Fields are translated in the schema."""
 
     class Schema(pa.SchemaModel):
         a: Series[int] = pa.Field(eq=9, neq=0)
@@ -85,6 +93,7 @@ def test_schemamodel_with_fields():
 
 
 def test_field_to_column():
+    """Test that Field outputs the correct column options."""
     for flag in ["nullable", "allow_duplicates", "coerce", "regex"]:
         for value in [True, False]:
             col = pa.Field(**{flag: value}).to_column(pa.DateTime, required=value)
@@ -95,6 +104,7 @@ def test_field_to_column():
 
 
 def test_field_to_index():
+    """Test that Field outputs the correct index options."""
     for flag in ["nullable", "allow_duplicates"]:
         for value in [True, False]:
             index = pa.Field(**{flag: value}).to_index(pa.DateTime)
@@ -104,6 +114,7 @@ def test_field_to_index():
 
 
 def test_field_no_checks():
+    """Test Field without checks."""
     assert not pa.Field().to_column(str).checks
 
 
@@ -127,12 +138,15 @@ def test_field_no_checks():
     ],
 )
 def test_field_checks(arg: str, value: Any, expected: pa.Check):
+    """Test that all built-in checks are available in a Field."""
     checks = pa.Field(**{arg: value}).to_column(str).checks
     assert len(checks) == 1
     assert checks[0] == expected
 
 
 def test_multiindex():
+    """Test that multiple Index annotations create a MultiIndex."""
+
     class Schema(pa.SchemaModel):
         a: Index[int]
[…]
