### 2020-09-18 257716d
implement index in SeriesSchema
