### 2020-09-19 841da42
test coercion logic
