### 2020-09-19 b89eb40
feature/199: SeriesSchema supports index (#268)
* implement index in SeriesSchema

* add tests

* test coercion logic
