### 2020-09-19 4e0b43c
add compariison operator aliases to built-in checks
