+++ pandera/checks.py
@@ -433,6 +433,54 @@ class Check(_CheckBase):
     """Check a pandas Series or DataFrame for certain properties."""
 
     @classmethod
+    @register_check_statistics(["value"])
+    def equal_to(cls, value, **kwargs) -> 'Check':
+        """Ensure all elements of a series equal a certain value.
+
+        :param value: All elements of a given :class:`pandas.Series` must have
+            this value
+        :param kwargs: key-word arguments passed into the `Check` initializer.
+
+        :returns: :class:`Check` object
+        """
+        def _equal(series: pd.Series) -> pd.Series:
+            """Comparison function for check"""
+            return series == value
+
+        return cls(
+            _equal,
+            name=cls.equal_to.__name__,
+            error="equal_to(%s)" % value,
+            **kwargs,
+        )
+
+    eq = equal_to
+
+    @classmethod
+    @register_check_statistics(["value"])
+    def not_equal_to(cls, value, **kwargs) -> 'Check':
+        """Ensure no elements of a series equals a certain value.
+
+        :param value: This value must not occur in the checked
+            :class:`pandas.Series`.
+        :param kwargs: key-word arguments passed into the `Check` initializer.
+
+        :returns: :class:`Check` object
+        """
+        def _not_equal(series: pd.Series) -> pd.Series:
+            """Comparison function for check"""
+            return series != value
+
+        return cls(
+            _not_equal,
+            name=cls.not_equal_to.__name__,
+            error="not_equal_to(%s)" % value,
+            **kwargs,
+        )
+
+    ne = not_equal_to
+
+    @classmethod
     @register_check_statistics(["min_value"])
     def greater_than(cls, min_value, **kwargs) -> 'Check':
         """Ensure values of a series are strictly greater than a minimum value.
@@ -457,6 +505,8 @@ class Check(_CheckBase):
             error="greater_than(%s)" % min_value,
             **kwargs,
         )
+
+    gt = greater_than
 
     @classmethod
     @register_check_statistics(["min_value"])
@@ -484,6 +534,8 @@ class Check(_CheckBase):
             **kwargs,
         )
 
+    ge = greater_than_or_equal_to
+
     @classmethod
     @register_check_statistics(["max_value"])
     def less_than(cls, max_value, **kwargs) -> 'Check':
@@ -510,10 +562,12 @@ class Check(_CheckBase):
             **kwargs,
         )
 
+    lt = less_than
+
     @classmethod
     @register_check_statistics(["max_value"])
     def less_than_or_equal_to(cls, max_value, **kwargs) -> 'Check':
-        """Ensure no value of a series exceeds a certain value.
+        """Ensure values are less than or equal to a maximum value.
 
         :param max_value: Upper bound not to be exceeded. Must be a type
             comparable to the dtype of the :class:`pandas.Series` to be
@@ -535,6 +589,8 @@ class Check(_CheckBase):
             error="less_than_or_equal_to(%s)" % max_value,
             **kwargs
         )
+
+    le = less_than_or_equal_to
 
     @classmethod
     @register_check_statistics([
@@ -582,50 +638,6 @@ class Check(_CheckBase):
             _in_range,
             name=cls.in_range.__name__,
             error="in_range(%s, %s)" % (min_value, max_value),
-            **kwargs,
-        )
-
-    @classmethod
-    @register_check_statistics(["value"])
-    def equal_to(cls, value, **kwargs) -> 'Check':
-        """Ensure all elements of a series equal a certain value.
-
-        :param value: All elements of a given :class:`pandas.Series` must have
-            this value
-        :param kwargs: key-word arguments passed into the `Check` initializer.
-
-        :returns: :class:`Check` object
-        """
-        def _equal(series: pd.Series) -> pd.Series:
[…]
