### 2020-09-20 472e7f4
force model checks as classmethod + clean up code organization
