+++ pandera/schema_model.py
@@ -2,10 +2,8 @@
 import inspect
 import re
 import warnings
-from collections import namedtuple
 from typing import (
     Any,
-    Callable,
     Dict,
     Iterable,
     List,
@@ -13,7 +11,6 @@ from typing import (
     Set,
     Tuple,
     Type,
-    TypeVar,
     Union,
     cast,
     get_type_hints,
@@ -22,176 +19,23 @@ from typing import (
 from . import schema_components
 from .checks import Check
 from .errors import SchemaInitError
-from .hypotheses import Hypothesis
-from .schema_components import SeriesSchemaBase
-from .schemas import CheckList, DataFrameSchema, PandasDtypeInputTypes
+from .model_components import (
+    CHECK_KEY,
+    DATAFRAME_CHECK_KEY,
+    CheckInfo,
+    FieldCheckInfo,
+    FieldInfo,
+)
+from .schemas import DataFrameSchema
 from .typing import Index, Series, parse_annotation
 
-__all__ = [
-    "Field",
-    "SchemaModel",
-    "BaseConfig",
-    "dateframe_check",
-    "dateframe_check",
-]
-
 SchemaIndex = Union[schema_components.Index, schema_components.MultiIndex]
-CheckOrHypothesis = Union[Check, Hypothesis]
-SchemaComponent = TypeVar("SchemaComponent", bound=SeriesSchemaBase)
 
-_ValidatorConfig = namedtuple("_ValidatorConfig", ["fields", "regex", "check"])
-_CHECK_KEY = "__check_config__"
-_DATAFRAME_CHECK_KEY = "__dateframe_check_config__"
+
 _CONFIG_KEY = "Config"
 
 
-class FieldInfo:
-    """Captures extra information about a field."""
-
-    __slots__ = ("checks", "nullable", "allow_duplicates", "coerce", "regex")
-
-    def __init__(
-        self,
-        checks: Optional[CheckList] = None,
-        nullable: bool = False,
-        allow_duplicates: bool = True,
-        coerce: bool = False,
-        regex: bool = False,
-    ) -> None:
-        self.checks = _to_checklist(checks)
-        self.nullable = nullable
-        self.allow_duplicates = allow_duplicates
-        self.coerce = coerce
-        self.regex = regex
-
-    def _to_schema_component(
-        self,
-        pandas_dtype: PandasDtypeInputTypes,
-        component: Type[SchemaComponent],
-        checks: CheckList = None,
-        **kwargs: Any,
-    ) -> SchemaComponent:
-        checks = _to_checklist(checks)
-        return component(pandas_dtype, checks=self.checks + checks, **kwargs)
-
-    def to_column(
-        self,
-        pandas_dtype: PandasDtypeInputTypes,
-        checks: CheckList = None,
-        required: bool = True,
-        name: str = None,
-    ) -> schema_components.Column:
-        """Create a schema_components.Column from a field."""
-        return self._to_schema_component(
-            pandas_dtype,
-            schema_components.Column,
-            nullable=self.nullable,
-            allow_duplicates=self.allow_duplicates,
-            coerce=self.coerce,
-            regex=self.regex,
-            required=required,
-            name=name,
-            checks=checks,
-        )
-
-    def to_index(
-        self,
-        pandas_dtype: PandasDtypeInputTypes,
-        checks: CheckList = None,
-        name: str = None,
-    ) -> schema_components.Index:
-        """Create a schema_components.Index from a field."""
-        return self._to_schema_component(
-            pandas_dtype,
-            schema_components.Index,
-            nullable=self.nullable,
-            allow_duplicates=self.allow_duplicates,
-            coerce=self.coerce,
[…]

[…]
