### 2020-09-20 b7a00f6
add compariison operator aliases to built-in checks (#269)
