+++ tests/test_checks_builtin.py
@@ -76,12 +76,14 @@ def check_raise_error_or_warning(failure_values, check):
 class TestGreaterThan:
     """Tests for Check.greater_than"""
     @staticmethod
-    def test_argument_check():
+    @pytest.mark.parametrize("check_fn", [Check.greater_than, Check.gt])
+    def test_argument_check(check_fn):
         """Test if None is accepted as boundary"""
         with pytest.raises(ValueError):
-            Check.greater_than(min_value=None)
+            check_fn(min_value=None)
 
     @staticmethod
+    @pytest.mark.parametrize("check_fn", [Check.greater_than, Check.gt])
     @pytest.mark.parametrize('values, min_val', [
         ((1, 2, 3), 0),
         ((1, 2, 3), -1),
@@ -91,11 +93,12 @@ class TestGreaterThan:
          pd.Timestamp("2015-01-01")),
         (("b", "c"), "a")
     ])
-    def test_succeeding(values, min_val):
+    def test_succeeding(check_fn, values, min_val):
         """Run checks which should succeed"""
-        check_values(values, Check.greater_than(min_val), {})
+        check_values(values, check_fn(min_val), {})
 
     @staticmethod
+    @pytest.mark.parametrize("check_fn", [Check.greater_than, Check.gt])
     @pytest.mark.parametrize('values, min_val, failure_cases', [
         ((1, 2, 3), 1, {1}),
         ((3, 2, 1), 1, {1}),
@@ -107,33 +110,40 @@ class TestGreaterThan:
          {pd.Timestamp("2015-02-01"), pd.Timestamp("2015-02-02")}),
         (("b", "c"), "b", {"b"})
     ])
-    def test_failing(values, min_val, failure_cases):
+    def test_failing(check_fn, values, min_val, failure_cases):
         """Run checks which should fail"""
-        check_values(values, Check.greater_than(min_val), failure_cases)
-        check_raise_error_or_warning(values, Check.greater_than(min_val))
+        check_values(values, check_fn(min_val), failure_cases)
+        check_raise_error_or_warning(values, check_fn(min_val))
 
     @staticmethod
+    @pytest.mark.parametrize("check_fn", [Check.greater_than, Check.gt])
     @pytest.mark.parametrize('values, min_val', [
         [(2, None), 1],
         [(pd.Timestamp("2015-02-02"), None), pd.Timestamp("2015-02-01")],
         [("b", None), "a"]
     ])
-    def test_failing_with_none(values, min_val):
+    def test_failing_with_none(check_fn, values, min_val):
         """Validate the check works also on dataframes with None values"""
         check_none_failures(
-            values, Check.greater_than(min_val, ignore_na=False)
+            values, check_fn(min_val, ignore_na=False)
         )
 
 
 class TestGreaterThanOrEqualTo:
     """Tests for Check.greater_than_or_equal_to"""
     @staticmethod
-    def test_argument_check():
+    @pytest.mark.parametrize(
+        "check_fn", [Check.greater_than_or_equal_to, Check.ge]
+    )
+    def test_argument_check(check_fn):
         """Test if None is accepted as boundary"""
         with pytest.raises(ValueError):
-            Check.greater_than_or_equal_to(min_value=None)
+            check_fn(min_value=None)
 
     @staticmethod
+    @pytest.mark.parametrize(
+        "check_fn", [Check.greater_than_or_equal_to, Check.ge]
+    )
     @pytest.mark.parametrize('values, min_val', [
         ((1, 2, 3), 1),
         ((1, 2, 3), -1),
@@ -143,11 +153,14 @@ class TestGreaterThanOrEqualTo:
          pd.Timestamp("2015-02-01")),
         (("b", "a"), "a")
     ])
[…]
