### 2020-09-20 24ef45b
increase class-api coverage to 100%
