### 2020-09-20 c789c83
bump 0.4.5: document new feature versions
add citation information for paper
