+++ pandera/version.py
@@ -1,3 +1,3 @@
 """Version file"""
 
-__version__ = "0.4.4"
+__version__ = "0.4.5"
