### 2020-09-20 6dcd9f7
add support for literal dtypes in SchemaModel
