+++ environment.yml
@@ -37,6 +37,7 @@ dependencies:
   - wrapt
   - pyyaml
   - typing_inspect
+  - typing_extensions
 
   - pip:
     - sphinx_autodoc_typehints==1.10.3
