### 2020-09-20 ad908d2
change readme to use built-in python types
