+++ docs/source/index.rst
@@ -57,9 +57,9 @@ Quick Start
 
     # define schema
     schema = pa.DataFrameSchema({
-        "column1": pa.Column(pa.Int, checks=pa.Check.less_than_or_equal_to(10)),
-        "column2": pa.Column(pa.Float, checks=pa.Check.less_than(-1.2)),
-        "column3": pa.Column(pa.String, checks=[
+        "column1": pa.Column(int, checks=pa.Check.less_than_or_equal_to(10)),
+        "column2": pa.Column(float, checks=pa.Check.less_than(-1.2)),
+        "column3": pa.Column(str, checks=[
             pa.Check.str_startswith("value_"),
             # define custom checks as functions that take a series as input and
             # outputs a boolean or boolean Series
