### 2020-09-21 038876d
bump 0.4.5: document new feature versions (#271)
* bump 0.4.5: document new feature versions

add citation information for paper

* change readme to use built-in python types
