+++ docs/source/schema_inference.rst
@@ -182,7 +182,7 @@ is a convenience method for this functionality.
    :skipif: SKIP
 
     schema_type: dataframe
-    version: 0.4.4
+    version: 0.4.5
     columns:
       column1:
         pandas_dtype: int64
