### 2020-09-22 d1b13dd
drop support and ci for python=3.5
