### 2020-09-22 29b0329
update environment.yml and requirements.txt
