### 2020-09-22 ea3780a
update travis config
