### 2020-09-24 e33597f
fix pylint error, remove io import in __init__
