### 2020-09-24 4c6698d
fix to_script tests
