### 2020-09-24 91e21a2
feature/142: drop support for python3.5 (#272)
* drop support and ci for python=3.5

* update environment.yml and requirements.txt

* update travis config

* update travis config

* update travis config

* remove autopep8

* fix pylint error, remove io import in __init__

* add test for DataFrameSchema.to_script

* fix to_script tests

* ignore test on windows
