### 2020-09-25 88bcf87
implement check_io decorator
