### 2020-09-27 2d8916d
parse pandas version with packaging
