### 2020-09-28 a4bec25
increase test coverage of check_dtypes
