### 2020-09-28 1443047
fix pylint and mypy errors
