### 2020-09-28 b22e51e
increase dtype.py coverage
