### 2020-10-01 6efc956
Rename neq Field check to ne
