### 2020-10-01 e41d0a1
Store SchemaModel.__config__ as a class
