### 2020-10-01 d13b0d0
Fix typo in check_types docstring
