### 2020-10-01 32a350f
Remove unused is_frame_or_series_hint function
