### 2020-10-02 48059ab
class-based api (#258)
* first working class-based api

* first working class-based api

* add pandera.typing.Field

Used to provide extra information about a field of a SchemaModel

* add support for MultiIndex to SchemaModel

* allow check_types without empty parentheses

* add validator decorator

* fix schema_model imports

* add support for SchemaModel inheritance

* add dataframe_validator and dataframe_transformer decorators

* add support for schema-wide model options

* fix postponed annotations

* clean up class-api code

* remove transformer from class-api

* fix test_schema_model e
[…]
