+++ pandera/decorators.py
@@ -1,17 +1,17 @@
 """Decorators for integrating pandera into existing data pipelines."""
 
+import functools
 import inspect
 import warnings
-
 from collections import OrderedDict
-from typing import Any, Callable, List, Union, Tuple, Dict, Optional, NoReturn
-
-import pandas as pd
+from typing import Any, Callable, Dict, List, NoReturn, Optional, Tuple, Union, cast
 
 import wrapt
+import pandas as pd
 
-from . import schemas
-from . import errors
+from . import errors, schemas
+from . typing import parse_annotation
+from . model import SchemaModel
 
 
 def _get_fn_argnames(fn: Callable) -> List[str]:
@@ -29,10 +29,12 @@ def _get_fn_argnames(fn: Callable) -> List[str]:
 
 
 def _handle_schema_error(
-        fn: Callable,
-        schema: Union[schemas.DataFrameSchema, schemas.SeriesSchema],
-        arg_df: pd.DataFrame,
-        schema_error: errors.SchemaError) -> NoReturn:
+    decorator_name,
+    fn: Callable,
+    schema: Union[schemas.DataFrameSchema, schemas.SeriesSchema],
+    arg_df: pd.DataFrame,
+    schema_error: errors.SchemaError,
+) -> NoReturn:
     """Reraise schema validation error with decorator context.
 
     :param fn: check the DataFrame or Series input of this function.
@@ -43,8 +45,8 @@ def _handle_schema_error(
         checks.
     """
     msg = (
-        "error in check_input decorator of function '%s': %s" %
-        (fn.__name__, schema_error)
+        "error in %s decorator of function '%s': %s" %
+        (decorator_name, fn.__name__, schema_error)
     )
     raise errors.SchemaError(
         schema, arg_df, msg,
@@ -173,7 +175,7 @@ def check_input(
             try:
                 args[0] = schema.validate(args[0], *validate_args)
             except errors.SchemaError as e:
-                _handle_schema_error(fn, schema, args[0], e)
+                _handle_schema_error("check_input", fn, schema, args[0], e)
         elif obj_getter is None and kwargs:
             # get the first key in the same order specified in the
             # function argument.
@@ -184,7 +186,9 @@ def check_input(
                     kwargs[args_names[0]], *validate_args
                 )
             except errors.SchemaError as e:
-                _handle_schema_error(fn, schema, kwargs[args_names[0]], e)
+                _handle_schema_error(
+                    "check_input", fn, schema, kwargs[args_names[0]], e
+                )
         else:
             raise ValueError(
                 "obj_getter is unrecognized type: %s" % type(obj_getter))
@@ -293,22 +297,78 @@ def check_output(
         elif callable(obj_getter):
             obj = obj_getter(out)
         else:
-            raise ValueError(
-                "obj_getter is unrecognized type: %s" % type(obj_getter))
+            raise ValueError("obj_getter is unrecognized type: %s" % type(obj_getter))
         try:
             schema.validate(obj, head, tail, sample, random_state, lazy)
         except errors.SchemaError as e:
-            msg = (
-                "error in check_output decorator of function '%s': %s" %
-                (fn.__name__, e)
-            )
-            raise errors.SchemaError(
-                schema, obj, msg,
-                failure_cases=e.failure_cases,
-                check=e.check,
-                check_index=e.check_index,
-            )
+            _handle_schema_error("check_output", fn, schema, obj, e)
 
         return out
 
     return _wrapper
+
+
+def check_types(
+    wrapped=None,
+    *,
+    head: Optional[int] = None,
+    tail: Optional[int] = None,
+    sample: Optional[int] = None,
+    random_state: Optional[int] = None,
+    lazy: bool = False,
+) -> Callable:
[…]
