+++ tests/test_decorators.py
@@ -1,13 +1,27 @@
 """Testing the Decorators that check a functions input or output."""
 
+from typing import Optional
+
 import numpy as np
 import pandas as pd
 import pytest
 
-from pandera import errors
 from pandera import (
-    Column, DataFrameSchema, Check, DateTime, Float, Int, String, check_input,
-    check_output)
+    Check,
+    Column,
+    DataFrameSchema,
+    DateTime,
+    Field,
+    Float,
+    Int,
+    SchemaModel,
+    String,
+    check_input,
+    check_output,
+    check_types,
+    errors,
+)
+from pandera.typing import DataFrame, Index, Series
 
 
 def test_check_function_decorators():
@@ -247,3 +261,153 @@ def test_check_input_method_decorators():
         transformer.transform_secord_arg_with_list_getter(None, dataframe))
     _assert_expectation(
         transformer.transform_secord_arg_with_dict_getter(None, dataframe))
+
+
+# required to be a global: see https://pydantic-docs.helpmanual.io/usage/postponed_annotations/
+class OnlyZeroesSchema(SchemaModel):  # pylint:disable=too-few-public-methods
+    """Schema with a single column containing zeroes."""
+
+    a: Series[int] = Field(eq=0)
+
+
+def test_check_types_arguments():
+    """Test that check_types forwards key-words arguments to validate."""
+    df = pd.DataFrame({"a": [0, 0]})
+
+    @check_types()
+    def transform_empty_parenthesis(
+        df: DataFrame[OnlyZeroesSchema],
+    ) -> DataFrame[OnlyZeroesSchema]:  # pylint: disable=unused-argument
+        return df
+
+    transform_empty_parenthesis(df)
+
+    @check_types(head=1)
+    def transform_head(
+        df: DataFrame[OnlyZeroesSchema],  # pylint: disable=unused-argument
+    ) -> DataFrame[OnlyZeroesSchema]:
+        return pd.DataFrame({"a": [0, 0]})
+
+    transform_head(df)
+
+    @check_types(tail=1)
+    def transform_tail(
+        df: DataFrame[OnlyZeroesSchema],  # pylint: disable=unused-argument
+    ) -> DataFrame[OnlyZeroesSchema]:
+        return pd.DataFrame({"a": [1, 0]})
+
+    transform_tail(df)
+
+    @check_types(lazy=True)
+    def transform_lazy(
+        df: DataFrame[OnlyZeroesSchema],  # pylint: disable=unused-argument
+    ) -> DataFrame[OnlyZeroesSchema]:
+        return pd.DataFrame({"a": [1, 1]})
+
+    with pytest.raises(errors.SchemaErrors, match="Usage Tip"):
+        transform_lazy(df)
+
+
+def test_check_types_unchanged():
+    """Test the check_types behaviour when the dataframe is unchanged within the
+    function being checked."""
+
+    @check_types
+    def transform(
+        df: DataFrame[OnlyZeroesSchema], notused: int  # pylint: disable=unused-argument
+    ) -> DataFrame[OnlyZeroesSchema]:
+        return df
+
+    df = pd.DataFrame({"a": [0]})
+    pd.testing.assert_frame_equal(transform(df, 2), df)
+
+
+# required to be globals: see https://pydantic-docs.helpmanual.io/usage/postponed_annotations/
+class InSchema(SchemaModel):  # pylint:disable=too-few-public-methods
+    """Test schema used as input."""
+
+    a: Series[int]
+    idx: Index[str]
+
+
+class OutSchema(SchemaModel):  # pylint: disable=too-few-public-methods
+    """Test schema used as output."""
+
+    b: Series[int]
+
+
+def test_check_types_multiple_inputs():
+    """Test that check_types behaviour when multiple inputs are annotated."""
+
+    @check_types
[…]
