### 2020-10-04 47bc7b8
fix incorrect docstring (#280)
* fix incorrect docstring

* Update schema_components.py

Co-authored-by: Niels Bantilan <niels.bantilan@gmail.com>
