### 2020-10-06 3bd50f2
cache to_schema call
