### 2020-10-06 390f25a
Added support for pycharm's type checking with dataframe
Made DataFrame a regular pandas dataframe when in IDE and use the pandera DataFrame on runtime.
