### 2020-10-06 dbe73d7
Moved optional packages to extras_require with "all" key
