### 2020-10-06 898893b
use __class_getitem__
this prevents the pylint error `unsubscriptable-object`
