### 2020-10-06 aea17cb
bugfix: caching __schema__ breaks schemas that inherit (#282)
* bugfix: caching __schema__ breaks schemas that inherit

schemas that inherit from base schema don't work as intended
in the `check_types` decorator because the __schema__ class
attribute is cached and used by the child class.

* cache to_schema call
