### 2020-10-06 8439c6d
update contributing guide
