### 2020-10-06 4649558
Updated index.rst with installation options
