### 2020-10-06 2b51aac
Updated README with installation options
