### 2020-10-06 3c4589d
update contributing guide (#285)
