### 2020-10-06 3ff2990
exclude if block from code coverage
