### 2020-10-06 99353f6
Added support for pycharm's type checking with dataframe (#283)
* Added support for pycharm's type checking with dataframe

Made DataFrame a regular pandas dataframe when in IDE and use the pandera DataFrame on runtime.

* use __class_getitem__

this prevents the pylint error `unsubscriptable-object`

* exclude if block from code coverage

Co-authored-by: Niels Bantilan <niels.bantilan@gmail.com>
