+++ README.md
@@ -46,11 +46,10 @@ Using pip:
 ```
 pip install pandera
 
-# Alternatively, If you want to include hypothesis checks
-pip install pandera[hypothesis]
-
-# Alternatively, If you want to include code formatting support
-pip install pandera[io]
+# install optional functionality
+pip install pandera[hypothesis]  # hypothesis checks
+pip install pandera[io]  # yaml/script schema io utilities
+pip install pandera[all]  # all packages
 ```
 
 Using conda:
