### 2020-10-07 930dde1
Added catches for import errors and raise the right error that explains how to download the right dependencies for pandera.
