+++ pandera/io.py
@@ -4,7 +4,14 @@ from functools import partial
 from pathlib import Path
 
 import pandas as pd
-import yaml
+try:
+    import yaml
+except ImportError:  # pragma: no cover
+    raise RuntimeError((
+        'IO and formatting requires "pyyaml" to be installed. \n'
+        'You can install pandera together with the IO dependencies with: \n'
+        "pip install pandera[io]\n"
+    ))
 
 from .dtypes import PandasDtype
 from .schema_statistics import get_dataframe_schema_statistics
@@ -286,7 +293,15 @@ def _format_index(index_statistics):
 
 
 def _format_script(script):
-    import black  # pylint: disable=import-outside-toplevel
+    try:
+        import black  # pylint: disable=import-outside-toplevel
+    except ImportError:  # pragma: no cover
+        raise RuntimeError((
+            'IO and formatting requires "black" to be installed. \n'
+            'You can install pandera together with the IO dependencies with: \n'
+            "pip install pandera[io]\n"
+        ))
+
     formatter = partial(
         black.format_str, mode=black.FileMode(line_length=80)
     )
