### 2020-10-07 09d003d
Fixed pylint issues regarding raise-missing-from
