+++ pandera/hypotheses.py
@@ -8,7 +8,7 @@ import pandas as pd
 try:
     from scipy import stats
 except ImportError as exc:  # pragma: no cover
-    raise RuntimeError((
+    raise ImportError((
         'Hypothesis checks requires "scipy" to be installed. \n'
         'You can install pandera together with the Hypothesis dependencies with: \n'
         "pip install pandera[hypothesis]\n"
