### 2020-10-09 271bc1e
Small changes request by pylint
