### 2021-08-31 #12 @astrojuanlu
I tried to locate the pandera paper on JOSS, without success. Am I missing anything'
