### 2021-08-31 #12 @lwasser
hey there @astrojuanlu i *believe* that @cosmicBboy hasn't yet submitted to JOSS. I briefly chatted over twitter i think  or maybe at scipy and it wasn't submitted yet. it may not be  under review yet.  @cosmicBboy can you confirm? i can  also remove that  tag if you don't  plan on submitting there but it sounded like you were  interested in doing that at some point.  the submission process is fast with JOSS once it goes through our review.
