### 2022-09-28 #12 @lwasser
hey there @cosmicBboy @mbjoseph  👋 Just a friendly reminder to take 5-10 minutes to fill out our survey . We really appreciate it.  Thank you in advance for helping us by filling out the survey!! 🙌  Niels, it's really important for us to collect information from our maintainers so that we can both stay in touch with you regarding package maintenance and also support you through time. We really appreciate your time in filling this out. Also are you the sole maintainer of this package? if not, please have your co-maintainers also fill it out and please list them here as well. Many thanks in advance!

✨ Ivan you only need to do this once :) ping me on slack with any questions!!  🙌

[🔗 HERE IS THE SURVEY LINK 🔗 ](https://forms.gle/ZZXcD1hr18tYmY2p6)
