### 2022-10-19 #12 @lwasser
hi again @cosmicBboy and @mbjoseph i'd be super appreciative if your filling our our survey

[🔗 HERE IS THE SURVEY LINK 🔗](https://forms.gle/ZZXcD1hr18tYmY2p6)!

I know you are busy and Niels I know you have super exciting job transition life happening now. But i'd appreciate your time. We'd like to check in with maintainers once a year to ensure all is well with package maintenance. Also your input on the survey helps us improve and show funders we are doing good things! Many thanks for your time!
