### 2023-04-24 #12 @NickleDave
Hi @cosmicBboy we are updating our metadata to be consistent.

When you have a second, can you please confirm for me that at the time of this review you were the only core maintainer? I have added that in the "all current maintainers" field above (as in #109)
